Visualize: Bundle poset code using webpack
