+++ M2/Macaulay2/packages/Visualize/js/visPoset.js
@@ -1,1589 +1,1 @@
-  // Initialize variables.
-  var width  = null,
-      height = null,
-      colors = null;
-
-  var svg = null;
-  var nodes = null,
-    lastNodeId = null,
-    links = null;
-
-  var constrString = null;
-  var incMatrix = null;
-  var adjMatrix = null;
-  var incMatrixString = null;
-  var adjMatrixString = null;
-
-  var maxGroup = 1;
-  var rowSep = 0;
-  var hPadding = 30;
-  var vPadding = 30;
-
-  var force = null;
-
-  var drag_line = null;
-
-  // Handles to link and node element groups.
-  var path = null,
-      circle = null;
-
-  // Mouse event variables.
-  var selected_node = null,
-      selected_link = null,
-      mousedown_link = null,
-      mousedown_node = null,
-      mouseup_node = null;
-
-  var drag = null;
-
- // Helps determine what menu button was clicked.
-  var clickTest = null; 
-
-  var scriptSource = (function(scripts) {
-    var scripts = document.getElementsByTagName('script'),
-        script = scripts[scripts.length - 1];
-
-    if (script.getAttribute.length !== undefined) {
-        return script.src
-    }
-
-    return script.getAttribute('src', -1)
-    }());
-    
-    // Just get the current directory that contains the html file.
-    scriptSource = scriptSource.substring(0, scriptSource.length - 14);
-
-function initializeBuilder() {
-  // Set up SVG for D3.
-  width  = window.innerWidth-document.getElementById("side").clientWidth;
-  height = window.innerHeight-10;
-  colors = d3.scale.category10();
-  
-  svg = d3.select('body')
-    .append('svg')
-    .attr('width', width)
-    .attr('height', height)
-    .attr('id', 'canvasElement2d');
-    
-  if(fixExtremalNodes){
-      document.getElementById("extremalNodeToggle").text = "Don't fix extremal nodes";
-  }
-
-  // Compute the minimal covering relations from the poset relation matrix.  This is necessary before computing the group for each node.
-  dataCovRel = minimalPosetRelations(dataRelMatrix);
-    
-  // Calculate appropriate groups for elements based on rank or height and store them in the global variable dataGroupList.
-  dataGroupList = computeNodeGroups(dataRelMatrix);
-    
-  // Compute the maximum level of the nodes in the poset.
-  maxGroup = d3.max(dataGroupList);
-  // Compute the distance between levels in the poset.  Make sure the denominator is positive.
-  rowSep = (height-2*vPadding)/d3.max([maxGroup,1]);
-      
-  // Set up initial nodes and links
-  //  - nodes are known by 'id', not by index in array.
-  //  - links are always source < target; edge directions are set by 'left' and 'right'.
-  //var data = dataData;
-  //var names = labelData;
-    
-  // Initialize nodes so that they will be equally spaced along each level.
-  nodes = nodesFromLabelsGroups(dataLabels,dataGroupList);
-  lastNodeId = nodes.length;
-  setAllNodesFixed();
-
-  // Initialize the links based on the minimal covering relations obtained from the relation matrix.
-  links = linksFromNodesRelations(nodes,dataCovRel);
-
-  // Initialize D3 force layout.
-  force = d3.layout.force()
-      .nodes(nodes)
-      .links(links)
-      .size([width, height])
-      .linkDistance(forceLinkDist)
-      .charge(forceCharge)
-      .on('tick', tick);
-
-  // After the force variable is initialized, set the sliders to update the force variables.
-  chargeSlider.noUiSlider.on('slide', updateForceCharge);
-  linkDistSlider.noUiSlider.on('slide', updateForceLinkDist);
-
-  // define arrow markers for graph links
-  svg.append('svg:defs').append('svg:marker')
-    .attr('id', 'end-arrow')
-    .attr('viewBox', '0 -5 10 10')
-    .attr('refX', 6)
-    .attr('markerWidth', 3)
-    .attr('markerHeight', 3)
-    .attr('orient', 'auto')
-    .append('svg:path')
-    .attr('d', 'M0,-5L10,0L0,5')
-    .attr('fill', '#000');
-    
-  // When a node begins to be dragged by the user, call the function dragstart.
-  drag = force.drag()
-    .on("dragstart", dragstart);
-
-  // Line displayed when dragging new nodes.
-  drag_line = svg.append('svg:path')
-    .attr('class', 'link dragline hidden')
-    .attr('d', 'M0,0L0,0');
-
-  // Handles to link and node element groups.
-  path = svg.append('svg:g').selectAll('path');
-  circle = svg.append('svg:g').selectAll('g');
-
-  // Mouse event variables.
-  selected_node = null;
-  selected_link = null;
-  mousedown_link = null;
-  mousedown_node = null;
-  mouseup_node = null;
-    
-  // Define which functions should be called for various mouse events on the svg.
-  svg.on('mousedown', mousedown)
-    .on('mousemove', mousemove)
-    .on('mouseup', mouseup);
-
-  // Define which functions should be called when a key is pressed and released.
-  d3.select(window)
-    .on('keydown', keydown)
-    .on('keyup', keyup);
-    
-  // The restart() function updates the graph.
-  restart();
-  
-  // Brett: Need to fix this.
-  /*
-  var maxLength = d3.max(nodes, function(d) { return d.name.length; });
-
-  console.log("maxLength: " + maxLength + "\n");
-
-  if(maxLength < 4){
-        document.getElementById("nodeText").style.fill = 'white';
-  } else {
-        document.getElementById("nodeText").style.fill = 'black';
-  }
-  */
-
-}
-
-function resetPoset() {
-  
-  // Set the x-coordinate of each node to the original fixed layout.
-  var groupFreq = [];
-  var groupCount = [];
-  for (var i=0; i<maxGroup+1; i++){
-    groupFreq.push(0);
-    groupCount.push(0);
-  }
-  nodes.forEach(function(d){groupFreq[d.group]=groupFreq[d.group]+1;});
-    
-  for( var i = 0; i < nodes.length; i++ ){
-    groupCount[nodes[i].group]=groupCount[nodes[i].group]+1;
-    nodes[i].x = groupCount[nodes[i].group]*((width-2*hPadding)/(groupFreq[nodes[i].group]+1));
-    nodes[i].px = groupCount[nodes[i].group]*((width-2*hPadding)/(groupFreq[nodes[i].group]+1));
-  }
-   
-  setAllNodesFixed();
-  
-  // Calling tick() here is crucial so that the locations of the nodes are updated.
-  tick();
-    
-  // Update the side menu bar to reflect that all nodes are now fixed in their original positions.
-  if(forceOn) toggleForce();
-  
-  restart();
-}
-
-function dragstart(d) {
-  // When dragging a node, set it to be fixed so that the user can give it a static position.
-  d3.select(this).classed(d.fixed = true);
-}
-
-function resetMouseVars() {
-  // Reset all mouse variables.
-  mousedown_node = null;
-  mouseup_node = null;
-  mousedown_link = null;
-}
-
-// Update force layout (called automatically by the force layout simulation each iteration).
-function tick() {
-  
-  // Make sure all nodes stay at the height corresponding to their group.
-  for( var i = 0; i < nodes.length; i++ ){
-    nodes[i].y = height-vPadding-(nodes[i].group)*rowSep;
-    nodes[i].py = height-vPadding-(nodes[i].group)*rowSep;
-  }
-    
-    // Restrict the nodes to be contained within a 15 pixel margin around the svg.
-  circle.attr('transform', function(d) {
-    if (d.x > width - 15) {
-      d.x = width - 15;
-    }
-    else if (d.x < 15) {
-      d.x = 15;
-    }
-    
-    // Visually update the locations of the nodes based on the force simulation.
-    return 'translate(' + d.x + ',' + d.y + ')';
-  });
-    
-  // Draw directed edges with proper padding from node centers.
-  path.attr('d', function(d) {
-    // For each edge, calculate the distance from the source to the target
-    // then normalize the x- and y-distances between the source and target.
-    var deltaX = d.target.x - d.source.x,
-        deltaY = d.target.y - d.source.y,
-        dist = Math.sqrt(deltaX * deltaX + deltaY * deltaY),
-        normX = deltaX / dist,
-        normY = deltaY / dist,
-        // If the edge is directed towards the source, then create extra padding (17) away from the source node to show the arrow,
-        // else set the sourcePadding to 12.
-        sourcePadding = d.left ? 17 : 12,
-        // If the edge is directed towards the target, then create extra padding (17) away from the target node to show the arrow,
-        // else set the targetPadding to 12.
-        targetPadding = d.right ? 17 : 12,
-        // Create new x and y coordinates for the source and the target based on whether extra padding was needed
-        // to account for directed edges.
-        sourceX = d.source.x + (sourcePadding * normX),
-        sourceY = d.source.y + (sourcePadding * normY),
-        targetX = d.target.x - (targetPadding * normX),
-        targetY = d.target.y - (targetPadding * normY);
-    
-    // Restrict the padded x and y coordinates of the source and target to be within a 15 pixel margin around the svg.
-    if (sourceX > width - 15) {
-      sourceX = width - 15;
-    }
-    else if (sourceX < 15) {
-      sourceX = 15;
-    }
-    if (targetX > width - 15) {
-      targetX = width -15;
-    }
-    else if (targetX < 15) {
-      targetX = 15;
-    }
-    if (sourceY > height - 15) {
-      sourceY = height - 15;
-    }
-    else if (sourceY < 15) {
-      sourceY = 15;
-    }
-    if (targetY  > height - 15) {
-      targetY = height - 15;
-    }
-    else if (targetY  < 15) {
-      targetY = 15;
-    }
-    // For each edge, set the attribute 'd' to have the form "MsourcexCoord,sourceyCoord LtargetxCoord,targetyCoord".
-    // Then the appropriate coordinates to use for padding the directed edges away from the nodes can be obtained by
-    // the 'd' attribute.
-    return 'M' + sourceX + ',' + sourceY + 'L' + targetX + ',' + targetY;
-  });
-
-}
-
-// Update graph (called when needed).
-function restart() {
-  // Construct the group of edges from the 'links' array.
-  path = path.data(links);
-
-  // Update existing links.
-  // If a link is currently selected, set 'selected: true'.  If a link should be highlighted, set 'highlighted: true'.
-  path.classed('highlighted', function(d) {return d.highlighted; })
-    .classed('selected', function(d) { return d === selected_link; })
-    // If the edge is directed towards the source or target, attach an arrow.
-    .style('marker-start', function(d) { return d.left ? 'url(#start-arrow)' : ''; })
-    .style('marker-end', function(d) { return d.right ? 'url(#end-arrow)' : ''; });
-
-  // Add new links.
-  path.enter().append('svg:path')
-    .attr('class', 'link')
-    // If a link should be highlighted, set 'highlighted: true'.
-    .classed('highlighted', function(d) {return d.highlighted; })
-    // If a link is currently selected, set 'selected: true'.
-    .classed('selected', function(d) { return d === selected_link; })
-    // If the edge is directed towards the source or target, attach an arrow.
-    .style('marker-start', function(d) { return d.left ? 'url(#start-arrow)' : ''; })
-    .style('marker-end', function(d) { return d.right ? 'url(#end-arrow)' : ''; })
-    .on('mousedown', function(d) {
-      // If the user clicks on a path while either the shift key is pressed or curEdit is false, do nothing.
-      if(d3.event.shiftKey || !curEdit) return;
-
-      // If the user clicks on a path while the shift key is not pressed and curEdit is true, set mousedown_link
-      // to be the path that the user clicked on.
-      mousedown_link = d;
-      
-      // If the link was already selected, then unselect it.
-      if(mousedown_link === selected_link) selected_link = null;
-      
-      // (Brett) Isn't 'if (curEdit)' redundant since we already checked it above?  Remove this line?
-//      else if (curEdit) selected_link = mousedown_link;
-      
-      // If the link was not already selected, then select it.
-      else selected_link = mousedown_link;
-      
-      // Since we selected or unselected a link, set all nodes to be unselected.
-      selected_node = null;
-      // If highlighting neighbors is turned on, un-highlight all nodes and links since there is no currently selected node.
-      if(curHighlight) unHighlightAll();
-      
-      // Update all properties of the graph.
-      restart();
-    });
-
-  // Remove old links.
-  path.exit().remove();
-
-  // Create the circle (node) group.
-  // Note: the function argument is crucial here!  Nodes are known by id, not by index!
-  circle = circle.data(nodes, function(d) { return d.id; });
-
-  // Update existing nodes (highlighted & selected visual states).
-  circle.selectAll('circle')
-    // If a node is currently selected, then make it brighter.
-    .style('fill', function(d) { return (d === selected_node) ? d3.rgb(colors(d.id)).brighter().toString() : (d.highlighted ? '#FF0000' : colors(d.id)); })
-    .classed('highlighted', function(d) { return d.highlighted; })
-    .attr('group', function(d) {return d.group;});
-  
-  // Update the text on each circle with the name of the corresponding node.
-  //circle.select("text").text(function(d) {console.log(d.name); return d.name;});
-
-  // Add new nodes.
-  var g = circle.enter().append('svg:g');
-
-  g.append('svg:circle')
-    .attr('class', 'node')
-    .attr('r', 12)
-    .style('fill', function(d) { return (d === selected_node) ? d3.rgb(colors(d.id)).brighter().toString() : (d.highlighted ? '#FF0000' : colors(d.id)); })
-    .style('stroke', function(d) { return d3.rgb(colors(d.id)).darker().toString(); })
-    .classed('highlighted',function(d) {return d.highlighted;})
-    .on('mouseover', function(d) {
-      // If no node has been previously clicked on or if the user has not dragged the cursor to a different node after clicking,
-      // then do nothing.
-      if (!mousedown_node || d === mousedown_node) return;
-      // Otherwise enlarge the target node.
-      d3.select(this).attr('transform', 'scale(1.1)');
-    })
-    .on('mouseout', function(d) {
-      // If no node has been previously clicked on or if the user has not dragged the cursor to a different node after clicking, then do nothing.
-      if (!mousedown_node || d === mousedown_node) return;
-      // Otherwise unenlarge the target node.  (The user has chosen to not create an edge to this node and has moved the cursor elsewhere.)
-      d3.select(this).attr('transform', '');
-    })
-    .on('mousedown', function(d) {
-      // If either the shift key is held down or editing is disabled, do nothing.
-      // Brett: Add back in the following line if we don't want selected nodes brightened in non-editing mode.
-      //if(d3.event.shiftKey || !curEdit) return;
-      if(d3.event.shiftKey) return;
-
-      // Otherwise, select node.
-      mousedown_node = d;
-      
-      // If the node that the user clicked was already selected, then unselect it.
-      if(mousedown_node === selected_node) { selected_node = null; 
-            if(curHighlight) unHighlightAll(); }
-      //Brett: Add the following line back in if we don't want nodes to be brightened in non-editing mode.
-      //else if(curEdit) { selected_node = mousedown_node;
-      else {selected_node = mousedown_node;
-            if(curHighlight) highlightAllComparable(selected_node);
-      };
-      selected_link = null;
-
-      // reposition drag line
-      if(curEdit){
-        drag_line
-          .style('marker-end', 'url(#end-arrow)')
-          .classed('hidden', false)
-          .attr('d', 'M' + mousedown_node.x + ',' + mousedown_node.y + 'L' + mousedown_node.x + ',' + mousedown_node.y);
-      }
-
-      restart();
-    })
-    .on('mouseup', function(d) {
-      if(!mousedown_node) return;
-
-      // Hide the drag line.
-      drag_line
-        .classed('hidden', true)
-        .style('marker-end', '');
-
-      // If the user dragged a line from a node to itself, reset all mouse variables.
-      mouseup_node = d;
-      if(mouseup_node === mousedown_node) { resetMouseVars(); return; }
-
-      // Unenlarge the node that the mouse was released on.
-      d3.select(this).attr('transform', '');
-
-      // The user has attempted to add a new relation.
-      var link = null;
-      var sourceID = mousedown_node.id;
-      var targetID = mouseup_node.id;
-      // If the source node is already less than the target node, do nothing.
-      if(dataRelMatrix[sourceID][targetID] != 1){
-        // Otherwise, determine whether the relation can be added.
-        var tempMatrix = completeTransitiveClosure(dataRelMatrix,sourceID,targetID);
-        if(!tempMatrix){
-            alert("Creating this relation would violate antisymmetry.");
-        } else {
-            force.stop();
-            dataRelMatrix = tempMatrix.slice();
-            dataCovRel = minimalPosetRelations(dataRelMatrix);
-            dataGroupList = computeNodeGroups(dataRelMatrix);
-            maxGroup = d3.max(dataGroupList);
-            // Make sure the denominator is positive.
-            rowSep = (height-2*vPadding)/d3.max([maxGroup,1]);
-            
-            // Running restart on empty arrays here clears out the circle and path groups so that they are rebuilt from scratch referring to the new node and links below.
-            nodes = [];
-            links = [];
-            restart();
-            
-            nodes = nodesFromLabelsGroups(dataLabels,dataGroupList);
-            lastNodeId = nodes.length;
-            setAllNodesFixed();
-            links = linksFromNodesRelations(nodes,dataCovRel);
-            force.nodes(nodes)
-              .links(links);
-            tick();
-            resetMouseVars();
-            menuDefaults();
-            // 'link' is the relation that was just added.
-            link = links.filter(function(l) {
-                return (l.source.id == d3.min([sourceID,targetID]) && l.target.id == d3.max([sourceID,targetID]));
-            })[0];
-                        
-            // Update the side menu bar to reflect that all nodes are now fixed in their original positions.
-            if(forceOn) toggleForce();
-        }
-      }
-      
-      // Select the newly created link and unselect the previously selected node.
-      if (curEdit) selected_link = link;
-      selected_node = null;
-      if (curHighlight) unHighlightAll();
-      restart();
-    })
-
-  .on('dblclick', function(d) {
-      name = "";
-      // The "/^" in the following regular expression allows ^ as a possible character in names, which may be useful if we want to label the poset elements by monomials, for example.
-      var letters = /^[/^0-9a-zA-Z_]+$/;
-      while (name=="") {
-        name = prompt('Enter new label name.', d.name);
-        // Check whether the user has entered any illegal characters (including spaces).
-        if (!(letters.test(name))) {
-            alert('Please input alphanumeric characters only with no spaces.');
-            name = "";
-        }
-        if (name==d.name) {
-          return;
-        }
-        // Check to see whether there already exists a node with the given name.
-        else if (checkName(name)) {
-          alert('Sorry, a node with that name already exists.')
-          name = "";
-        }
-      }
-            
-      if(name != "null") {
-        d.name = name;
-        d3.select(this.parentNode).select("text").text(function(d) {return d.name});
-        // Update the appropriate entry of the global variable dataLabels with the new node name.
-        dataLabels[d.id] = name;
-      }
-
-    });
-
-  if(labelsOn){
-  // Add text for names for any new nodes.
-    g.append('svg:text')
-        .attr('x', 0)
-        .attr('y', 4)
-        .attr('class', 'id noselect')
-        .attr("pointer-events", "none")
-        .text(function(d) { return d.name; });
-  }
-  /*
-  var maxLength = d3.max(nodes, function(d) {
-        return d.name.length;
-  });
-      
-  if(maxLength < 4){
-        document.getElementById("nodeText").style.fill = 'white';
-  } else {
-        document.getElementById("nodeText").style.fill = 'black';
-  }
-  */
-
-  // Remove the old nodes.
-  circle.exit().remove();
-
-  // Set the force layout in motion.
-  force.start();
-}
-
-function checkName(name) {
-  for (var i = 0; i<nodes.length; i++) {
-    if (nodes[i].name == name) {
-      return true;
-    }
-  }
-  return false;
-}
-
-function getNextAlpha(alpha) {
-  return String.fromCharCode(alpha.charCodeAt(0) + 1);
-}
-
-function mousedown() {
-  // prevent I-bar on drag
-  //d3.event.preventDefault();
-
-  // because :active only works in WebKit?
-  svg.classed('active', true);
-
-  // If editing is not on, the shift key is being held down, or a node or link has been selected, do nothing.
-  if(!curEdit || d3.event.shiftKey || mousedown_node || mousedown_link) return;
-
-  var point = d3.mouse(this);
-  var curName = lastNodeId + 1;
-  while(checkName(curName.toString())){
-      curName += 1;
-  }
-  curName = curName.toString();
-  /*
-  if (checkName(curName)) {
-    curName += 'a';
-  }
-  while (checkName(curName)) {
-    curName = curName.substring(0, curName.length - 1) + getNextAlpha(curName.slice(-1));
-  }
-  */
-
-  // Adding a new minimal node with x-value given by the mouse position on click.
-  var node = {id: lastNodeId++, group: 0, name: curName, highlighted: false};
-  node.x = point[0];
-  node.y = height-vPadding;
-  if (!forceOn) {
-    node.fixed = true;
-  }
-  nodes.push(node);
-  // Add the name of the new node to the global array dataLabels.
-  dataLabels.push(curName);
-  // Add a new row and column to the global array dataRelMatrix with a 1 in the bottom row corresponding to the new node.
-  dataRelMatrix = extendMatrix(dataRelMatrix);
-    
-  // Graph is updated here so we change some items to default.
-  menuDefaults();
-
-  restart();
-}
-
-function mousemove() {
-  // If the user has not selected a node, do nothing.
-  if(!mousedown_node) return;
-
-  // Update the drag line.
-  if(curEdit){
-    drag_line.attr('d', 'M' + mousedown_node.x + ',' + mousedown_node.y + 'L' + d3.mouse(this)[0] + ',' + d3.mouse(this)[1]);
-  }
-    
-  restart();
-}
-
-function mouseup() {
-  if(mousedown_node) {
-    // Hide the drag line.
-    drag_line
-      .classed('hidden', true)
-      .style('marker-end', '');
-  }
-
-  // because :active only works in WebKit?
-  svg.classed('active', false);
-
-  // Reset mouse event variables.
-  resetMouseVars();
-
-  restart();
-
-}
-
-// Remove all links involving a given node.
-function spliceLinksForNode(node) {
-  var toSplice = links.filter(function(l) {
-    return (l.source === node || l.target === node);
-  });
-  toSplice.map(function(l) {
-    links.splice(links.indexOf(l), 1);
-  });
-}
-
-// only respond once per keydown
-var lastKeyDown = -1;
-
-function keydown() {
-  //d3.event.preventDefault();
-
-  // If no key has been pressed, do nothing.
-  if(lastKeyDown !== -1) return;
-  lastKeyDown = d3.event.keyCode;
-
-  // shift
-  if(d3.event.keyCode === 16) {
-    circle.call(drag);
-    svg.classed('shift', true);
-  }
-
-  // The rest of the key presses only apply when a node or link is currently selected.
-  if(!selected_node && !selected_link) return;
-    
-  // -----------------------------------------------------
-    
-  switch(d3.event.keyCode) {
-    case 8: // backspace
-    case 46: // delete
-      if(curEdit && selected_node) {
-        // Delete the selected node and update the poset.
-        dataLabels.splice(selected_node.id,1);
-        force.stop();
-        dataRelMatrix = deleteRowCol(dataRelMatrix,selected_node.id);
-        dataCovRel = minimalPosetRelations(dataRelMatrix);
-        dataGroupList = computeNodeGroups(dataRelMatrix);
-        maxGroup = d3.max(dataGroupList);
-        // Make sure the denominator is positive.
-        rowSep = (height-2*vPadding)/d3.max([maxGroup,1]);
-        
-        nodes = [];
-        links = [];
-        restart();
-          
-        nodes = nodesFromLabelsGroups(dataLabels,dataGroupList);
-        lastNodeId = nodes.length;
-        setAllNodesFixed();
-        links = linksFromNodesRelations(nodes,dataCovRel);
-        force.nodes(nodes)
-          .links(links);
-        resetMouseVars();
-
-        if(curHighlight) unHighlightAll();
-        // Update the side menu bar to reflect that all nodes are now fixed in their original positions.
-        if(forceOn) toggleForce();
-          
-      } else if(curEdit && selected_link) {
-        // Delete the selected link and update the poset.
-        var sourceID = selected_link.source.id;
-        var targetID = selected_link.target.id;
-        // Make the nodes from the selected covering relation incomparable.
-        dataRelMatrix[sourceID][targetID] = 0;
-        dataRelMatrix[targetID][sourceID] = 0;
-        // Update the minimal covering relation and node ranks.
-        dataCovRel = minimalPosetRelations(dataRelMatrix);
-        dataGroupList = computeNodeGroups(dataRelMatrix);
-        maxGroup = d3.max(dataGroupList);
-        // Make sure the denominator is positive.
-        rowSep = (height-2*vPadding)/d3.max([maxGroup,1]);
-        
-        nodes = [];
-        links = [];
-        restart();
-        
-        nodes = nodesFromLabelsGroups(dataLabels,dataGroupList);
-        lastNodeId = nodes.length;
-        setAllNodesFixed();
-        links = linksFromNodesRelations(nodes,dataCovRel);
-        force.nodes(nodes)
-          .links(links);
-        resetMouseVars();
-        
-        if(curHighlight) unHighlightAll();
-        
-        // Update the side menu bar to reflect that all nodes are now fixed in their original positions.
-        if(forceOn) toggleForce();
-      }
-      selected_link = null;
-      if(curEdit) {selected_node = null;}
-
-      // Graph Changed :: deleted nodes and links
-      // as a result we change some items to default    
-      menuDefaults();
-     
-      restart();
-      break;
-  }
-  // ------------------------------------------------------
-  restart();
-}
-
-function keyup() {
-  lastKeyDown = -1;
-
-  // shift
-  if(d3.event.keyCode === 16) {
-    circle
-      .on('mousedown.drag', null)
-      .on('touchstart.drag', null);
-    svg.classed('shift', false);
-  }
-}
-
-function disableEditing() {
-  circle.call(drag);
-  svg.classed('shift', true);
-  //selected_node = null;
-  selected_link = null;
-  //if(curHighlight) unHighlightAll();
-
-  restart();
-}
-
-function enableEditing() {
-  circle
-      .on('mousedown.drag', null)
-      .on('touchstart.drag', null);
-  svg.classed('shift', false);
-}
-
-function enableHighlight() {
-  // If there is no currently selected node, then just return (negating the value of curHighlight).
-  dataMaxChains = posetMaximalChains(dataRelMatrix);
-  if(!selected_node) return;
-  highlightAllComparable(selected_node);
-}
-
-function unHighlightAll() {
-    // Un-highlight all nodes.
-    for (var i = 0; i<nodes.length; i++) {
-       nodes[i].highlighted = false;
-    }
-    
-    // Un-highlight all links.
-    for (var i = 0; i<links.length; i++) {
-       links[i].highlighted = false;
-    }
-    
-    // Update graph based on changes to nodes and links.
-    restart();
-}
-
-// This function uses the global variable dataRelMatrix to determine all of the nodes which are comparable to the given node, then highlights all of these nodes and any links involving two such nodes.
-function highlightAllComparable(n) {
-    unHighlightAll();
-    var rel = allPosetRelations(dataRelMatrix);
-    var compElt = [];
-    for(var i=0; i < rel.length; i++){
-        // Create array of all elements comparable to n.
-        if(rel[i][1] == n.id){compElt.push(rel[i][0]);}
-        if(rel[i][0] == n.id){compElt.push(rel[i][1]);}
-    }
-    // Highlight all comparable nodes.
-    for(var i=0; i < compElt.length; i++){
-        nodes[compElt[i]].highlighted = true;
-    }
-    
-    // Highlight all links whose source and target are both comparable to n.
-    for (var i = 0; i<links.length; i++) {
-       links[i].highlighted = (containedIn(links[i].source.id,compElt) && containedIn(links[i].target.id,compElt));
-    }
-    
-    // Update graph based on changes to nodes and links.
-    restart();
-}
-
-function areNeighbors(node1,node2) {
-    return links.some( function(l) {return (((l.source === node1) && (l.target === node2)) || ((l.target === node1) && (l.source === node2)));});
-}
-
-function setAllNodesFixed() {
-  for (var i = 0; i<nodes.length; i++) {
-    nodes[i].fixed = true;
-  }
-}
-
-function setAllNodesUnfixed() {
-  for (var i = 0; i<nodes.length; i++) {
-    nodes[i].fixed = false;
-  }
-}
-
-function hideLabels() {
-    circle.select("text").remove();    
-}
-
-function showLabels() {
-     circle.append('svg:text')
-      .attr('x', 0)
-      .attr('y', 4)
-      .attr('class', 'id noselect')
-      .attr("pointer-events", "none")
-      .text(function(d) { return d.name; });
-}
-
-// This method returns the groups of the nodes based on rank or height, as appropriate.
-function computeNodeGroups(relMatrix){
-  if(fixExtremalNodes){
-      return posetRelHeightFunction(relMatrix);
-  } else {
-      if(posetIsRanked(dataRelMatrix)){
-          return posetRankFunction(relMatrix);
-      } else {
-          return posetHeightFunction(relMatrix);
-      }
-  }
-}
-
-// This function returns an array of nodes with appropriate x and y values so that the poset elements will be equally distributed along each level.
-function nodesFromLabelsGroups(labelList,groupList) {
-  var temp = [];
-  for(var i=0; i < labelList.length; i++){
-      temp.push({id: i, name: labelList[i], group: groupList[i], highlighted: false});
-  }
-  var groupFreq = [];
-  var groupCount = [];
-  for (var i=0; i<maxGroup+1; i++){
-    groupFreq.push(0);
-    groupCount.push(0);
-  }
-  temp.forEach(function(d){groupFreq[d.group]=groupFreq[d.group]+1;});
-  
-  for(var i=0; i < temp.length; i++){
-      // Set the nodes as fixed by default and specify their initial x and y values to be evenly spaced along their level.
-	  temp[i].y = height-vPadding-temp[i].group*rowSep;
-      temp[i].py = height-vPadding-temp[i].group*rowSep;
-      groupCount[temp[i].group]=groupCount[temp[i].group]+1; 
-      temp[i].x = groupCount[temp[i].group]*((width-2*hPadding)/(groupFreq[temp[i].group]+1));
-      temp[i].px = groupCount[temp[i].group]*((width-2*hPadding)/(groupFreq[temp[i].group]+1));
-  }
-  return temp;
-}
-
-// This function returns a list of the links representing minimal covering relations based on the constructed nodes and a list of minimal relations.
-function linksFromNodesRelations(nodeList,relList){
-  var temp = [];
-  for(var i=0; i < relList.length; i++){
-      // Since the links always go from smaller node id to larger node id, determine which elements in each covering relation have smaller and larger node id.
-      var minNode = d3.min(dataCovRel[i]);
-      var maxNode = d3.max(dataCovRel[i]);
-      
-      // If the first element in the covering relation has the smaller node id, then we should set the link to point 'right' (from smaller node id to larger id, which matches the order given in the covering relation).
-      //var rightLink = (minNode == dataCovRel[i][0]);
-      //var leftLink = !rightLink;
-      
-      // Brett: Setting left or right to true creates a little padding around the target node.  Do we need to do this now that we are keeping track of the relation matrix?
-      // links.push({source: minNode, target: maxNode, left: leftLink, right: rightLink});
-      
-      temp.push({source: nodeList[minNode], target: nodeList[maxNode], left: false, right: false, highlighted: false});
-  }
-  return temp;
-}
-
-function updateWindowSize2d() {
-    //var svg = document.getElementById("canvasElement2d");
-    
-    // get width/height with container selector (body also works)
-    // or use other method of calculating desired values
-    if(!menuOpen){
-        width = window.innerWidth-10;
-    } else {
-        width = window.innerWidth - document.getElementById("side").clientWidth;
-    }
-    height = window.innerHeight-10;
-    // Make sure the denominator is positive.
-    rowSep = (height-2*vPadding)/d3.max([maxGroup,1]);
-
-    // set attrs and 'resume' force 
-    svg.attr('width', width);
-    svg.attr('height', height);
-
-    if(!forceOn){resetPoset();}
-    force.size([width, height]).resume();
-}
-
-// Functions to construct M2 constructors for poset, incidence matrix, and adjacency matrix.  This references the global variable dataCovRel, which is updated with the minimal covering relations each time the poset is updated.
-
-function poset2M2Constructor( labels ){
-  if (labels.length == 0) {
-      return "error \"No constructor for an empty poset.\"";
-  }
-  var covRel = idRelationsToLabelRelations(dataCovRel,labels);
-  var relString = nestedArraytoM2List(covRel);
-  var labelString = "{";
-  var m = labels.length;
-  for( var i = 0; i < m; i++ ){
-    if(i != (m-1)){
-      labelString = labelString + "\"" + labels[i].toString() + "\", ";
-    }
-    else{
-      labelString = labelString + "\"" + labels[i].toString() + "\"}";
-    }
-  }
-  return "poset("+labelString+","+relString+")";
-}
-
-// Constructs the adjacency matrix for a graph as a multidimensional array.
-function getAdjacencyMatrix (nodeSet, edgeSet){
-  var adjMatrix = []; // The next two loops create an initial (nodes.length) x (nodes.length) matrix of zeros.
-  for(var i = 0; i < nodeSet.length; i++){
-    adjMatrix[i] = [];
-    for(var j = 0; j < nodeSet.length; j++){
-      adjMatrix[i][j] = 0;
-    }
-  }
-
-  for (var i = 0; i < edgeSet.length; i++) {
-    adjMatrix[edgeSet[i].source.id][edgeSet[i].target.id] = 1; // Set matrix entries corresponding to adjacencies to 1.
-    adjMatrix[edgeSet[i].target.id][edgeSet[i].source.id] = 1;
-  }
-
-  return adjMatrix;
-}
-
-function updateForceCharge(){
-    if(!forceOn){toggleForce()};
-    forceCharge = -chargeSlider.noUiSlider.get();
-    force.charge(forceCharge).start();
-}
-
-function updateForceLinkDist(){
-    if(!forceOn){toggleForce()};
-    forceLinkDist = linkDistSlider.noUiSlider.get();
-    force.linkDistance(forceLinkDist).start();
-}
-
-
-// ----------- Functions for computations with posets ---------
-
-// Given the relation matrix for a poset, this function returns null if the poset is not ranked and otherwise returns a list of the ranks of the elements.  This algorithm is taken from Posets.m2.
-function posetRankFunction(relMatrix){
-    var rk = [];
-    // Initialize rk with arrays of the form [node index,0].  The second entry will be updated in the next loop and will eventually be the rank that the node is assigned.
-    for(var i=0; i < relMatrix.length; i++){
-        rk.push([i,0]);
-    }
-    var covRel = dataCovRel; // Current minimal covering relations.
-    for(var i=0; i < covRel.length; i++){
-        // 
-        var tmp = rk[covRel[i][1]][rk[covRel[i][1]].length-1] - rk[covRel[i][0]][rk[covRel[i][1]].length-1] - 1;
-        var u = rk[covRel[i][0]][0];
-        var v = rk[covRel[i][1]][0];
-        if(u == v){
-            if(tmp == 0) {
-                continue;
-            } else {
-                return null;
-            }
-        }
-        var temprk = [];
-        if(tmp > 0){
-            for(var j=0; j < rk.length; j++){
-                if(rk[j][0] == u){
-                    temprk.push([v,rk[j][rk[j].length-1] + tmp]);
-                } else {
-                    temprk.push(rk[j]);
-                }
-            }
-        } else {
-            for(var j=0; j < rk.length; j++){
-                if(rk[j][0] == v){
-                    temprk.push([u,rk[j][rk[j].length-1] - tmp]);
-                } else {
-                    temprk.push(rk[j]);
-                }
-            }
-        }
-        rk = temprk;
-    }
-    var rkOutput = [];
-    for(var i=0; i < rk.length; i++){
-        rkOutput.push(rk[i][1]);
-    }
-    return rkOutput;
-}
-
-// Given the relation matrix for a poset, this function determines whether the poset is ranked or not.
-function posetIsRanked(relMatrix){
-    return posetRankFunction(relMatrix) != null;
-}
-
-// Given the relation matrix for a poset, this function computes a filtration of the poset of the form [F_0,F_1,...].  F_0 consists of the minimal elements of the poset, F_1 consists of the minimal elements of P - F_0, and so on.  This algorithm is taken from Posets.m2, which was in turn taken from John Stembridge's Maple package for computations with posets.
-function posetFiltration(relMatrix){
-    //var covRel = minimalPosetRelations(relMatrix);
-    var covRel = dataCovRel;
-    var cnt = [];
-    var cvrby = [];
-    // For each element of the poset, determine how many times it occurs as the larger element in any minimal covering relation (listed in cnt).  Also, for each element of the poset, determine the minimal elements that cover it (listed in cvrby).
-    for(var i=0; i < relMatrix.length; i++){
-        var tempCount = 0;
-        var tempCvrs = [];
-        for(var j=0; j < covRel.length; j++){
-            if(covRel[j][1] == i){tempCount = tempCount+1;}
-            if(covRel[j][0] == i){tempCvrs.push(covRel[j][1]);}
-        }
-        cnt.push(tempCount);
-        cvrby.push(tempCvrs);
-    }
-    // Find indices of all elements that do not minimally cover any element (i.e., the minimal elements in the poset).
-    var neu = [];
-    for(var i=0; i < cnt.length; i++){
-        if(cnt[i] == 0){neu.push(i)};        
-    }
-    // We need neu.slice() here so that the neu array is cloned and ret points to the new occurrence of neu;
-    var ret = [neu.slice()];
-    while(neu.length > 0){
-        var tempMinCvrs = [];
-        for(var i=0; i < neu.length; i++){
-            tempMinCvrs.push(cvrby[neu[i]]);
-        }
-        tempMinCvrs = flattenArray(tempMinCvrs);
-        var tempArr = [];
-        for(var i=0; i < tempMinCvrs.length; i++){
-            if(cnt[tempMinCvrs[i]] == 1){
-                tempArr.push(tempMinCvrs[i]);
-            } else {
-                cnt[tempMinCvrs[i]] = cnt[tempMinCvrs[i]] - 1;
-                continue;
-            }
-        }
-        neu = tempArr.slice();
-        ret.push(tempArr.slice());
-    }
-    ret.splice(ret.length-1,1);
-    return ret;
-}
-
-// Given the relation matrix for a poset, compute the "height" of each element, which is simply the corresponding level of the filtration of the poset that contains it.  This is meant to be used in place of a rank function if the poset is not ranked.
-function posetHeightFunction(relMatrix){
-    var filt = posetFiltration(relMatrix);
-    var ht = [];
-    for(var i=0; i < relMatrix.length; i++){
-        for(var j=0; j < filt.length; j++){
-            for(var k=0; k < filt[j].length; k++){
-                if(filt[j][k] == i){ht.push(j);}
-            }
-        }
-    }
-    return ht;
-}
-
-// Given the relation matrix for a poset, compute the minimal elements.
-function posetMinimalElements(relMatrix){
-    var output = [];
-    for(var i=0; i < relMatrix.length; i++){
-        var isMin = true;
-        for(var j=0; j < relMatrix.length; j++){
-            // If the ith element covers another element, then it is not minimal.
-            if((relMatrix[j][i] == 1) && (i != j)){isMin = false; break;}
-        }
-        if(isMin){output.push(i)};
-    }
-    return output;
-}
-
-// Given the relation matrix for a poset, compute the maximal elements.
-function posetMaximalElements(relMatrix){
-    var output = [];
-    for(var i=0; i < relMatrix.length; i++){
-        var isMax = true;
-        for(var j=0; j < relMatrix.length; j++){
-            // If the ith element is covered by another element, then it is not maximal.
-            if((relMatrix[i][j] == 1) && (i != j)){isMax = false; break;}
-        }
-        if(isMax){output.push(i)};
-    }
-    return output;
-}
-
-// Given the relation matrix for a poset, compute the maximal chains.
-function posetMaximalChains(relMatrix){
-    var minElt = posetMinimalElements(relMatrix);
-    var nonMaximalChains = [];
-    for(var i=0; i < minElt.length; i++){
-        nonMaximalChains.push([minElt[i]]);
-    }
-    var covRel = dataCovRel;
-    var cvrby = [];
-    for(var i=0; i < relMatrix.length; i++){
-        var tempCvrs = [];
-        for(var j=0; j < covRel.length; j++){
-            if(covRel[j][0] == i){tempCvrs.push(covRel[j][1]);}
-        }
-        cvrby.push(tempCvrs);
-    }
-    var maxChains = [];
-    while(nonMaximalChains.length > 0){
-        var tempArr2 = [];
-        for(var i=0; i < nonMaximalChains.length; i++){
-            var tempArr = [];
-            tempCvrs = cvrby[nonMaximalChains[i][nonMaximalChains[i].length-1]];
-            if(tempCvrs.length == 0){
-                // If a maximal chain is found, add it to maxChains.
-                maxChains.push(nonMaximalChains[i].slice());
-                continue;
-            } else {
-                // Otherwise, the chain can be extended further.  Create all possible extensions of the current chain with minimal covering elements.
-                for (var j=0; j < tempCvrs.length; j++){
-                    tempArr.push(nonMaximalChains[i].concat(tempCvrs[j]));
-                }
-                tempArr2.push(tempArr.slice());
-            }
-        }
-        nonMaximalChains = flattenArray(tempArr2.slice());
-    }
-    return maxChains;
-}
-
-// // Given the relation matrix for a poset, compute the "relative height" of each element, which tries to keep elements evenly spaced relative to the length of the maximal chains in the poset.  This is meant to be used in place of a rank function is the poset is not ranked.
-function posetRelHeightFunction(relMatrix){
-    var maxChains = posetMaximalChains(relMatrix);
-    var heightList = posetHeightFunction(relMatrix);
-    // For each element of the poset, create a list of all maximal chains that involve that element.
-    var maxChainList = [];
-    for(var i=0; i < relMatrix.length; i++){
-        var tempArr = [];
-        for(var j=0; j < maxChains.length; j++){
-            for(var k=0; k < maxChains[j].length; k++){
-                // If the ith element appears in the jth maximal chain, then push the jth maximal chain to tempArr.
-                if(maxChains[j][k] == i){tempArr.push(maxChains[j]);}
-            }
-        }
-        maxChainList.push(tempArr.slice());
-    }
-    // For each element of the poset, find the lengths of all maximal chains that involve it.
-    var chainLengthList = [];
-    for(var i=0; i < maxChainList.length; i++){
-        tempArr = [];
-        for(var j=0; j < maxChainList[i].length; j++){
-            tempArr.push(maxChainList[i][j].length);
-        }
-        chainLengthList.push(tempArr.slice());
-    }
-    var relHeightList = [];
-    for(var i=0; i < chainLengthList.length; i++){
-        relHeightList.push(d3.max(chainLengthList[i]) - 1);
-    }
-    var totalHeight = lcm(relHeightList);
-    var output = [];
-    for(var i=0; i < relMatrix.length; i++){
-        output.push((totalHeight/relHeightList[i])*heightList[i]);
-    }
-    return output;
-}
-
-// Given the relation matrix for a set under a binary relation, this function determines whether the relation is antisymmetric (required for a poset) or not.
-function posetIsAntisymmetric(relMatrix){
-    var n = relMatrix.length;
-    for(var i=0; i < n-1; i++){
-        for(var j=i+1; j < n; j++){
-            if((relMatrix[i][j] == 1) && (relMatrix[j][i] == 1)){
-                return false;
-            }            
-        }
-    }
-    return true;
-}
-
-// Given the relation matrix for a poset, this function returns an array consisting of all relations in the poset (with nodes labeled by id).  This assumes that the rows and columns in the relation matrix are indexed according to the id of the nodes.  The relMatrix is given such that relMatrix[i][j] == 1 if and only if node_i <= node_j in the partial order.
-function allPosetRelations (relMatrix){
-    var tempArr = [];
-    var n = relMatrix.length;
-    for(var i=0; i < n; i++){
-        for(var j=i; j < n; j++){
-            // relMatrix[i][j] and relMatrix[j][i] can't both be 1 if i != j or else the poset would not be antisymmetric.
-            if(relMatrix[i][j] == 1){tempArr.push([i,j]);}
-            else {
-                if(relMatrix[j][i] == 1){tempArr.push([j,i]);}
-            }
-        }
-    }
-    
-    return tempArr;
-}
-
-// Given the relation matrix for a poset, this function returns an array consisting of minimal covering relations in the poset (with nodes labeled by id).  This assumes that the rows and columns in the relation matrix are indexed according to the id of the nodes.  The relMatrix is given such that relMatrix[i][j] == 1 if and only if node_j <= node_i in the partial order.  This algorithm is the same as the one used in Posets.m2.
-function minimalPosetRelations (relMatrix){
-    var n = relMatrix.length;
-    var outputArr = [];
-    var gtp = [];
-    for(var i=0; i < n; i++){
-        var temp = [];
-        for(var j=0; j < n; j++){
-            if((i != j) && (relMatrix[i][j] == 1)){temp.push(j)};
-        }
-        gtp.push(temp);
-    }
-    for(var i=0; i < n; i++){
-        var gtgtp = [];
-        var tempIndices = gtp[i];
-        for(var j=0; j < tempIndices.length; j++){
-            gtgtp.push(gtp[tempIndices[j]]);            
-        }
-        gtgtp = eliminateDuplicates(flattenArray(gtgtp));
-        var trimIndices = setDifference(tempIndices,gtgtp);
-        for(var j=0; j < trimIndices.length; j++){
-            outputArr.push([i,trimIndices[j]]);            
-        }        
-    }
-    
-    return outputArr;    
-}
-
-// Given a square array of arrays (representing the relation matrix of a poset) and two distinct non-negative integers r and s, set the (r,s) entry equal to 1 (adding the relation r <= s) and also set all other (i,j) entries such that i <= r and s <= j (under the poset relation) equal to 1.  The method returns null if the resulting matrix is the relation matrix for a non-antisymmetric relation.
-function completeTransitiveClosure(relMatrix,r,s){
-    // If we already have s <= r in the poset, then adding the relation r <= s will make it non-antisymmetric.
-    if(relMatrix[s][r] == 1){return null;}
-    
-    var rel = allPosetRelations(relMatrix);
-    var pred = [];
-    var anc = [];
-    for(var i=0; i < rel.length; i++){
-        // Create arrays of all predecessors of r and all successors of s.
-        if(rel[i][1] == r){pred.push(rel[i][0]);}
-        if(rel[i][0] == s){anc.push(rel[i][1]);}
-    }
-    for(var i=0; i < pred.length; i++){
-        for(var j=0; j < anc.length; j++){
-            // If some successor of s is already <= some predecessor of r then the resulting relation will not be antisymmetric, so return null.
-            if(relMatrix[anc[j]][pred[i]] == 1){return null;}
-            relMatrix[pred[i]][anc[j]] = 1;
-        }
-    }
-    return relMatrix;
-}
-
-// Given a list of relations labeled by node id, return the corresponding list of relations labeled by the name of the corresponding node.
-function idRelationsToLabelRelations (relArr,labelArr){
-    var out = [];
-    for(var i=0; i < relArr.length; i++){
-        out.push([labelArr[relArr[i][0]],labelArr[relArr[i][1]]]);
-    }
-    return out;
-}
-
-// ----------- Helper functions for dealing with arrays. ----------
-
-// Given a nested array of arrays, this flattens the array by one level.
-function flattenArray (arr) {
-    return [].concat.apply([], arr);
-}
-
-// Eliminate all duplicate entries in an array.
-function eliminateDuplicates(arr) {
-  var i,
-      len=arr.length,
-      out=[],
-      obj={};
-
-  for (i=0;i<len;i++) {
-    obj[arr[i]]=0;
-  }
-  for (i in obj) {
-    out.push(i);
-  }
-  return out;
-}
-
-// Take the set-theoretic difference of two arrays.  (i.e., Return the result of removing from the first array all common elements of the second array.)
-function setDifference(arr1,arr2) {
-    var len1 = arr1.length;
-    var len2 = arr2.length;
-    var delIndices = [];
-    for(var i =0; i < len1; i++){
-        for(var j=0; j < len2; j++){
-            // If j[i] appears in arr2, then delete it.
-            if(arr1[i] == arr2[j]){delIndices.push(i);}            
-        }
-    }
-    delIndices = eliminateDuplicates(delIndices);
-    var offset = 0;
-    // Remove all elements of arr1 where common elements with arr2 were found.
-    for(var i=0; i < delIndices.length; i++){
-        arr1.splice(delIndices[i]-offset,1);
-        offset = offset+1;        
-    }
-    
-    return arr1;
-}
-
-// Compute the lcm of an array of integers.
-function lcm(A) {
-    var n = A.length, a = Math.abs(A[0]);
-    for (var i = 1; i < n; i++)
-     { var b = Math.abs(A[i]), c = a;
-       while (a && b){ a > b ? a %= b : b %= a; } 
-       a = Math.abs(c*A[i])/(a+b);
-     }
-    return a;
-}
-
-// Takes a rectangular array of arrays and returns a string which can be copy/pasted into M2.
-function nestedArraytoM2List (arr){
-  var str = "{{";
-  for(var i = 0; i < arr.length; i++){
-    for(var j = 0; j < arr[i].length; j++){
-      str = str + "\"" + arr[i][j].toString() + "\"";
-      if(j == arr[i].length - 1){
-        str = str + "}";
-            } else {
-        str = str + ",";
-      }
-    }
-    if(i < arr.length-1){
-      str = str + ",{";
-    } else {
-      str = str + "}";
-    }
-  }
-
-  return str;
-}
-
-// Given an array and an element, this method returns true if the element is contained in the array, else it returns false.
-function containedIn(n,arr){
-    for(var i=0; i < arr.length; i++){
-        if(arr[i] == n){return true;}
-    }
-    return false;
-}
-
-// Given a rectangular array of arrays (representing the relation matrix of a poset), this method appends a row and column with a 1 in the bottom right spot.
-function extendMatrix (relMatrix){
-    var tempRow = [];
-    for(var i=0; i < relMatrix.length; i++){
-        relMatrix[i].push(0);
-        tempRow.push(0);
-    }
-    tempRow.push(1);
-    relMatrix.push(tempRow);
-    return relMatrix;
-}
-
-// Given a rectangular array of arrays (representing the relation matrix of a poset) and a non-negative integer n, this method deletes the nth row and nth column of the matrix.
-function deleteRowCol (relMatrix,n){
-    relMatrix.splice(n,1);
-    for(var i=0; i < relMatrix.length; i++){
-        relMatrix[i].splice(n,1);
-    }
-    return relMatrix;
-}
-
-// -----------------------------------------------------------
-
-
-// for making unique timestamps in LaTeX. Numbers are not allowed in macros.
-function makeid()
-{
-    var randomtext = "";
-    var randompossible = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz";
-
-    for( var i=0; i < 5; i++ )
-        randomtext += randompossible.charAt(Math.floor(Math.random() * randompossible.length));
-
-    return randomtext;
-}
-
-
-function exportTikz (event){
-  var points = [];
-  for(var i = 0; i < nodes.length; i++){
-    points[i] = [nodes[i].x.toString()+"/"+nodes[i].y.toString()+"/"+nodes[i].id+"/"+nodes[i].name];
-  }
-
-  var edges = [];
-  for(var j = 0; j < links.length; j++){
-    edges[j] = [ links[j].source.id.toString()+"/"+links[j].target.id.toString() ];
-  }
-
-  var timestamp = makeid();
-
-  var tikzTex = "";
-  tikzTex =  "\\begin{tikzpicture}\n" +
-    "\\newcommand*\\points" +timestamp + "{" + points + "}\n" +
-    "\\newcommand*\\edges" + timestamp + "{" + edges + "}\n" +
-    "\\newcommand*\\scale" + timestamp + "{0.02}\n" +
-    "\\foreach \\x/\\y/\\z/\\w in \\points" + timestamp + "\n" +
-    "  \\node (\\z) at (\\scale" + timestamp + "*\\x,-\\scale" + timestamp +
-    "*\\y) [circle,draw,inner sep=0pt] {$\\w$};\n" +
-    "\\foreach \\x/\\y in \\edges" + timestamp + "\n" +
-    "  \\draw (\\x) -- (\\y);\n" +
-    "\\end{tikzpicture}\n" +
-    "% \\points" + timestamp +
-    " is point set in the form x-coord/y-coord/node ID/node label\n" +
-    "% \\edges" + timestamp +
-    " is edge set in the form Source ID/Target ID\n" +
-    "% \\scale" + timestamp +
-    " makes the picture able to be viewed on the page\n";
-
-    
-  if(!tikzGenerated){
-    var tikzDiv = document.createElement("div");
-    tikzDiv.id = "tikzHolder";
-    tikzDiv.className = "list-group-item";
-    tikzDiv.setAttribute('href','#');
-    var tikzInput = document.createElement("textarea");
-    tikzInput.value = "";
-    tikzInput.id = "tikzTextBox";
-    tikzInput.rows = 20;
-    tikzInput.style = "vertical-align:middle; width: 100%";
-    var tikzButton = document.createElement("button");
-    tikzButton.id = "copyButton";
-    tikzButton.style = "vertical-align:middle;";
-    //tikzButton.dataClipboardTarget = "#tikzTextBox";
-    tikzButton.type = "button";
-    var clipboardImg = document.createElement("i");
-    clipboardImg.classList.add("fa-solid", "fa-clipboard");
-    tikzButton.appendChild(clipboardImg);
-    tikzDiv.appendChild(tikzInput);
-    tikzDiv.appendChild(tikzButton);
-    var listGroup = document.getElementById("menuList");
-    listGroup.insertBefore(tikzDiv,listGroup.childNodes[16]);
-    document.getElementById("copyButton").setAttribute("data-clipboard-target","#tikzTextBox");
-    clipboard = new ClipboardJS('#copyButton');
-    clipboard.on('error', function(e) {
-        window.alert("Press enter, then CTRL-C or CMD-C to copy")
-    });  
-    tikzGenerated = true;
-  }
-  document.getElementById("tikzTextBox").value = tikzTex;    
-}
-
-// -----------------------------------------
-// Begin Server Stuff
-// -----------------------------------------
-
-// Add a response for each id from the side menu
-function onclickResults(m2Response) {
-    
-    if (clickTest == "isAtomic"){
-      d3.select("#isAtomic").html("&nbsp;&nbsp; isAtomic :: <b>"+m2Response+"</b>");
-    } 
-    
-    else if (clickTest == "isBounded"){
-      d3.select("#isBounded").html("&nbsp;&nbsp; isBounded :: <b>"+m2Response+"</b>");
-    } 
-    
-    else if (clickTest == "isDistributive"){
-      d3.select("#isDistributive").html("&nbsp;&nbsp; isDistributive :: <b>"+m2Response+"</b>");
-    } 
-    
-    else if (clickTest == "isGeometric"){
-      d3.select("#isGeometric").html("&nbsp;&nbsp; isGeometric :: <b>"+m2Response+"</b>");
-    } 
-    
-    else if (clickTest == "isGraded"){
-      d3.select("#isGraded").html("&nbsp;&nbsp; isGraded :: <b>"+m2Response+"</b>");
-    } 
-    
-    else if (clickTest == "isLattice"){
-      d3.select("#isLattice").html("&nbsp;&nbsp; isLattice :: <b>"+m2Response+"</b>");
-    } 
-    
-    else if (clickTest == "isLowerSemilattice"){
-      d3.select("#isLowerSemilattice").html("&nbsp;&nbsp; isLowerSemilattice :: <b>"+m2Response+"</b>");
-    } 
-    
-    else if (clickTest == "isLowerSemimodular"){
-      d3.select("#isLowerSemimodular").html("&nbsp;&nbsp; isLowerSemimodular :: <b>"+m2Response+"</b>");
-    } 
-    
-    else if (clickTest == "isModular"){
-      d3.select("#isModular").html("&nbsp;&nbsp; isModular :: <b>"+m2Response+"</b>");
-    } 
-    
-    else if (clickTest == "isRanked"){
-      d3.select("#isRanked").html("&nbsp;&nbsp; isRanked :: <b>"+m2Response+"</b>");
-    } 
-    
-    else if (clickTest == "isSperner"){
-      d3.select("#isSperner").html("&nbsp;&nbsp; isSperner :: <b>"+m2Response+"</b>");
-    } 
-    
-    else if (clickTest == "isStrictSperner"){
-      d3.select("#isStrictSperner").html("&nbsp;&nbsp; isStrictSperner :: <b>"+m2Response+"</b>");
-    } 
-    
-    else if (clickTest == "isUpperSemilattice"){
-      d3.select("#isUpperSemilattice").html("&nbsp;&nbsp; isUpperSemilattice :: <b>"+m2Response+"</b>");
-    } 
-    
-    else if (clickTest == "isUpperSemimodular"){
-      d3.select("#isUpperSemimodular").html("&nbsp;&nbsp; isUpperSemimodular :: <b>"+m2Response+"</b>");
-    } 
-    
-    else if (clickTest == "dilworthNumber"){
-      d3.select("#dilworthNumber").html("&nbsp;&nbsp; dilworthNumber :: <b>"+m2Response+"</b>");
-    } 
-}
-
-// Anytime the graph is edited by user we call this function.
-// It changes the menu items to default.
-function menuDefaults() {
-  d3.select("#isAtomic").html("&nbsp;&nbsp; isAtomic");
-  d3.select("#isBounded").html("&nbsp;&nbsp; isBounded");
-  d3.select("#isDistributive").html("&nbsp;&nbsp; isDistributive");
-  d3.select("#isGeometric").html("&nbsp;&nbsp; isGeometric");
-  d3.select("#isGraded").html("&nbsp;&nbsp; isGraded");
-  d3.select("#isLattice").html("&nbsp;&nbsp; isLattice");
-  d3.select("#isLowerSemilattice").html("&nbsp;&nbsp; isLowerSemilattice");
-  d3.select("#isLowerSemimodular").html("&nbsp;&nbsp; isLowerSemimodular");
-  d3.select("#isModular").html("&nbsp;&nbsp; isModular");
-  d3.select("#isRanked").html("&nbsp;&nbsp; isRanked");
-  d3.select("#isSperner").html("&nbsp;&nbsp; isSperner");
-  d3.select("#isStrictSperner").html("&nbsp;&nbsp; isStrictSperner");
-  d3.select("#isUpperSemilattice").html("&nbsp;&nbsp; isUpperSemilattice");
-  d3.select("#isUpperSemimodular").html("&nbsp;&nbsp; isUpperSemimodular");
-  d3.select("#dilworthNumber").html("&nbsp;&nbsp; dilworthNumber");
-  if (tikzGenerated) {
-      d3.select("#tikzHolder").node().remove();
-      tikzGenerated = false;
-  }
-}
-
-// Create the XHR object.
-function createCORSRequest(method, url) {
-  var xhr = new XMLHttpRequest();                    
-  if ("withCredentials" in xhr) {
-    // XHR for Chrome/Firefox/Opera/Safari.
-    xhr.open(method, url, true);
-  } else if (typeof XDomainRequest != "undefined") {
-    // XDomainRequest for IE.
-    xhr = new XDomainRequest();
-    xhr.open(method, url);
-  } else {
-    // CORS not supported.
-    xhr = null;
-  }
-
-  return xhr;
-}
- 
-// Make the actual CORS request.
-function makeCorsRequest(method,url,browserData) {
-  // All HTML5 Rocks properties support CORS.
-  // var url ='http://localhost:8000/fcn2/';
- 
-  var xhr = createCORSRequest(method, url);
-  if (!xhr) {
-    alert('CORS not supported');
-    return;
-  }
- 
-  // Response handlers.
-  xhr.onload = function() {
-    var responseText = xhr.responseText;
-
-    onclickResults(responseText);      
-
-  };
- 
-  //xhr.onerror = function() {
-  //  alert('Woops, there was an error making the request.');
-  //};
-
-  xhr.send(browserData);
-}
-
-// -----------------------------------------
-// End Server Stuff
-// -----------------------------------------
+(()=>{var n={4143:(n,t,e)=>{var r,i;!function(){var o={version:"3.5.17"},u=[].slice,a=function(n){return u.call(n)},l=this.document;function c(n){return n&&(n.ownerDocument||n.document||n).documentElement}function s(n){return n&&(n.ownerDocument&&n.ownerDocument.defaultView||n.document&&n||n.defaultView)}if(l)try{a(l.documentElement.childNodes)[0].nodeType}catch(n){a=function(n){for(var t=n.length,e=new Array(t);t--;)e[t]=n[t];return e}}if(Date.now||(Date.now=function(){return+new Date}),l)try{l.createElement("DIV").style.setProperty("opacity",0,"")}catch(n){var f=this.Element.prototype,h=f.setAttribute,p=f.setAttributeNS,d=this.CSSStyleDeclaration.prototype,g=d.setProperty;f.setAttribute=function(n,t){h.call(this,n,t+"")},f.setAttributeNS=function(n,t,e){p.call(this,n,t,e+"")},d.setProperty=function(n,t,e){g.call(this,n,t+"",e)}}function v(n,t){return n<t?-1:n>t?1:n>=t?0:NaN}function m(n){return null===n?NaN:+n}function y(n){return!isNaN(n)}function x(n){return{left:function(t,e,r,i){for(arguments.length<3&&(r=0),arguments.length<4&&(i=t.length);r<i;){var o=r+i>>>1;n(t[o],e)<0?r=o+1:i=o}return r},right:function(t,e,r,i){for(arguments.length<3&&(r=0),arguments.length<4&&(i=t.length);r<i;){var o=r+i>>>1;n(t[o],e)>0?i=o:r=o+1}return r}}}o.ascending=v,o.descending=function(n,t){return t<n?-1:t>n?1:t>=n?0:NaN},o.min=function(n,t){var e,r,i=-1,o=n.length;if(1===arguments.length){for(;++i<o;)if(null!=(r=n[i])&&r>=r){e=r;break}for(;++i<o;)null!=(r=n[i])&&e>r&&(e=r)}else{for(;++i<o;)if(null!=(r=t.call(n,n[i],i))&&r>=r){e=r;break}for(;++i<o;)null!=(r=t.call(n,n[i],i))&&e>r&&(e=r)}return e},o.max=function(n,t){var e,r,i=-1,o=n.length;if(1===arguments.length){for(;++i<o;)if(null!=(r=n[i])&&r>=r){e=r;break}for(;++i<o;)null!=(r=n[i])&&r>e&&(e=r)}else{for(;++i<o;)if(null!=(r=t.call(n,n[i],i))&&r>=r){e=r;break}for(;++i<o;)null!=(r=t.call(n,n[i],i))&&r>e&&(e=r)}return e},o.extent=function(n,t){var e,r,i,o=-1,u=n.length;if(1===arguments.length){for(;++o<u;)if(null!=(r=n[o])&&r>=r){e=i=r;break}for(;++o<u;)null!=(r=n[o])&&(e>r&&(e=r),i<r&&(i=r))}else{for(;++o<u;)if(null!=(r=t.call(n,n[o],o))&&r>=r){e=i=r;break}for(;++o<u;)null!=(r=t.call(n,n[o],o))&&(e>r&&(e=r),i<r&&(i=r))}return[e,i]},o.sum=function(n,t){var e,r=0,i=n.length,o=-1;if(1===arguments.length)for(;++o<i;)y(e=+n[o])&&(r+=e);else for(;++o<i;)y(e=+t.call(n,n[o],o))&&(r+=e);return r},o.mean=function(n,t){var e,r=0,i=n.length,o=-1,u=i;if(1===arguments.length)for(;++o<i;)y(e=m(n[o]))?r+=e:--u;else for(;++o<i;)y(e=m(t.call(n,n[o],o)))?r+=e:--u;if(u)return r/u},o.quantile=function(n,t){var e=(n.length-1)*t+1,r=Math.floor(e),i=+n[r-1],o=e-r;return o?i+o*(n[r]-i):i},o.median=function(n,t){var e,r=[],i=n.length,u=-1;if(1===arguments.length)for(;++u<i;)y(e=m(n[u]))&&r.push(e);else for(;++u<i;)y(e=m(t.call(n,n[u],u)))&&r.push(e);if(r.length)return o.quantile(r.sort(v),.5)},o.variance=function(n,t){var e,r,i=n.length,o=0,u=0,a=-1,l=0;if(1===arguments.length)for(;++a<i;)y(e=m(n[a]))&&(u+=(r=e-o)*(e-(o+=r/++l)));else for(;++a<i;)y(e=m(t.call(n,n[a],a)))&&(u+=(r=e-o)*(e-(o+=r/++l)));if(l>1)return u/(l-1)},o.deviation=function(){var n=o.variance.apply(this,arguments);return n?Math.sqrt(n):n};var M=x(v);function b(n){return n.length}o.bisectLeft=M.left,o.bisect=o.bisectRight=M.right,o.bisector=function(n){return x(1===n.length?function(t,e){return v(n(t),e)}:n)},o.shuffle=function(n,t,e){(o=arguments.length)<3&&(e=n.length,o<2&&(t=0));for(var r,i,o=e-t;o;)i=Math.random()*o--|0,r=n[o+t],n[o+t]=n[i+t],n[i+t]=r;return n},o.permute=function(n,t){for(var e=t.length,r=new Array(e);e--;)r[e]=n[t[e]];return r},o.pairs=function(n){for(var t=0,e=n.length-1,r=n[0],i=new Array(e<0?0:e);t<e;)i[t]=[r,r=n[++t]];return i},o.transpose=function(n){if(!(i=n.length))return[];for(var t=-1,e=o.min(n,b),r=new Array(e);++t<e;)for(var i,u=-1,a=r[t]=new Array(i);++u<i;)a[u]=n[u][t];return r},o.zip=function(){return o.transpose(arguments)},o.keys=function(n){var t=[];for(var e in n)t.push(e);return t},o.values=function(n){var t=[];for(var e in n)t.push(n[e]);return t},o.entries=function(n){var t=[];for(var e in n)t.push({key:e,value:n[e]});return t},o.merge=function(n){for(var t,e,r,i=n.length,o=-1,u=0;++o<i;)u+=n[o].length;for(e=new Array(u);--i>=0;)for(t=(r=n[i]).length;--t>=0;)e[--u]=r[t];return e};var w=Math.abs;function _(n,t){for(var e in t)Object.defineProperty(n.prototype,e,{value:t[e],enumerable:!1})}function S(){this._=Object.create(null)}function k(n){return"__proto__"==(n+="")||"\0"===n[0]?"\0"+n:n}function E(n){return"\0"===(n+="")[0]?n.slice(1):n}function N(n){return k(n)in this._}function L(n){return(n=k(n))in this._&&delete this._[n]}function C(){var n=[];for(var t in this._)n.push(E(t));return n}function A(){var n=0;for(var t in this._)++n;return n}function T(){for(var n in this._)return!1;return!0}function R(){this._=Object.create(null)}function z(n){return n}function D(n,t,e){return function(){var r=e.apply(t,arguments);return r===t?n:r}}function q(n,t){if(t in n)return t;t=t.charAt(0).toUpperCase()+t.slice(1);for(var e=0,r=O.length;e<r;++e){var i=O[e]+t;if(i in n)return i}}o.range=function(n,t,e){if(arguments.length<3&&(e=1,arguments.length<2&&(t=n,n=0)),(t-n)/e===1/0)throw new Error("infinite range");var r,i=[],o=function(n){for(var t=1;n*t%1;)t*=10;return t}(w(e)),u=-1;if(n*=o,t*=o,(e*=o)<0)for(;(r=n+e*++u)>t;)i.push(r/o);else for(;(r=n+e*++u)<t;)i.push(r/o);return i},o.map=function(n,t){var e=new S;if(n instanceof S)n.forEach(function(n,t){e.set(n,t)});else if(Array.isArray(n)){var r,i=-1,o=n.length;if(1===arguments.length)for(;++i<o;)e.set(i,n[i]);else for(;++i<o;)e.set(t.call(n,r=n[i],i),r)}else for(var u in n)e.set(u,n[u]);return e},_(S,{has:N,get:function(n){return this._[k(n)]},set:function(n,t){return this._[k(n)]=t},remove:L,keys:C,values:function(){var n=[];for(var t in this._)n.push(this._[t]);return n},entries:function(){var n=[];for(var t in this._)n.push({key:E(t),value:this._[t]});return n},size:A,empty:T,forEach:function(n){for(var t in this._)n.call(this,E(t),this._[t])}}),o.nest=function(){var n,t,e={},r=[],i=[];function u(i,o,a){if(a>=r.length)return t?t.call(e,o):n?o.sort(n):o;for(var l,c,s,f,h=-1,p=o.length,d=r[a++],g=new S;++h<p;)(f=g.get(l=d(c=o[h])))?f.push(c):g.set(l,[c]);return i?(c=i(),s=function(n,t){c.set(n,u(i,t,a))}):(c={},s=function(n,t){c[n]=u(i,t,a)}),g.forEach(s),c}function a(n,t){if(t>=r.length)return n;var e=[],o=i[t++];return n.forEach(function(n,r){e.push({key:n,values:a(r,t)})}),o?e.sort(function(n,t){return o(n.key,t.key)}):e}return e.map=function(n,t){return u(t,n,0)},e.entries=function(n){return a(u(o.map,n,0),0)},e.key=function(n){return r.push(n),e},e.sortKeys=function(n){return i[r.length-1]=n,e},e.sortValues=function(t){return n=t,e},e.rollup=function(n){return t=n,e},e},o.set=function(n){var t=new R;if(n)for(var e=0,r=n.length;e<r;++e)t.add(n[e]);return t},_(R,{has:N,add:function(n){return this._[k(n+="")]=!0,n},remove:L,values:C,size:A,empty:T,forEach:function(n){for(var t in this._)n.call(this,E(t))}}),o.behavior={},o.rebind=function(n,t){for(var e,r=1,i=arguments.length;++r<i;)n[e=arguments[r]]=D(n,t,t[e]);return n};var O=["webkit","ms","moz","Moz","o","O"];function P(){}function U(){}function H(n){var t=[],e=new S;function r(){for(var e,r=t,i=-1,o=r.length;++i<o;)(e=r[i].on)&&e.apply(this,arguments);return n}return r.on=function(r,i){var o,u=e.get(r);return arguments.length<2?u&&u.on:(u&&(u.on=null,t=t.slice(0,o=t.indexOf(u)).concat(t.slice(o+1)),e.remove(r)),i&&t.push(e.set(r,{on:i})),n)},r}function j(){o.event.preventDefault()}function I(){for(var n,t=o.event;n=t.sourceEvent;)t=n;return t}function $(n){for(var t=new U,e=0,r=arguments.length;++e<r;)t[arguments[e]]=H(t);return t.of=function(e,r){return function(i){try{var u=i.sourceEvent=o.event;i.target=n,o.event=i,t[i.type].apply(e,r)}finally{o.event=u}}},t}o.dispatch=function(){for(var n=new U,t=-1,e=arguments.length;++t<e;)n[arguments[t]]=H(n);return n},U.prototype.on=function(n,t){var e=n.indexOf("."),r="";if(e>=0&&(r=n.slice(e+1),n=n.slice(0,e)),n)return arguments.length<2?this[n].on(r):this[n].on(r,t);if(2===arguments.length){if(null==t)for(n in this)this.hasOwnProperty(n)&&this[n].on(r,null);return this}},o.event=null,o.requote=function(n){return n.replace(F,"\\$&")};var F=/[\\\^\$\*\+\?\|\[\]\(\)\.\{\}]/g,B={}.__proto__?function(n,t){n.__proto__=t}:function(n,t){for(var e in t)n[e]=t[e]};function G(n){return B(n,X),n}var Y=function(n,t){return t.querySelector(n)},Z=function(n,t){return t.querySelectorAll(n)},V=function(n,t){var e=n.matches||n[q(n,"matchesSelector")];return V=function(n,t){return e.call(n,t)},V(n,t)};"function"==typeof Sizzle&&(Y=function(n,t){return Sizzle(n,t)[0]||null},Z=Sizzle,V=Sizzle.matchesSelector),o.selection=function(){return o.select(l.documentElement)};var X=o.selection.prototype=[];function W(n){return"function"==typeof n?n:function(){return Y(n,this)}}function J(n){return"function"==typeof n?n:function(){return Z(n,this)}}X.select=function(n){var t,e,r,i,o=[];n=W(n);for(var u=-1,a=this.length;++u<a;){o.push(t=[]),t.parentNode=(r=this[u]).parentNode;for(var l=-1,c=r.length;++l<c;)(i=r[l])?(t.push(e=n.call(i,i.__data__,l,u)),e&&"__data__"in i&&(e.__data__=i.__data__)):t.push(null)}return G(o)},X.selectAll=function(n){var t,e,r=[];n=J(n);for(var i=-1,o=this.length;++i<o;)for(var u=this[i],l=-1,c=u.length;++l<c;)(e=u[l])&&(r.push(t=a(n.call(e,e.__data__,l,i))),t.parentNode=e);return G(r)};var K="http://www.w3.org/1999/xhtml",Q={svg:"http://www.w3.org/2000/svg",xhtml:K,xlink:"http://www.w3.org/1999/xlink",xml:"http://www.w3.org/XML/1998/namespace",xmlns:"http://www.w3.org/2000/xmlns/"};function nn(n,t){return n=o.ns.qualify(n),null==t?n.local?function(){this.removeAttributeNS(n.space,n.local)}:function(){this.removeAttribute(n)}:"function"==typeof t?n.local?function(){var e=t.apply(this,arguments);null==e?this.removeAttributeNS(n.space,n.local):this.setAttributeNS(n.space,n.local,e)}:function(){var e=t.apply(this,arguments);null==e?this.removeAttribute(n):this.setAttribute(n,e)}:n.local?function(){this.setAttributeNS(n.space,n.local,t)}:function(){this.setAttribute(n,t)}}function tn(n){return n.trim().replace(/\s+/g," ")}function en(n){return new RegExp("(?:^|\\s+)"+o.requote(n)+"(?:\\s+|$)","g")}function rn(n){return(n+"").trim().split(/^|\s+/)}function on(n,t){var e=(n=rn(n).map(un)).length;return"function"==typeof t?function(){for(var r=-1,i=t.apply(this,arguments);++r<e;)n[r](this,i)}:function(){for(var r=-1;++r<e;)n[r](this,t)}}function un(n){var t=en(n);return function(e,r){if(i=e.classList)return r?i.add(n):i.remove(n);var i=e.getAttribute("class")||"";r?(t.lastIndex=0,t.test(i)||e.setAttribute("class",tn(i+" "+n))):e.setAttribute("class",tn(i.replace(t," ")))}}function an(n,t,e){return null==t?function(){this.style.removeProperty(n)}:"function"==typeof t?function(){var r=t.apply(this,arguments);null==r?this.style.removeProperty(n):this.style.setProperty(n,r,e)}:function(){this.style.setProperty(n,t,e)}}function ln(n,t){return null==t?function(){delete this[n]}:"function"==typeof t?function(){var e=t.apply(this,arguments);null==e?delete this[n]:this[n]=e}:function(){this[n]=t}}function cn(n){return"function"==typeof n?n:(n=o.ns.qualify(n)).local?function(){return this.ownerDocument.createElementNS(n.space,n.local)}:function(){var t=this.ownerDocument,e=this.namespaceURI;return e===K&&t.documentElement.namespaceURI===K?t.createElement(n):t.createElementNS(e,n)}}function sn(){var n=this.parentNode;n&&n.removeChild(this)}function fn(n){return{__data__:n}}function hn(n){return function(){return V(this,n)}}function pn(n){return arguments.length||(n=v),function(t,e){return t&&e?n(t.__data__,e.__data__):!t-!e}}function dn(n,t){for(var e=0,r=n.length;e<r;e++)for(var i,o=n[e],u=0,a=o.length;u<a;u++)(i=o[u])&&t(i,u,e);return n}function gn(n){return B(n,vn),n}o.ns={prefix:Q,qualify:function(n){var t=n.indexOf(":"),e=n;return t>=0&&"xmlns"!==(e=n.slice(0,t))&&(n=n.slice(t+1)),Q.hasOwnProperty(e)?{space:Q[e],local:n}:n}},X.attr=function(n,t){if(arguments.length<2){if("string"==typeof n){var e=this.node();return(n=o.ns.qualify(n)).local?e.getAttributeNS(n.space,n.local):e.getAttribute(n)}for(t in n)this.each(nn(t,n[t]));return this}return this.each(nn(n,t))},X.classed=function(n,t){if(arguments.length<2){if("string"==typeof n){var e=this.node(),r=(n=rn(n)).length,i=-1;if(t=e.classList){for(;++i<r;)if(!t.contains(n[i]))return!1}else for(t=e.getAttribute("class");++i<r;)if(!en(n[i]).test(t))return!1;return!0}for(t in n)this.each(on(t,n[t]));return this}return this.each(on(n,t))},X.style=function(n,t,e){var r=arguments.length;if(r<3){if("string"!=typeof n){for(e in r<2&&(t=""),n)this.each(an(e,n[e],t));return this}if(r<2){var i=this.node();return s(i).getComputedStyle(i,null).getPropertyValue(n)}e=""}return this.each(an(n,t,e))},X.property=function(n,t){if(arguments.length<2){if("string"==typeof n)return this.node()[n];for(t in n)this.each(ln(t,n[t]));return this}return this.each(ln(n,t))},X.text=function(n){return arguments.length?this.each("function"==typeof n?function(){var t=n.apply(this,arguments);this.textContent=null==t?"":t}:null==n?function(){this.textContent=""}:function(){this.textContent=n}):this.node().textContent},X.html=function(n){return arguments.length?this.each("function"==typeof n?function(){var t=n.apply(this,arguments);this.innerHTML=null==t?"":t}:null==n?function(){this.innerHTML=""}:function(){this.innerHTML=n}):this.node().innerHTML},X.append=function(n){return n=cn(n),this.select(function(){return this.appendChild(n.apply(this,arguments))})},X.insert=function(n,t){return n=cn(n),t=W(t),this.select(function(){return this.insertBefore(n.apply(this,arguments),t.apply(this,arguments)||null)})},X.remove=function(){return this.each(sn)},X.data=function(n,t){var e,r,i=-1,o=this.length;if(!arguments.length){for(n=new Array(o=(e=this[0]).length);++i<o;)(r=e[i])&&(n[i]=r.__data__);return n}function u(n,e){var r,i,o,u=n.length,s=e.length,f=Math.min(u,s),h=new Array(s),p=new Array(s),d=new Array(u);if(t){var g,v=new S,m=new Array(u);for(r=-1;++r<u;)(i=n[r])&&(v.has(g=t.call(i,i.__data__,r))?d[r]=i:v.set(g,i),m[r]=g);for(r=-1;++r<s;)(i=v.get(g=t.call(e,o=e[r],r)))?!0!==i&&(h[r]=i,i.__data__=o):p[r]=fn(o),v.set(g,!0);for(r=-1;++r<u;)r in m&&!0!==v.get(m[r])&&(d[r]=n[r])}else{for(r=-1;++r<f;)i=n[r],o=e[r],i?(i.__data__=o,h[r]=i):p[r]=fn(o);for(;r<s;++r)p[r]=fn(e[r]);for(;r<u;++r)d[r]=n[r]}p.update=h,p.parentNode=h.parentNode=d.parentNode=n.parentNode,a.push(p),l.push(h),c.push(d)}var a=gn([]),l=G([]),c=G([]);if("function"==typeof n)for(;++i<o;)u(e=this[i],n.call(e,e.parentNode.__data__,i));else for(;++i<o;)u(e=this[i],n);return l.enter=function(){return a},l.exit=function(){return c},l},X.datum=function(n){return arguments.length?this.property("__data__",n):this.property("__data__")},X.filter=function(n){var t,e,r,i=[];"function"!=typeof n&&(n=hn(n));for(var o=0,u=this.length;o<u;o++){i.push(t=[]),t.parentNode=(e=this[o]).parentNode;for(var a=0,l=e.length;a<l;a++)(r=e[a])&&n.call(r,r.__data__,a,o)&&t.push(r)}return G(i)},X.order=function(){for(var n=-1,t=this.length;++n<t;)for(var e,r=this[n],i=r.length-1,o=r[i];--i>=0;)(e=r[i])&&(o&&o!==e.nextSibling&&o.parentNode.insertBefore(e,o),o=e);return this},X.sort=function(n){n=pn.apply(this,arguments);for(var t=-1,e=this.length;++t<e;)this[t].sort(n);return this.order()},X.each=function(n){return dn(this,function(t,e,r){n.call(t,t.__data__,e,r)})},X.call=function(n){var t=a(arguments);return n.apply(t[0]=this,t),this},X.empty=function(){return!this.node()},X.node=function(){for(var n=0,t=this.length;n<t;n++)for(var e=this[n],r=0,i=e.length;r<i;r++){var o=e[r];if(o)return o}return null},X.size=function(){var n=0;return dn(this,function(){++n}),n};var vn=[];function mn(n,t,e){var r="__on"+n,i=n.indexOf("."),u=xn;i>0&&(n=n.slice(0,i));var l=yn.get(n);function c(){var t=this[r];t&&(this.removeEventListener(n,t,t.$),delete this[r])}return l&&(n=l,u=Mn),i?t?function(){var i=u(t,a(arguments));c.call(this),this.addEventListener(n,this[r]=i,i.$=e),i._=t}:c:t?P:function(){var t,e=new RegExp("^__on([^.]+)"+o.requote(n)+"$");for(var r in this)if(t=r.match(e)){var i=this[r];this.removeEventListener(t[1],i,i.$),delete this[r]}}}o.selection.enter=gn,o.selection.enter.prototype=vn,vn.append=X.append,vn.empty=X.empty,vn.node=X.node,vn.call=X.call,vn.size=X.size,vn.select=function(n){for(var t,e,r,i,o,u=[],a=-1,l=this.length;++a<l;){r=(i=this[a]).update,u.push(t=[]),t.parentNode=i.parentNode;for(var c=-1,s=i.length;++c<s;)(o=i[c])?(t.push(r[c]=e=n.call(i.parentNode,o.__data__,c,a)),e.__data__=o.__data__):t.push(null)}return G(u)},vn.insert=function(n,t){var e,r,i;return arguments.length<2&&(e=this,t=function(n,t,o){var u,a=e[o].update,l=a.length;for(o!=i&&(i=o,r=0),t>=r&&(r=t+1);!(u=a[r])&&++r<l;);return u}),X.insert.call(this,n,t)},o.select=function(n){var t;return"string"==typeof n?(t=[Y(n,l)]).parentNode=l.documentElement:(t=[n]).parentNode=c(n),G([t])},o.selectAll=function(n){var t;return"string"==typeof n?(t=a(Z(n,l))).parentNode=l.documentElement:(t=a(n)).parentNode=null,G([t])},X.on=function(n,t,e){var r=arguments.length;if(r<3){if("string"!=typeof n){for(e in r<2&&(t=!1),n)this.each(mn(e,n[e],t));return this}if(r<2)return(r=this.node()["__on"+n])&&r._;e=!1}return this.each(mn(n,t,e))};var yn=o.map({mouseenter:"mouseover",mouseleave:"mouseout"});function xn(n,t){return function(e){var r=o.event;o.event=e,t[0]=this.__data__;try{n.apply(this,t)}finally{o.event=r}}}function Mn(n,t){var e=xn(n,t);return function(n){var t=this,r=n.relatedTarget;r&&(r===t||8&r.compareDocumentPosition(t))||e.call(t,n)}}l&&yn.forEach(function(n){"on"+n in l&&yn.remove(n)});var bn,wn=0;function _n(n){var t=".dragsuppress-"+ ++wn,e="click"+t,r=o.select(s(n)).on("touchmove"+t,j).on("dragstart"+t,j).on("selectstart"+t,j);if(null==bn&&(bn=!("onselectstart"in n)&&q(n.style,"userSelect")),bn){var i=c(n).style,u=i[bn];i[bn]="none"}return function(n){if(r.on(t,null),bn&&(i[bn]=u),n){var o=function(){r.on(e,null)};r.on(e,function(){j(),o()},!0),setTimeout(o,0)}}}o.mouse=function(n){return kn(n,I())};var Sn=this.navigator&&/WebKit/.test(this.navigator.userAgent)?-1:0;function kn(n,t){t.changedTouches&&(t=t.changedTouches[0]);var e=n.ownerSVGElement||n;if(e.createSVGPoint){var r=e.createSVGPoint();if(Sn<0){var i=s(n);if(i.scrollX||i.scrollY){var u=(e=o.select("body").append("svg").style({position:"absolute",top:0,left:0,margin:0,padding:0,border:"none"},"important"))[0][0].getScreenCTM();Sn=!(u.f||u.e),e.remove()}}return Sn?(r.x=t.pageX,r.y=t.pageY):(r.x=t.clientX,r.y=t.clientY),[(r=r.matrixTransform(n.getScreenCTM().inverse())).x,r.y]}var a=n.getBoundingClientRect();return[t.clientX-a.left-n.clientLeft,t.clientY-a.top-n.clientTop]}function En(){return o.event.changedTouches[0].identifier}o.touch=function(n,t,e){if(arguments.length<3&&(e=t,t=I().changedTouches),t)for(var r,i=0,o=t.length;i<o;++i)if((r=t[i]).identifier===e)return kn(n,r)},o.behavior.drag=function(){var n=$(i,"drag","dragstart","dragend"),t=null,e=u(P,o.mouse,s,"mousemove","mouseup"),r=u(En,o.touch,z,"touchmove","touchend");function i(){this.on("mousedown.drag",e).on("touchstart.drag",r)}function u(e,r,i,u,a){return function(){var l,c=this,s=o.event.target.correspondingElement||o.event.target,f=c.parentNode,h=n.of(c,arguments),p=0,d=e(),g=".drag"+(null==d?"":"-"+d),v=o.select(i(s)).on(u+g,function(){var n,t,e=r(f,d);e&&(n=e[0]-y[0],t=e[1]-y[1],p|=n|t,y=e,h({type:"drag",x:e[0]+l[0],y:e[1]+l[1],dx:n,dy:t}))}).on(a+g,function(){r(f,d)&&(v.on(u+g,null).on(a+g,null),m(p),h({type:"dragend"}))}),m=_n(s),y=r(f,d);l=t?[(l=t.apply(c,arguments)).x-y[0],l.y-y[1]]:[0,0],h({type:"dragstart"})}}return i.origin=function(n){return arguments.length?(t=n,i):t},o.rebind(i,n,"on")},o.touches=function(n,t){return arguments.length<2&&(t=I().touches),t?a(t).map(function(t){var e=kn(n,t);return e.identifier=t.identifier,e}):[]};var Nn=1e-6,Ln=Nn*Nn,Cn=Math.PI,An=2*Cn,Tn=An-Nn,Rn=Cn/2,zn=Cn/180,Dn=180/Cn;function qn(n){return n>0?1:n<0?-1:0}function On(n,t,e){return(t[0]-n[0])*(e[1]-n[1])-(t[1]-n[1])*(e[0]-n[0])}function Pn(n){return n>1?0:n<-1?Cn:Math.acos(n)}function Un(n){return n>1?Rn:n<-1?-Rn:Math.asin(n)}function Hn(n){return((n=Math.exp(n))+1/n)/2}function jn(n){return(n=Math.sin(n/2))*n}var In=Math.SQRT2;o.interpolateZoom=function(n,t){var e,r,i=n[0],o=n[1],u=n[2],a=t[0],l=t[1],c=t[2],s=a-i,f=l-o,h=s*s+f*f;if(h<Ln)r=Math.log(c/u)/In,e=function(n){return[i+n*s,o+n*f,u*Math.exp(In*n*r)]};else{var p=Math.sqrt(h),d=(c*c-u*u+4*h)/(2*u*2*p),g=(c*c-u*u-4*h)/(2*c*2*p),v=Math.log(Math.sqrt(d*d+1)-d),m=Math.log(Math.sqrt(g*g+1)-g);r=(m-v)/In,e=function(n){var t,e=n*r,a=Hn(v),l=u/(2*p)*(a*(t=In*e+v,((t=Math.exp(2*t))-1)/(t+1))-function(n){return((n=Math.exp(n))-1/n)/2}(v));return[i+l*s,o+l*f,u*a/Hn(In*e+v)]}}return e.duration=1e3*r,e},o.behavior.zoom=function(){var n,t,e,r,i,u,a,c,f,h={x:0,y:0,k:1},p=[960,500],d=Bn,g=250,v=0,m="mousedown.zoom",y="mousemove.zoom",x="mouseup.zoom",M="touchstart.zoom",b=$(w,"zoomstart","zoom","zoomend");function w(n){n.on(m,T).on(Fn+".zoom",z).on("dblclick.zoom",D).on(M,R)}function _(n){return[(n[0]-h.x)/h.k,(n[1]-h.y)/h.k]}function S(n){h.k=Math.max(d[0],Math.min(d[1],n))}function k(n,t){t=function(n){return[n[0]*h.k+h.x,n[1]*h.k+h.y]}(t),h.x+=n[0]-t[0],h.y+=n[1]-t[1]}function E(n,e,r,i){n.__chart__={x:h.x,y:h.y,k:h.k},S(Math.pow(2,i)),k(t=e,r),n=o.select(n),g>0&&(n=n.transition().duration(g)),n.call(w.event)}function N(){a&&a.domain(u.range().map(function(n){return(n-h.x)/h.k}).map(u.invert)),f&&f.domain(c.range().map(function(n){return(n-h.y)/h.k}).map(c.invert))}function L(n){v++||n({type:"zoomstart"})}function C(n){N(),n({type:"zoom",scale:h.k,translate:[h.x,h.y]})}function A(n){--v||(n({type:"zoomend"}),t=null)}function T(){var n=this,t=b.of(n,arguments),e=0,r=o.select(s(n)).on(y,function(){e=1,k(o.mouse(n),i),C(t)}).on(x,function(){r.on(y,null).on(x,null),u(e),A(t)}),i=_(o.mouse(n)),u=_n(n);Aa.call(n),L(t)}function R(){var n,t=this,e=b.of(t,arguments),r={},u=0,a=".zoom-"+o.event.changedTouches[0].identifier,l="touchmove"+a,c="touchend"+a,s=[],f=o.select(t),p=_n(t);function d(){var e=o.touches(t);return n=h.k,e.forEach(function(n){n.identifier in r&&(r[n.identifier]=_(n))}),e}function g(){var n=o.event.target;o.select(n).on(l,v).on(c,y),s.push(n);for(var e=o.event.changedTouches,a=0,f=e.length;a<f;++a)r[e[a].identifier]=null;var p=d(),g=Date.now();if(1===p.length){if(g-i<500){var m=p[0];E(t,m,r[m.identifier],Math.floor(Math.log(h.k)/Math.LN2)+1),j()}i=g}else if(p.length>1){m=p[0];var x=p[1],M=m[0]-x[0],b=m[1]-x[1];u=M*M+b*b}}function v(){var a,l,c,s,f=o.touches(t);Aa.call(t);for(var h=0,p=f.length;h<p;++h,s=null)if(c=f[h],s=r[c.identifier]){if(l)break;a=c,l=s}if(s){var d=(d=c[0]-a[0])*d+(d=c[1]-a[1])*d,g=u&&Math.sqrt(d/u);a=[(a[0]+c[0])/2,(a[1]+c[1])/2],l=[(l[0]+s[0])/2,(l[1]+s[1])/2],S(g*n)}i=null,k(a,l),C(e)}function y(){if(o.event.touches.length){for(var n=o.event.changedTouches,t=0,i=n.length;t<i;++t)delete r[n[t].identifier];for(var u in r)return void d()}o.selectAll(s).on(a,null),f.on(m,T).on(M,R),p(),A(e)}g(),L(e),f.on(m,null).on(M,g)}function z(){var i=b.of(this,arguments);r?clearTimeout(r):(Aa.call(this),n=_(t=e||o.mouse(this)),L(i)),r=setTimeout(function(){r=null,A(i)},50),j(),S(Math.pow(2,.002*$n())*h.k),k(t,n),C(i)}function D(){var n=o.mouse(this),t=Math.log(h.k)/Math.LN2;E(this,n,_(n),o.event.shiftKey?Math.ceil(t)-1:Math.floor(t)+1)}return Fn||(Fn="onwheel"in l?($n=function(){return-o.event.deltaY*(o.event.deltaMode?120:1)},"wheel"):"onmousewheel"in l?($n=function(){return o.event.wheelDelta},"mousewheel"):($n=function(){return-o.event.detail},"MozMousePixelScroll")),w.event=function(n){n.each(function(){var n=b.of(this,arguments),e=h;za?o.select(this).transition().each("start.zoom",function(){h=this.__chart__||{x:0,y:0,k:1},L(n)}).tween("zoom:zoom",function(){var r=p[0],i=p[1],u=t?t[0]:r/2,a=t?t[1]:i/2,l=o.interpolateZoom([(u-h.x)/h.k,(a-h.y)/h.k,r/h.k],[(u-e.x)/e.k,(a-e.y)/e.k,r/e.k]);return function(t){var e=l(t),i=r/e[2];this.__chart__=h={x:u-e[0]*i,y:a-e[1]*i,k:i},C(n)}}).each("interrupt.zoom",function(){A(n)}).each("end.zoom",function(){A(n)}):(this.__chart__=h,L(n),C(n),A(n))})},w.translate=function(n){return arguments.length?(h={x:+n[0],y:+n[1],k:h.k},N(),w):[h.x,h.y]},w.scale=function(n){return arguments.length?(h={x:h.x,y:h.y,k:null},S(+n),N(),w):h.k},w.scaleExtent=function(n){return arguments.length?(d=null==n?Bn:[+n[0],+n[1]],w):d},w.center=function(n){return arguments.length?(e=n&&[+n[0],+n[1]],w):e},w.size=function(n){return arguments.length?(p=n&&[+n[0],+n[1]],w):p},w.duration=function(n){return arguments.length?(g=+n,w):g},w.x=function(n){return arguments.length?(a=n,u=n.copy(),h={x:0,y:0,k:1},w):a},w.y=function(n){return arguments.length?(f=n,c=n.copy(),h={x:0,y:0,k:1},w):f},o.rebind(w,b,"on")};var $n,Fn,Bn=[0,1/0];function Gn(){}function Yn(n,t,e){return this instanceof Yn?(this.h=+n,this.s=+t,void(this.l=+e)):arguments.length<2?n instanceof Yn?new Yn(n.h,n.s,n.l):pt(""+n,dt,Yn):new Yn(n,t,e)}o.color=Gn,Gn.prototype.toString=function(){return this.rgb()+""},o.hsl=Yn;var Zn=Yn.prototype=new Gn;function Vn(n,t,e){var r,i;function o(n){return Math.round(255*function(n){return n>360?n-=360:n<0&&(n+=360),n<60?r+(i-r)*n/60:n<180?i:n<240?r+(i-r)*(240-n)/60:r}(n))}return n=isNaN(n)?0:(n%=360)<0?n+360:n,t=isNaN(t)||t<0?0:t>1?1:t,r=2*(e=e<0?0:e>1?1:e)-(i=e<=.5?e*(1+t):e+t-e*t),new lt(o(n+120),o(n),o(n-120))}function Xn(n,t,e){return this instanceof Xn?(this.h=+n,this.c=+t,void(this.l=+e)):arguments.length<2?n instanceof Xn?new Xn(n.h,n.c,n.l):function(n,t,e){return n>0?new Xn(Math.atan2(e,t)*Dn,Math.sqrt(t*t+e*e),n):new Xn(NaN,NaN,n)}(n instanceof Kn?n.l:(n=gt((n=o.rgb(n)).r,n.g,n.b)).l,n.a,n.b):new Xn(n,t,e)}Zn.brighter=function(n){return n=Math.pow(.7,arguments.length?n:1),new Yn(this.h,this.s,this.l/n)},Zn.darker=function(n){return n=Math.pow(.7,arguments.length?n:1),new Yn(this.h,this.s,n*this.l)},Zn.rgb=function(){return Vn(this.h,this.s,this.l)},o.hcl=Xn;var Wn=Xn.prototype=new Gn;function Jn(n,t,e){return isNaN(n)&&(n=0),isNaN(t)&&(t=0),new Kn(e,Math.cos(n*=zn)*t,Math.sin(n)*t)}function Kn(n,t,e){return this instanceof Kn?(this.l=+n,this.a=+t,void(this.b=+e)):arguments.length<2?n instanceof Kn?new Kn(n.l,n.a,n.b):n instanceof Xn?Jn(n.h,n.c,n.l):gt((n=lt(n)).r,n.g,n.b):new Kn(n,t,e)}Wn.brighter=function(n){return new Xn(this.h,this.c,Math.min(100,this.l+Qn*(arguments.length?n:1)))},Wn.darker=function(n){return new Xn(this.h,this.c,Math.max(0,this.l-Qn*(arguments.length?n:1)))},Wn.rgb=function(){return Jn(this.h,this.c,this.l).rgb()},o.lab=Kn;var Qn=18,nt=.95047,tt=1,et=1.08883,rt=Kn.prototype=new Gn;function it(n,t,e){var r=(n+16)/116,i=r+t/500,o=r-e/200;return new lt(at(3.2404542*(i=ot(i)*nt)-1.5371385*(r=ot(r)*tt)-.4985314*(o=ot(o)*et)),at(-.969266*i+1.8760108*r+.041556*o),at(.0556434*i-.2040259*r+1.0572252*o))}function ot(n){return n>.206893034?n*n*n:(n-4/29)/7.787037}function ut(n){return n>.008856?Math.pow(n,1/3):7.787037*n+4/29}function at(n){return Math.round(255*(n<=.00304?12.92*n:1.055*Math.pow(n,1/2.4)-.055))}function lt(n,t,e){return this instanceof lt?(this.r=~~n,this.g=~~t,void(this.b=~~e)):arguments.length<2?n instanceof lt?new lt(n.r,n.g,n.b):pt(""+n,lt,Vn):new lt(n,t,e)}function ct(n){return new lt(n>>16,n>>8&255,255&n)}function st(n){return ct(n)+""}rt.brighter=function(n){return new Kn(Math.min(100,this.l+Qn*(arguments.length?n:1)),this.a,this.b)},rt.darker=function(n){return new Kn(Math.max(0,this.l-Qn*(arguments.length?n:1)),this.a,this.b)},rt.rgb=function(){return it(this.l,this.a,this.b)},o.rgb=lt;var ft=lt.prototype=new Gn;function ht(n){return n<16?"0"+Math.max(0,n).toString(16):Math.min(255,n).toString(16)}function pt(n,t,e){var r,i,o,u=0,a=0,l=0;if(r=/([a-z]+)\((.*)\)/.exec(n=n.toLowerCase()))switch(i=r[2].split(","),r[1]){case"hsl":return e(parseFloat(i[0]),parseFloat(i[1])/100,parseFloat(i[2])/100);case"rgb":return t(mt(i[0]),mt(i[1]),mt(i[2]))}return(o=yt.get(n))?t(o.r,o.g,o.b):(null==n||"#"!==n.charAt(0)||isNaN(o=parseInt(n.slice(1),16))||(4===n.length?(u=(3840&o)>>4,u|=u>>4,a=240&o,a|=a>>4,l=15&o,l|=l<<4):7===n.length&&(u=(16711680&o)>>16,a=(65280&o)>>8,l=255&o)),t(u,a,l))}function dt(n,t,e){var r,i,o=Math.min(n/=255,t/=255,e/=255),u=Math.max(n,t,e),a=u-o,l=(u+o)/2;return a?(i=l<.5?a/(u+o):a/(2-u-o),r=n==u?(t-e)/a+(t<e?6:0):t==u?(e-n)/a+2:(n-t)/a+4,r*=60):(r=NaN,i=l>0&&l<1?0:r),new Yn(r,i,l)}function gt(n,t,e){var r=ut((.4124564*(n=vt(n))+.3575761*(t=vt(t))+.1804375*(e=vt(e)))/nt),i=ut((.2126729*n+.7151522*t+.072175*e)/tt);return Kn(116*i-16,500*(r-i),200*(i-ut((.0193339*n+.119192*t+.9503041*e)/et)))}function vt(n){return(n/=255)<=.04045?n/12.92:Math.pow((n+.055)/1.055,2.4)}function mt(n){var t=parseFloat(n);return"%"===n.charAt(n.length-1)?Math.round(2.55*t):t}ft.brighter=function(n){n=Math.pow(.7,arguments.length?n:1);var t=this.r,e=this.g,r=this.b,i=30;return t||e||r?(t&&t<i&&(t=i),e&&e<i&&(e=i),r&&r<i&&(r=i),new lt(Math.min(255,t/n),Math.min(255,e/n),Math.min(255,r/n))):new lt(i,i,i)},ft.darker=function(n){return new lt((n=Math.pow(.7,arguments.length?n:1))*this.r,n*this.g,n*this.b)},ft.hsl=function(){return dt(this.r,this.g,this.b)},ft.toString=function(){return"#"+ht(this.r)+ht(this.g)+ht(this.b)};var yt=o.map({aliceblue:15792383,antiquewhite:16444375,aqua:65535,aquamarine:8388564,azure:15794175,beige:16119260,bisque:16770244,black:0,blanchedalmond:16772045,blue:255,blueviolet:9055202,brown:10824234,burlywood:14596231,cadetblue:6266528,chartreuse:8388352,chocolate:13789470,coral:16744272,cornflowerblue:6591981,cornsilk:16775388,crimson:14423100,cyan:65535,darkblue:139,darkcyan:35723,darkgoldenrod:12092939,darkgray:11119017,darkgreen:25600,darkgrey:11119017,darkkhaki:12433259,darkmagenta:9109643,darkolivegreen:5597999,darkorange:16747520,darkorchid:10040012,darkred:9109504,darksalmon:15308410,darkseagreen:9419919,darkslateblue:4734347,darkslategray:3100495,darkslategrey:3100495,darkturquoise:52945,darkviolet:9699539,deeppink:16716947,deepskyblue:49151,dimgray:6908265,dimgrey:6908265,dodgerblue:2003199,firebrick:11674146,floralwhite:16775920,forestgreen:2263842,fuchsia:16711935,gainsboro:14474460,ghostwhite:16316671,gold:16766720,goldenrod:14329120,gray:8421504,green:32768,greenyellow:11403055,grey:8421504,honeydew:15794160,hotpink:16738740,indianred:13458524,indigo:4915330,ivory:16777200,khaki:15787660,lavender:15132410,lavenderblush:16773365,lawngreen:8190976,lemonchiffon:16775885,lightblue:11393254,lightcoral:15761536,lightcyan:14745599,lightgoldenrodyellow:16448210,lightgray:13882323,lightgreen:9498256,lightgrey:13882323,lightpink:16758465,lightsalmon:16752762,lightseagreen:2142890,lightskyblue:8900346,lightslategray:7833753,lightslategrey:7833753,lightsteelblue:11584734,lightyellow:16777184,lime:65280,limegreen:3329330,linen:16445670,magenta:16711935,maroon:8388608,mediumaquamarine:6737322,mediumblue:205,mediumorchid:12211667,mediumpurple:9662683,mediumseagreen:3978097,mediumslateblue:8087790,mediumspringgreen:64154,mediumturquoise:4772300,mediumvioletred:13047173,midnightblue:1644912,mintcream:16121850,mistyrose:16770273,moccasin:16770229,navajowhite:16768685,navy:128,oldlace:16643558,olive:8421376,olivedrab:7048739,orange:16753920,orangered:16729344,orchid:14315734,palegoldenrod:15657130,palegreen:10025880,paleturquoise:11529966,palevioletred:14381203,papayawhip:16773077,peachpuff:16767673,peru:13468991,pink:16761035,plum:14524637,powderblue:11591910,purple:8388736,rebeccapurple:6697881,red:16711680,rosybrown:12357519,royalblue:4286945,saddlebrown:9127187,salmon:16416882,sandybrown:16032864,seagreen:3050327,seashell:16774638,sienna:10506797,silver:12632256,skyblue:8900331,slateblue:6970061,slategray:7372944,slategrey:7372944,snow:16775930,springgreen:65407,steelblue:4620980,tan:13808780,teal:32896,thistle:14204888,tomato:16737095,turquoise:4251856,violet:15631086,wheat:16113331,white:16777215,whitesmoke:16119285,yellow:16776960,yellowgreen:10145074});function xt(n){return"function"==typeof n?n:function(){return n}}function Mt(n){return function(t,e,r){return 2===arguments.length&&"function"==typeof e&&(r=e,e=null),bt(t,e,n,r)}}function bt(n,t,e,r){var i={},u=o.dispatch("beforesend","progress","load","error"),l={},c=new XMLHttpRequest,s=null;function f(){var n,t=c.status;if(!t&&function(n){var t=n.responseType;return t&&"text"!==t?n.response:n.responseText}(c)||t>=200&&t<300||304===t){try{n=e.call(i,c)}catch(n){return void u.error.call(i,n)}u.load.call(i,n)}else u.error.call(i,c)}return this.XDomainRequest&&!("withCredentials"in c)&&/^(http(s)?:)?\/\//.test(n)&&(c=new XDomainRequest),"onload"in c?c.onload=c.onerror=f:c.onreadystatechange=function(){c.readyState>3&&f()},c.onprogress=function(n){var t=o.event;o.event=n;try{u.progress.call(i,c)}finally{o.event=t}},i.header=function(n,t){return n=(n+"").toLowerCase(),arguments.length<2?l[n]:(null==t?delete l[n]:l[n]=t+"",i)},i.mimeType=function(n){return arguments.length?(t=null==n?null:n+"",i):t},i.responseType=function(n){return arguments.length?(s=n,i):s},i.response=function(n){return e=n,i},["get","post"].forEach(function(n){i[n]=function(){return i.send.apply(i,[n].concat(a(arguments)))}}),i.send=function(e,r,o){if(2===arguments.length&&"function"==typeof r&&(o=r,r=null),c.open(e,n,!0),null==t||"accept"in l||(l.accept=t+",*/*"),c.setRequestHeader)for(var a in l)c.setRequestHeader(a,l[a]);return null!=t&&c.overrideMimeType&&c.overrideMimeType(t),null!=s&&(c.responseType=s),null!=o&&i.on("error",o).on("load",function(n){o(null,n)}),u.beforesend.call(i,c),c.send(null==r?null:r),i},i.abort=function(){return c.abort(),i},o.rebind(i,u,"on"),null==r?i:i.get(function(n){return 1===n.length?function(t,e){n(null==t?e:null)}:n}(r))}yt.forEach(function(n,t){yt.set(n,ct(t))}),o.functor=xt,o.xhr=Mt(z),o.dsv=function(n,t){var e=new RegExp('["'+n+"\n]"),r=n.charCodeAt(0);function i(n,e,r){arguments.length<3&&(r=e,e=null);var i=bt(n,t,null==e?o:u(e),r);return i.row=function(n){return arguments.length?i.response(null==(e=n)?o:u(n)):e},i}function o(n){return i.parse(n.responseText)}function u(n){return function(t){return i.parse(t.responseText,n)}}function a(t){return t.map(l).join(n)}function l(n){return e.test(n)?'"'+n.replace(/\"/g,'""')+'"':n}return i.parse=function(n,t){var e;return i.parseRows(n,function(n,r){if(e)return e(n,r-1);var i=new Function("d","return {"+n.map(function(n,t){return JSON.stringify(n)+": d["+t+"]"}).join(",")+"}");e=t?function(n,e){return t(i(n),e)}:i})},i.parseRows=function(n,t){var e,i,o={},u={},a=[],l=n.length,c=0,s=0;function f(){if(c>=l)return u;if(i)return i=!1,o;var t=c;if(34===n.charCodeAt(t)){for(var e=t;e++<l;)if(34===n.charCodeAt(e)){if(34!==n.charCodeAt(e+1))break;++e}return c=e+2,13===(a=n.charCodeAt(e+1))?(i=!0,10===n.charCodeAt(e+2)&&++c):10===a&&(i=!0),n.slice(t+1,e).replace(/""/g,'"')}for(;c<l;){var a,s=1;if(10===(a=n.charCodeAt(c++)))i=!0;else if(13===a)i=!0,10===n.charCodeAt(c)&&(++c,++s);else if(a!==r)continue;return n.slice(t,c-s)}return n.slice(t)}for(;(e=f())!==u;){for(var h=[];e!==o&&e!==u;)h.push(e),e=f();t&&null==(h=t(h,s++))||a.push(h)}return a},i.format=function(t){if(Array.isArray(t[0]))return i.formatRows(t);var e=new R,r=[];return t.forEach(function(n){for(var t in n)e.has(t)||r.push(e.add(t))}),[r.map(l).join(n)].concat(t.map(function(t){return r.map(function(n){return l(t[n])}).join(n)})).join("\n")},i.formatRows=function(n){return n.map(a).join("\n")},i},o.csv=o.dsv(",","text/csv"),o.tsv=o.dsv("\t","text/tab-separated-values");var wt,_t,St,kt,Et=this[q(this,"requestAnimationFrame")]||function(n){setTimeout(n,17)};function Nt(n,t,e){var r=arguments.length;r<2&&(t=0),r<3&&(e=Date.now());var i={c:n,t:e+t,n:null};return _t?_t.n=i:wt=i,_t=i,St||(kt=clearTimeout(kt),St=1,Et(Lt)),i}function Lt(){var n=Ct(),t=At()-n;t>24?(isFinite(t)&&(clearTimeout(kt),kt=setTimeout(Lt,t)),St=0):(St=1,Et(Lt))}function Ct(){for(var n=Date.now(),t=wt;t;)n>=t.t&&t.c(n-t.t)&&(t.c=null),t=t.n;return n}function At(){for(var n,t=wt,e=1/0;t;)t.c?(t.t<e&&(e=t.t),t=(n=t).n):t=n?n.n=t.n:wt=t.n;return _t=n,e}function Tt(n,t){return t-(n?Math.ceil(Math.log(n)/Math.LN10):1)}o.timer=function(){Nt.apply(this,arguments)},o.timer.flush=function(){Ct(),At()},o.round=function(n,t){return t?Math.round(n*(t=Math.pow(10,t)))/t:Math.round(n)};var Rt=["y","z","a","f","p","n","µ","m","","k","M","G","T","P","E","Z","Y"].map(function(n,t){var e=Math.pow(10,3*w(8-t));return{scale:t>8?function(n){return n/e}:function(n){return n*e},symbol:n}});function zt(n){var t=n.decimal,e=n.thousands,r=n.grouping,i=n.currency,u=r&&e?function(n,t){for(var i=n.length,o=[],u=0,a=r[0],l=0;i>0&&a>0&&(l+a+1>t&&(a=Math.max(1,t-l)),o.push(n.substring(i-=a,i+a)),!((l+=a+1)>t));)a=r[u=(u+1)%r.length];return o.reverse().join(e)}:z;return function(n){var e=Dt.exec(n),r=e[1]||" ",a=e[2]||">",l=e[3]||"-",c=e[4]||"",s=e[5],f=+e[6],h=e[7],p=e[8],d=e[9],g=1,v="",m="",y=!1,x=!0;switch(p&&(p=+p.substring(1)),(s||"0"===r&&"="===a)&&(s=r="0",a="="),d){case"n":h=!0,d="g";break;case"%":g=100,m="%",d="f";break;case"p":g=100,m="%",d="r";break;case"b":case"o":case"x":case"X":"#"===c&&(v="0"+d.toLowerCase());case"c":x=!1;case"d":y=!0,p=0;break;case"s":g=-1,d="r"}"$"===c&&(v=i[0],m=i[1]),"r"!=d||p||(d="g"),null!=p&&("g"==d?p=Math.max(1,Math.min(21,p)):"e"!=d&&"f"!=d||(p=Math.max(0,Math.min(20,p)))),d=qt.get(d)||Ot;var M=s&&h;return function(n){var e=m;if(y&&n%1)return"";var i=n<0||0===n&&1/n<0?(n=-n,"-"):"-"===l?"":l;if(g<0){var c=o.formatPrefix(n,p);n=c.scale(n),e=c.symbol+m}else n*=g;var b,w,_=(n=d(n,p)).lastIndexOf(".");if(_<0){var S=x?n.lastIndexOf("e"):-1;S<0?(b=n,w=""):(b=n.substring(0,S),w=n.substring(S))}else b=n.substring(0,_),w=t+n.substring(_+1);!s&&h&&(b=u(b,1/0));var k=v.length+b.length+w.length+(M?0:i.length),E=k<f?new Array(k=f-k+1).join(r):"";return M&&(b=u(E+b,E.length?f-w.length:1/0)),i+=v,n=b+w,("<"===a?i+n+E:">"===a?E+i+n:"^"===a?E.substring(0,k>>=1)+i+n+E.substring(k):i+(M?n:E+n))+e}}}o.formatPrefix=function(n,t){var e=0;return(n=+n)&&(n<0&&(n*=-1),t&&(n=o.round(n,Tt(n,t))),e=1+Math.floor(1e-12+Math.log(n)/Math.LN10),e=Math.max(-24,Math.min(24,3*Math.floor((e-1)/3)))),Rt[8+e/3]};var Dt=/(?:([^{])?([<>=^]))?([+\- ])?([$#])?(0)?(\d+)?(,)?(\.-?\d+)?([a-z%])?/i,qt=o.map({b:function(n){return n.toString(2)},c:function(n){return String.fromCharCode(n)},o:function(n){return n.toString(8)},x:function(n){return n.toString(16)},X:function(n){return n.toString(16).toUpperCase()},g:function(n,t){return n.toPrecision(t)},e:function(n,t){return n.toExponential(t)},f:function(n,t){return n.toFixed(t)},r:function(n,t){return(n=o.round(n,Tt(n,t))).toFixed(Math.max(0,Math.min(20,Tt(n*(1+1e-15),t))))}});function Ot(n){return n+""}var Pt=o.time={},Ut=Date;function Ht(){this._=new Date(arguments.length>1?Date.UTC.apply(this,arguments):arguments[0])}Ht.prototype={getDate:function(){return this._.getUTCDate()},getDay:function(){return this._.getUTCDay()},getFullYear:function(){return this._.getUTCFullYear()},getHours:function(){return this._.getUTCHours()},getMilliseconds:function(){return this._.getUTCMilliseconds()},getMinutes:function(){return this._.getUTCMinutes()},getMonth:function(){return this._.getUTCMonth()},getSeconds:function(){return this._.getUTCSeconds()},getTime:function(){return this._.getTime()},getTimezoneOffset:function(){return 0},valueOf:function(){return this._.valueOf()},setDate:function(){jt.setUTCDate.apply(this._,arguments)},setDay:function(){jt.setUTCDay.apply(this._,arguments)},setFullYear:function(){jt.setUTCFullYear.apply(this._,arguments)},setHours:function(){jt.setUTCHours.apply(this._,arguments)},setMilliseconds:function(){jt.setUTCMilliseconds.apply(this._,arguments)},setMinutes:function(){jt.setUTCMinutes.apply(this._,arguments)},setMonth:function(){jt.setUTCMonth.apply(this._,arguments)},setSeconds:function(){jt.setUTCSeconds.apply(this._,arguments)},setTime:function(){jt.setTime.apply(this._,arguments)}};var jt=Date.prototype;function It(n,t,e){function r(t){var e=n(t),r=o(e,1);return t-e<r-t?e:r}function i(e){return t(e=n(new Ut(e-1)),1),e}function o(n,e){return t(n=new Ut(+n),e),n}function u(n,r,o){var u=i(n),a=[];if(o>1)for(;u<r;)e(u)%o||a.push(new Date(+u)),t(u,1);else for(;u<r;)a.push(new Date(+u)),t(u,1);return a}n.floor=n,n.round=r,n.ceil=i,n.offset=o,n.range=u;var a=n.utc=$t(n);return a.floor=a,a.round=$t(r),a.ceil=$t(i),a.offset=$t(o),a.range=function(n,t,e){try{Ut=Ht;var r=new Ht;return r._=n,u(r,t,e)}finally{Ut=Date}},n}function $t(n){return function(t,e){try{Ut=Ht;var r=new Ht;return r._=t,n(r,e)._}finally{Ut=Date}}}function Ft(n){var t=n.dateTime,e=n.date,r=n.time,i=n.periods,u=n.days,a=n.shortDays,l=n.months,c=n.shortMonths;function s(n){var t=n.length;function e(e){for(var r,i,o,u=[],a=-1,l=0;++a<t;)37===n.charCodeAt(a)&&(u.push(n.slice(l,a)),null!=(i=Bt[r=n.charAt(++a)])&&(r=n.charAt(++a)),(o=b[r])&&(r=o(e,null==i?"e"===r?" ":"0":i)),u.push(r),l=a+1);return u.push(n.slice(l,a)),u.join("")}return e.parse=function(t){var e={y:1900,m:0,d:1,H:0,M:0,S:0,L:0,Z:null};if(f(e,n,t,0)!=t.length)return null;"p"in e&&(e.H=e.H%12+12*e.p);var r=null!=e.Z&&Ut!==Ht,i=new(r?Ht:Ut);return"j"in e?i.setFullYear(e.y,0,e.j):"W"in e||"U"in e?("w"in e||(e.w="W"in e?1:0),i.setFullYear(e.y,0,1),i.setFullYear(e.y,0,"W"in e?(e.w+6)%7+7*e.W-(i.getDay()+5)%7:e.w+7*e.U-(i.getDay()+6)%7)):i.setFullYear(e.y,e.m,e.d),i.setHours(e.H+(e.Z/100|0),e.M+e.Z%100,e.S,e.L),r?i._:i},e.toString=function(){return n},e}function f(n,t,e,r){for(var i,o,u,a=0,l=t.length,c=e.length;a<l;){if(r>=c)return-1;if(37===(i=t.charCodeAt(a++))){if(u=t.charAt(a++),!(o=w[u in Bt?t.charAt(a++):u])||(r=o(n,e,r))<0)return-1}else if(i!=e.charCodeAt(r++))return-1}return r}s.utc=function(n){var t=s(n);function e(n){try{var e=new(Ut=Ht);return e._=n,t(e)}finally{Ut=Date}}return e.parse=function(n){try{Ut=Ht;var e=t.parse(n);return e&&e._}finally{Ut=Date}},e.toString=t.toString,e},s.multi=s.utc.multi=fe;var h=o.map(),p=Vt(u),d=Xt(u),g=Vt(a),v=Xt(a),m=Vt(l),y=Xt(l),x=Vt(c),M=Xt(c);i.forEach(function(n,t){h.set(n.toLowerCase(),t)});var b={a:function(n){return a[n.getDay()]},A:function(n){return u[n.getDay()]},b:function(n){return c[n.getMonth()]},B:function(n){return l[n.getMonth()]},c:s(t),d:function(n,t){return Zt(n.getDate(),t,2)},e:function(n,t){return Zt(n.getDate(),t,2)},H:function(n,t){return Zt(n.getHours(),t,2)},I:function(n,t){return Zt(n.getHours()%12||12,t,2)},j:function(n,t){return Zt(1+Pt.dayOfYear(n),t,3)},L:function(n,t){return Zt(n.getMilliseconds(),t,3)},m:function(n,t){return Zt(n.getMonth()+1,t,2)},M:function(n,t){return Zt(n.getMinutes(),t,2)},p:function(n){return i[+(n.getHours()>=12)]},S:function(n,t){return Zt(n.getSeconds(),t,2)},U:function(n,t){return Zt(Pt.sundayOfYear(n),t,2)},w:function(n){return n.getDay()},W:function(n,t){return Zt(Pt.mondayOfYear(n),t,2)},x:s(e),X:s(r),y:function(n,t){return Zt(n.getFullYear()%100,t,2)},Y:function(n,t){return Zt(n.getFullYear()%1e4,t,4)},Z:ce,"%":function(){return"%"}},w={a:function(n,t,e){g.lastIndex=0;var r=g.exec(t.slice(e));return r?(n.w=v.get(r[0].toLowerCase()),e+r[0].length):-1},A:function(n,t,e){p.lastIndex=0;var r=p.exec(t.slice(e));return r?(n.w=d.get(r[0].toLowerCase()),e+r[0].length):-1},b:function(n,t,e){x.lastIndex=0;var r=x.exec(t.slice(e));return r?(n.m=M.get(r[0].toLowerCase()),e+r[0].length):-1},B:function(n,t,e){m.lastIndex=0;var r=m.exec(t.slice(e));return r?(n.m=y.get(r[0].toLowerCase()),e+r[0].length):-1},c:function(n,t,e){return f(n,b.c.toString(),t,e)},d:re,e:re,H:oe,I:oe,j:ie,L:le,m:ee,M:ue,p:function(n,t,e){var r=h.get(t.slice(e,e+=2).toLowerCase());return null==r?-1:(n.p=r,e)},S:ae,U:Jt,w:Wt,W:Kt,x:function(n,t,e){return f(n,b.x.toString(),t,e)},X:function(n,t,e){return f(n,b.X.toString(),t,e)},y:ne,Y:Qt,Z:te,"%":se};return s}Pt.year=It(function(n){return(n=Pt.day(n)).setMonth(0,1),n},function(n,t){n.setFullYear(n.getFullYear()+t)},function(n){return n.getFullYear()}),Pt.years=Pt.year.range,Pt.years.utc=Pt.year.utc.range,Pt.day=It(function(n){var t=new Ut(2e3,0);return t.setFullYear(n.getFullYear(),n.getMonth(),n.getDate()),t},function(n,t){n.setDate(n.getDate()+t)},function(n){return n.getDate()-1}),Pt.days=Pt.day.range,Pt.days.utc=Pt.day.utc.range,Pt.dayOfYear=function(n){var t=Pt.year(n);return Math.floor((n-t-6e4*(n.getTimezoneOffset()-t.getTimezoneOffset()))/864e5)},["sunday","monday","tuesday","wednesday","thursday","friday","saturday"].forEach(function(n,t){t=7-t;var e=Pt[n]=It(function(n){return(n=Pt.day(n)).setDate(n.getDate()-(n.getDay()+t)%7),n},function(n,t){n.setDate(n.getDate()+7*Math.floor(t))},function(n){var e=Pt.year(n).getDay();return Math.floor((Pt.dayOfYear(n)+(e+t)%7)/7)-(e!==t)});Pt[n+"s"]=e.range,Pt[n+"s"].utc=e.utc.range,Pt[n+"OfYear"]=function(n){var e=Pt.year(n).getDay();return Math.floor((Pt.dayOfYear(n)+(e+t)%7)/7)}}),Pt.week=Pt.sunday,Pt.weeks=Pt.sunday.range,Pt.weeks.utc=Pt.sunday.utc.range,Pt.weekOfYear=Pt.sundayOfYear;var Bt={"-":"",_:" ",0:"0"},Gt=/^\s*\d+/,Yt=/^%/;function Zt(n,t,e){var r=n<0?"-":"",i=(r?-n:n)+"",o=i.length;return r+(o<e?new Array(e-o+1).join(t)+i:i)}function Vt(n){return new RegExp("^(?:"+n.map(o.requote).join("|")+")","i")}function Xt(n){for(var t=new S,e=-1,r=n.length;++e<r;)t.set(n[e].toLowerCase(),e);return t}function Wt(n,t,e){Gt.lastIndex=0;var r=Gt.exec(t.slice(e,e+1));return r?(n.w=+r[0],e+r[0].length):-1}function Jt(n,t,e){Gt.lastIndex=0;var r=Gt.exec(t.slice(e));return r?(n.U=+r[0],e+r[0].length):-1}function Kt(n,t,e){Gt.lastIndex=0;var r=Gt.exec(t.slice(e));return r?(n.W=+r[0],e+r[0].length):-1}function Qt(n,t,e){Gt.lastIndex=0;var r=Gt.exec(t.slice(e,e+4));return r?(n.y=+r[0],e+r[0].length):-1}function ne(n,t,e){Gt.lastIndex=0;var r,i=Gt.exec(t.slice(e,e+2));return i?(n.y=(r=+i[0])+(r>68?1900:2e3),e+i[0].length):-1}function te(n,t,e){return/^[+-]\d{4}$/.test(t=t.slice(e,e+5))?(n.Z=-t,e+5):-1}function ee(n,t,e){Gt.lastIndex=0;var r=Gt.exec(t.slice(e,e+2));return r?(n.m=r[0]-1,e+r[0].length):-1}function re(n,t,e){Gt.lastIndex=0;var r=Gt.exec(t.slice(e,e+2));return r?(n.d=+r[0],e+r[0].length):-1}function ie(n,t,e){Gt.lastIndex=0;var r=Gt.exec(t.slice(e,e+3));return r?(n.j=+r[0],e+r[0].length):-1}function oe(n,t,e){Gt.lastIndex=0;var r=Gt.exec(t.slice(e,e+2));return r?(n.H=+r[0],e+r[0].length):-1}function ue(n,t,e){Gt.lastIndex=0;var r=Gt.exec(t.slice(e,e+2));return r?(n.M=+r[0],e+r[0].length):-1}function ae(n,t,e){Gt.lastIndex=0;var r=Gt.exec(t.slice(e,e+2));return r?(n.S=+r[0],e+r[0].length):-1}function le(n,t,e){Gt.lastIndex=0;var r=Gt.exec(t.slice(e,e+3));return r?(n.L=+r[0],e+r[0].length):-1}function ce(n){var t=n.getTimezoneOffset(),e=t>0?"-":"+",r=w(t)/60|0,i=w(t)%60;return e+Zt(r,"0",2)+Zt(i,"0",2)}function se(n,t,e){Yt.lastIndex=0;var r=Yt.exec(t.slice(e,e+1));return r?e+r[0].length:-1}function fe(n){for(var t=n.length,e=-1;++e<t;)n[e][0]=this(n[e][0]);return function(t){for(var e=0,r=n[e];!r[1](t);)r=n[++e];return r[0](t)}}o.locale=function(n){return{numberFormat:zt(n),timeFormat:Ft(n)}};var he=o.locale({decimal:".",thousands:",",grouping:[3],currency:["$",""],dateTime:"%a %b %e %X %Y",date:"%m/%d/%Y",time:"%H:%M:%S",periods:["AM","PM"],days:["Sunday","Monday","Tuesday","Wednesday","Thursday","Friday","Saturday"],shortDays:["Sun","Mon","Tue","Wed","Thu","Fri","Sat"],months:["January","February","March","April","May","June","July","August","September","October","November","December"],shortMonths:["Jan","Feb","Mar","Apr","May","Jun","Jul","Aug","Sep","Oct","Nov","Dec"]});function pe(){}o.format=he.numberFormat,o.geo={},pe.prototype={s:0,t:0,add:function(n){ge(n,this.t,de),ge(de.s,this.s,this),this.s?this.t+=de.t:this.s=de.t},reset:function(){this.s=this.t=0},valueOf:function(){return this.s}};var de=new pe;function ge(n,t,e){var r=e.s=n+t,i=r-n,o=r-i;e.t=n-o+(t-i)}function ve(n,t){n&&ye.hasOwnProperty(n.type)&&ye[n.type](n,t)}o.geo.stream=function(n,t){n&&me.hasOwnProperty(n.type)?me[n.type](n,t):ve(n,t)};var me={Feature:function(n,t){ve(n.geometry,t)},FeatureCollection:function(n,t){for(var e=n.features,r=-1,i=e.length;++r<i;)ve(e[r].geometry,t)}},ye={Sphere:function(n,t){t.sphere()},Point:function(n,t){n=n.coordinates,t.point(n[0],n[1],n[2])},MultiPoint:function(n,t){for(var e=n.coordinates,r=-1,i=e.length;++r<i;)n=e[r],t.point(n[0],n[1],n[2])},LineString:function(n,t){xe(n.coordinates,t,0)},MultiLineString:function(n,t){for(var e=n.coordinates,r=-1,i=e.length;++r<i;)xe(e[r],t,0)},Polygon:function(n,t){Me(n.coordinates,t)},MultiPolygon:function(n,t){for(var e=n.coordinates,r=-1,i=e.length;++r<i;)Me(e[r],t)},GeometryCollection:function(n,t){for(var e=n.geometries,r=-1,i=e.length;++r<i;)ve(e[r],t)}};function xe(n,t,e){var r,i=-1,o=n.length-e;for(t.lineStart();++i<o;)r=n[i],t.point(r[0],r[1],r[2]);t.lineEnd()}function Me(n,t){var e=-1,r=n.length;for(t.polygonStart();++e<r;)xe(n[e],t,1);t.polygonEnd()}o.geo.area=function(n){return be=0,o.geo.stream(n,De),be};var be,we,_e,Se,ke,Ee,Ne,Le,Ce,Ae,Te,Re,ze=new pe,De={sphere:function(){be+=4*Cn},point:P,lineStart:P,lineEnd:P,polygonStart:function(){ze.reset(),De.lineStart=qe},polygonEnd:function(){var n=2*ze;be+=n<0?4*Cn+n:n,De.lineStart=De.lineEnd=De.point=P}};function qe(){var n,t,e,r,i;function o(n,t){t=t*zn/2+Cn/4;var o=(n*=zn)-e,u=o>=0?1:-1,a=u*o,l=Math.cos(t),c=Math.sin(t),s=i*c,f=r*l+s*Math.cos(a),h=s*u*Math.sin(a);ze.add(Math.atan2(h,f)),e=n,r=l,i=c}De.point=function(u,a){De.point=o,e=(n=u)*zn,r=Math.cos(a=(t=a)*zn/2+Cn/4),i=Math.sin(a)},De.lineEnd=function(){o(n,t)}}function Oe(n){var t=n[0],e=n[1],r=Math.cos(e);return[r*Math.cos(t),r*Math.sin(t),Math.sin(e)]}function Pe(n,t){return n[0]*t[0]+n[1]*t[1]+n[2]*t[2]}function Ue(n,t){return[n[1]*t[2]-n[2]*t[1],n[2]*t[0]-n[0]*t[2],n[0]*t[1]-n[1]*t[0]]}function He(n,t){n[0]+=t[0],n[1]+=t[1],n[2]+=t[2]}function je(n,t){return[n[0]*t,n[1]*t,n[2]*t]}function Ie(n){var t=Math.sqrt(n[0]*n[0]+n[1]*n[1]+n[2]*n[2]);n[0]/=t,n[1]/=t,n[2]/=t}function $e(n){return[Math.atan2(n[1],n[0]),Un(n[2])]}function Fe(n,t){return w(n[0]-t[0])<Nn&&w(n[1]-t[1])<Nn}o.geo.bounds=function(){var n,t,e,r,i,u,a,l,c,s,f,h={point:p,lineStart:g,lineEnd:v,polygonStart:function(){h.point=m,h.lineStart=y,h.lineEnd=x,c=0,De.polygonStart()},polygonEnd:function(){De.polygonEnd(),h.point=p,h.lineStart=g,h.lineEnd=v,ze<0?(n=-(e=180),t=-(r=90)):c>Nn?r=90:c<-Nn&&(t=-90),f[0]=n,f[1]=e}};function p(i,o){s.push(f=[n=i,e=i]),o<t&&(t=o),o>r&&(r=o)}function d(o,u){var a=Oe([o*zn,u*zn]);if(l){var c=Ue(l,a),s=Ue([c[1],-c[0],0],c);Ie(s),s=$e(s);var f=o-i,h=f>0?1:-1,d=s[0]*Dn*h,g=w(f)>180;if(g^(h*i<d&&d<h*o))(v=s[1]*Dn)>r&&(r=v);else if(g^(h*i<(d=(d+360)%360-180)&&d<h*o)){var v;(v=-s[1]*Dn)<t&&(t=v)}else u<t&&(t=u),u>r&&(r=u);g?o<i?M(n,o)>M(n,e)&&(e=o):M(o,e)>M(n,e)&&(n=o):e>=n?(o<n&&(n=o),o>e&&(e=o)):o>i?M(n,o)>M(n,e)&&(e=o):M(o,e)>M(n,e)&&(n=o)}else p(o,u);l=a,i=o}function g(){h.point=d}function v(){f[0]=n,f[1]=e,h.point=p,l=null}function m(n,t){if(l){var e=n-i;c+=w(e)>180?e+(e>0?360:-360):e}else u=n,a=t;De.point(n,t),d(n,t)}function y(){De.lineStart()}function x(){m(u,a),De.lineEnd(),w(c)>Nn&&(n=-(e=180)),f[0]=n,f[1]=e,l=null}function M(n,t){return(t-=n)<0?t+360:t}function b(n,t){return n[0]-t[0]}function _(n,t){return t[0]<=t[1]?t[0]<=n&&n<=t[1]:n<t[0]||t[1]<n}return function(i){if(r=e=-(n=t=1/0),s=[],o.geo.stream(i,h),c=s.length){s.sort(b);for(var u=1,a=[g=s[0]];u<c;++u)_((p=s[u])[0],g)||_(p[1],g)?(M(g[0],p[1])>M(g[0],g[1])&&(g[1]=p[1]),M(p[0],g[1])>M(g[0],g[1])&&(g[0]=p[0])):a.push(g=p);for(var l,c,p,d=-1/0,g=(u=0,a[c=a.length-1]);u<=c;g=p,++u)p=a[u],(l=M(g[1],p[0]))>d&&(d=l,n=p[0],e=g[1])}return s=f=null,n===1/0||t===1/0?[[NaN,NaN],[NaN,NaN]]:[[n,t],[e,r]]}}(),o.geo.centroid=function(n){we=_e=Se=ke=Ee=Ne=Le=Ce=Ae=Te=Re=0,o.geo.stream(n,Be);var t=Ae,e=Te,r=Re,i=t*t+e*e+r*r;return i<Ln&&(t=Ne,e=Le,r=Ce,_e<Nn&&(t=Se,e=ke,r=Ee),(i=t*t+e*e+r*r)<Ln)?[NaN,NaN]:[Math.atan2(e,t)*Dn,Un(r/Math.sqrt(i))*Dn]};var Be={sphere:P,point:Ge,lineStart:Ze,lineEnd:Ve,polygonStart:function(){Be.lineStart=Xe},polygonEnd:function(){Be.lineStart=Ze}};function Ge(n,t){n*=zn;var e=Math.cos(t*=zn);Ye(e*Math.cos(n),e*Math.sin(n),Math.sin(t))}function Ye(n,t,e){++we,Se+=(n-Se)/we,ke+=(t-ke)/we,Ee+=(e-Ee)/we}function Ze(){var n,t,e;function r(r,i){r*=zn;var o=Math.cos(i*=zn),u=o*Math.cos(r),a=o*Math.sin(r),l=Math.sin(i),c=Math.atan2(Math.sqrt((c=t*l-e*a)*c+(c=e*u-n*l)*c+(c=n*a-t*u)*c),n*u+t*a+e*l);_e+=c,Ne+=c*(n+(n=u)),Le+=c*(t+(t=a)),Ce+=c*(e+(e=l)),Ye(n,t,e)}Be.point=function(i,o){i*=zn;var u=Math.cos(o*=zn);n=u*Math.cos(i),t=u*Math.sin(i),e=Math.sin(o),Be.point=r,Ye(n,t,e)}}function Ve(){Be.point=Ge}function Xe(){var n,t,e,r,i;function o(n,t){n*=zn;var o=Math.cos(t*=zn),u=o*Math.cos(n),a=o*Math.sin(n),l=Math.sin(t),c=r*l-i*a,s=i*u-e*l,f=e*a-r*u,h=Math.sqrt(c*c+s*s+f*f),p=e*u+r*a+i*l,d=h&&-Pn(p)/h,g=Math.atan2(h,p);Ae+=d*c,Te+=d*s,Re+=d*f,_e+=g,Ne+=g*(e+(e=u)),Le+=g*(r+(r=a)),Ce+=g*(i+(i=l)),Ye(e,r,i)}Be.point=function(u,a){n=u,t=a,Be.point=o,u*=zn;var l=Math.cos(a*=zn);e=l*Math.cos(u),r=l*Math.sin(u),i=Math.sin(a),Ye(e,r,i)},Be.lineEnd=function(){o(n,t),Be.lineEnd=Ve,Be.point=Ge}}function We(n,t){function e(e,r){return e=n(e,r),t(e[0],e[1])}return n.invert&&t.invert&&(e.invert=function(e,r){return(e=t.invert(e,r))&&n.invert(e[0],e[1])}),e}function Je(){return!0}function Ke(n,t,e,r,i){var o=[],u=[];if(n.forEach(function(n){if(!((t=n.length-1)<=0)){var t,e=n[0],r=n[t];if(Fe(e,r)){i.lineStart();for(var a=0;a<t;++a)i.point((e=n[a])[0],e[1]);i.lineEnd()}else{var l=new nr(e,n,null,!0),c=new nr(e,null,l,!1);l.o=c,o.push(l),u.push(c),l=new nr(r,n,null,!1),c=new nr(r,null,l,!0),l.o=c,o.push(l),u.push(c)}}}),u.sort(t),Qe(o),Qe(u),o.length){for(var a=0,l=e,c=u.length;a<c;++a)u[a].e=l=!l;for(var s,f,h=o[0];;){for(var p=h,d=!0;p.v;)if((p=p.n)===h)return;s=p.z,i.lineStart();do{if(p.v=p.o.v=!0,p.e){if(d)for(a=0,c=s.length;a<c;++a)i.point((f=s[a])[0],f[1]);else r(p.x,p.n.x,1,i);p=p.n}else{if(d)for(a=(s=p.p.z).length-1;a>=0;--a)i.point((f=s[a])[0],f[1]);else r(p.x,p.p.x,-1,i);p=p.p}s=(p=p.o).z,d=!d}while(!p.v);i.lineEnd()}}}function Qe(n){if(t=n.length){for(var t,e,r=0,i=n[0];++r<t;)i.n=e=n[r],e.p=i,i=e;i.n=e=n[0],e.p=i}}function nr(n,t,e,r){this.x=n,this.z=t,this.o=e,this.e=r,this.v=!1,this.n=this.p=null}function tr(n,t,e,r){return function(i,u){var a,l=t(u),c=i.invert(r[0],r[1]),s={point:f,lineStart:p,lineEnd:d,polygonStart:function(){s.point=M,s.lineStart=b,s.lineEnd=w,a=[],g=[]},polygonEnd:function(){s.point=f,s.lineStart=p,s.lineEnd=d,a=o.merge(a);var n=function(n,t){var e=n[0],r=n[1],i=[Math.sin(e),-Math.cos(e),0],o=0,u=0;ze.reset();for(var a=0,l=t.length;a<l;++a){var c=t[a],s=c.length;if(s)for(var f=c[0],h=f[0],p=f[1]/2+Cn/4,d=Math.sin(p),g=Math.cos(p),v=1;;){v===s&&(v=0);var m=(n=c[v])[0],y=n[1]/2+Cn/4,x=Math.sin(y),M=Math.cos(y),b=m-h,w=b>=0?1:-1,_=w*b,S=_>Cn,k=d*x;if(ze.add(Math.atan2(k*w*Math.sin(_),g*M+k*Math.cos(_))),o+=S?b+w*An:b,S^h>=e^m>=e){var E=Ue(Oe(f),Oe(n));Ie(E);var N=Ue(i,E);Ie(N);var L=(S^b>=0?-1:1)*Un(N[2]);(r>L||r===L&&(E[0]||E[1]))&&(u+=S^b>=0?1:-1)}if(!v++)break;h=m,d=x,g=M,f=n}}return(o<-Nn||o<Nn&&ze<-Nn)^1&u}(c,g);a.length?(x||(u.polygonStart(),x=!0),Ke(a,ir,n,e,u)):n&&(x||(u.polygonStart(),x=!0),u.lineStart(),e(null,null,1,u),u.lineEnd()),x&&(u.polygonEnd(),x=!1),a=g=null},sphere:function(){u.polygonStart(),u.lineStart(),e(null,null,1,u),u.lineEnd(),u.polygonEnd()}};function f(t,e){var r=i(t,e);n(t=r[0],e=r[1])&&u.point(t,e)}function h(n,t){var e=i(n,t);l.point(e[0],e[1])}function p(){s.point=h,l.lineStart()}function d(){s.point=f,l.lineEnd()}var g,v,m=rr(),y=t(m),x=!1;function M(n,t){v.push([n,t]);var e=i(n,t);y.point(e[0],e[1])}function b(){y.lineStart(),v=[]}function w(){M(v[0][0],v[0][1]),y.lineEnd();var n,t=y.clean(),e=m.buffer(),r=e.length;if(v.pop(),g.push(v),v=null,r)if(1&t){var i,o=-1;if((r=(n=e[0]).length-1)>0){for(x||(u.polygonStart(),x=!0),u.lineStart();++o<r;)u.point((i=n[o])[0],i[1]);u.lineEnd()}}else r>1&&2&t&&e.push(e.pop().concat(e.shift())),a.push(e.filter(er))}return s}}function er(n){return n.length>1}function rr(){var n,t=[];return{lineStart:function(){t.push(n=[])},point:function(t,e){n.push([t,e])},lineEnd:P,buffer:function(){var e=t;return t=[],n=null,e},rejoin:function(){t.length>1&&t.push(t.pop().concat(t.shift()))}}}function ir(n,t){return((n=n.x)[0]<0?n[1]-Rn-Nn:Rn-n[1])-((t=t.x)[0]<0?t[1]-Rn-Nn:Rn-t[1])}var or=tr(Je,function(n){var t,e=NaN,r=NaN,i=NaN;return{lineStart:function(){n.lineStart(),t=1},point:function(o,u){var a=o>0?Cn:-Cn,l=w(o-e);w(l-Cn)<Nn?(n.point(e,r=(r+u)/2>0?Rn:-Rn),n.point(i,r),n.lineEnd(),n.lineStart(),n.point(a,r),n.point(o,r),t=0):i!==a&&l>=Cn&&(w(e-i)<Nn&&(e-=i*Nn),w(o-a)<Nn&&(o-=a*Nn),r=function(n,t,e,r){var i,o,u=Math.sin(n-e);return w(u)>Nn?Math.atan((Math.sin(t)*(o=Math.cos(r))*Math.sin(e)-Math.sin(r)*(i=Math.cos(t))*Math.sin(n))/(i*o*u)):(t+r)/2}(e,r,o,u),n.point(i,r),n.lineEnd(),n.lineStart(),n.point(a,r),t=0),n.point(e=o,r=u),i=a},lineEnd:function(){n.lineEnd(),e=r=NaN},clean:function(){return 2-t}}},function(n,t,e,r){var i;if(null==n)i=e*Rn,r.point(-Cn,i),r.point(0,i),r.point(Cn,i),r.point(Cn,0),r.point(Cn,-i),r.point(0,-i),r.point(-Cn,-i),r.point(-Cn,0),r.point(-Cn,i);else if(w(n[0]-t[0])>Nn){var o=n[0]<t[0]?Cn:-Cn;i=e*o/2,r.point(-o,i),r.point(0,i),r.point(o,i)}else r.point(t[0],t[1])},[-Cn,-Cn/2]);function ur(n,t,e,r){return function(i){var o,u=i.a,a=i.b,l=u.x,c=u.y,s=0,f=1,h=a.x-l,p=a.y-c;if(o=n-l,h||!(o>0)){if(o/=h,h<0){if(o<s)return;o<f&&(f=o)}else if(h>0){if(o>f)return;o>s&&(s=o)}if(o=e-l,h||!(o<0)){if(o/=h,h<0){if(o>f)return;o>s&&(s=o)}else if(h>0){if(o<s)return;o<f&&(f=o)}if(o=t-c,p||!(o>0)){if(o/=p,p<0){if(o<s)return;o<f&&(f=o)}else if(p>0){if(o>f)return;o>s&&(s=o)}if(o=r-c,p||!(o<0)){if(o/=p,p<0){if(o>f)return;o>s&&(s=o)}else if(p>0){if(o<s)return;o<f&&(f=o)}return s>0&&(i.a={x:l+s*h,y:c+s*p}),f<1&&(i.b={x:l+f*h,y:c+f*p}),i}}}}}}var ar=1e9;function lr(n,t,e,r){return function(l){var c,s,f,h,p,d,g,v,m,y,x,M=l,b=rr(),w=ur(n,t,e,r),_={point:E,lineStart:function(){_.point=N,s&&s.push(f=[]),y=!0,m=!1,g=v=NaN},lineEnd:function(){c&&(N(h,p),d&&m&&b.rejoin(),c.push(b.buffer())),_.point=E,m&&l.lineEnd()},polygonStart:function(){l=b,c=[],s=[],x=!0},polygonEnd:function(){l=M,c=o.merge(c);var t=function(n){for(var t=0,e=s.length,r=n[1],i=0;i<e;++i)for(var o,u=1,a=s[i],l=a.length,c=a[0];u<l;++u)o=a[u],c[1]<=r?o[1]>r&&On(c,o,n)>0&&++t:o[1]<=r&&On(c,o,n)<0&&--t,c=o;return 0!==t}([n,r]),e=x&&t,i=c.length;(e||i)&&(l.polygonStart(),e&&(l.lineStart(),S(null,null,1,l),l.lineEnd()),i&&Ke(c,u,t,S,l),l.polygonEnd()),c=s=f=null}};function S(o,u,l,c){var s=0,f=0;if(null==o||(s=i(o,l))!==(f=i(u,l))||a(o,u)<0^l>0)do{c.point(0===s||3===s?n:e,s>1?r:t)}while((s=(s+l+4)%4)!==f);else c.point(u[0],u[1])}function k(i,o){return n<=i&&i<=e&&t<=o&&o<=r}function E(n,t){k(n,t)&&l.point(n,t)}function N(n,t){var e=k(n=Math.max(-1e9,Math.min(ar,n)),t=Math.max(-1e9,Math.min(ar,t)));if(s&&f.push([n,t]),y)h=n,p=t,d=e,y=!1,e&&(l.lineStart(),l.point(n,t));else if(e&&m)l.point(n,t);else{var r={a:{x:g,y:v},b:{x:n,y:t}};w(r)?(m||(l.lineStart(),l.point(r.a.x,r.a.y)),l.point(r.b.x,r.b.y),e||l.lineEnd(),x=!1):e&&(l.lineStart(),l.point(n,t),x=!1)}g=n,v=t,m=e}return _};function i(r,i){return w(r[0]-n)<Nn?i>0?0:3:w(r[0]-e)<Nn?i>0?2:1:w(r[1]-t)<Nn?i>0?1:0:i>0?3:2}function u(n,t){return a(n.x,t.x)}function a(n,t){var e=i(n,1),r=i(t,1);return e!==r?e-r:0===e?t[1]-n[1]:1===e?n[0]-t[0]:2===e?n[1]-t[1]:t[0]-n[0]}}function cr(n){var t=0,e=Cn/3,r=zr(n),i=r(t,e);return i.parallels=function(n){return arguments.length?r(t=n[0]*Cn/180,e=n[1]*Cn/180):[t/Cn*180,e/Cn*180]},i}function sr(n,t){var e=Math.sin(n),r=(e+Math.sin(t))/2,i=1+e*(2*r-e),o=Math.sqrt(i)/r;function u(n,t){var e=Math.sqrt(i-2*r*Math.sin(t))/r;return[e*Math.sin(n*=r),o-e*Math.cos(n)]}return u.invert=function(n,t){var e=o-t;return[Math.atan2(n,e)/r,Un((i-(n*n+e*e)*r*r)/(2*r))]},u}o.geo.clipExtent=function(){var n,t,e,r,i,o,u={stream:function(n){return i&&(i.valid=!1),(i=o(n)).valid=!0,i},extent:function(a){return arguments.length?(o=lr(n=+a[0][0],t=+a[0][1],e=+a[1][0],r=+a[1][1]),i&&(i.valid=!1,i=null),u):[[n,t],[e,r]]}};return u.extent([[0,0],[960,500]])},(o.geo.conicEqualArea=function(){return cr(sr)}).raw=sr,o.geo.albers=function(){return o.geo.conicEqualArea().rotate([96,0]).center([-.6,38.7]).parallels([29.5,45.5]).scale(1070)},o.geo.albersUsa=function(){var n,t,e,r,i=o.geo.albers(),u=o.geo.conicEqualArea().rotate([154,0]).center([-2,58.5]).parallels([55,65]),a=o.geo.conicEqualArea().rotate([157,0]).center([-3,19.9]).parallels([8,18]),l={point:function(t,e){n=[t,e]}};function c(i){var o=i[0],u=i[1];return n=null,t(o,u),n||(e(o,u),n)||r(o,u),n}return c.invert=function(n){var t=i.scale(),e=i.translate(),r=(n[0]-e[0])/t,o=(n[1]-e[1])/t;return(o>=.12&&o<.234&&r>=-.425&&r<-.214?u:o>=.166&&o<.234&&r>=-.214&&r<-.115?a:i).invert(n)},c.stream=function(n){var t=i.stream(n),e=u.stream(n),r=a.stream(n);return{point:function(n,i){t.point(n,i),e.point(n,i),r.point(n,i)},sphere:function(){t.sphere(),e.sphere(),r.sphere()},lineStart:function(){t.lineStart(),e.lineStart(),r.lineStart()},lineEnd:function(){t.lineEnd(),e.lineEnd(),r.lineEnd()},polygonStart:function(){t.polygonStart(),e.polygonStart(),r.polygonStart()},polygonEnd:function(){t.polygonEnd(),e.polygonEnd(),r.polygonEnd()}}},c.precision=function(n){return arguments.length?(i.precision(n),u.precision(n),a.precision(n),c):i.precision()},c.scale=function(n){return arguments.length?(i.scale(n),u.scale(.35*n),a.scale(n),c.translate(i.translate())):i.scale()},c.translate=function(n){if(!arguments.length)return i.translate();var o=i.scale(),s=+n[0],f=+n[1];return t=i.translate(n).clipExtent([[s-.455*o,f-.238*o],[s+.455*o,f+.238*o]]).stream(l).point,e=u.translate([s-.307*o,f+.201*o]).clipExtent([[s-.425*o+Nn,f+.12*o+Nn],[s-.214*o-Nn,f+.234*o-Nn]]).stream(l).point,r=a.translate([s-.205*o,f+.212*o]).clipExtent([[s-.214*o+Nn,f+.166*o+Nn],[s-.115*o-Nn,f+.234*o-Nn]]).stream(l).point,c},c.scale(1070)};var fr,hr,pr,dr,gr,vr,mr={point:P,lineStart:P,lineEnd:P,polygonStart:function(){hr=0,mr.lineStart=yr},polygonEnd:function(){mr.lineStart=mr.lineEnd=mr.point=P,fr+=w(hr/2)}};function yr(){var n,t,e,r;function i(n,t){hr+=r*n-e*t,e=n,r=t}mr.point=function(o,u){mr.point=i,n=e=o,t=r=u},mr.lineEnd=function(){i(n,t)}}var xr={point:function(n,t){n<pr&&(pr=n),n>gr&&(gr=n),t<dr&&(dr=t),t>vr&&(vr=t)},lineStart:P,lineEnd:P,polygonStart:P,polygonEnd:P};function Mr(){var n=br(4.5),t=[],e={point:r,lineStart:function(){e.point=i},lineEnd:u,polygonStart:function(){e.lineEnd=a},polygonEnd:function(){e.lineEnd=u,e.point=r},pointRadius:function(t){return n=br(t),e},result:function(){if(t.length){var n=t.join("");return t=[],n}}};function r(e,r){t.push("M",e,",",r,n)}function i(n,r){t.push("M",n,",",r),e.point=o}function o(n,e){t.push("L",n,",",e)}function u(){e.point=r}function a(){t.push("Z")}return e}function br(n){return"m0,"+n+"a"+n+","+n+" 0 1,1 0,"+-2*n+"a"+n+","+n+" 0 1,1 0,"+2*n+"z"}var wr,_r={point:Sr,lineStart:kr,lineEnd:Er,polygonStart:function(){_r.lineStart=Nr},polygonEnd:function(){_r.point=Sr,_r.lineStart=kr,_r.lineEnd=Er}};function Sr(n,t){Se+=n,ke+=t,++Ee}function kr(){var n,t;function e(e,r){var i=e-n,o=r-t,u=Math.sqrt(i*i+o*o);Ne+=u*(n+e)/2,Le+=u*(t+r)/2,Ce+=u,Sr(n=e,t=r)}_r.point=function(r,i){_r.point=e,Sr(n=r,t=i)}}function Er(){_r.point=Sr}function Nr(){var n,t,e,r;function i(n,t){var i=n-e,o=t-r,u=Math.sqrt(i*i+o*o);Ne+=u*(e+n)/2,Le+=u*(r+t)/2,Ce+=u,Ae+=(u=r*n-e*t)*(e+n),Te+=u*(r+t),Re+=3*u,Sr(e=n,r=t)}_r.point=function(o,u){_r.point=i,Sr(n=e=o,t=r=u)},_r.lineEnd=function(){i(n,t)}}function Lr(n){var t=4.5,e={point:r,lineStart:function(){e.point=i},lineEnd:u,polygonStart:function(){e.lineEnd=a},polygonEnd:function(){e.lineEnd=u,e.point=r},pointRadius:function(n){return t=n,e},result:P};function r(e,r){n.moveTo(e+t,r),n.arc(e,r,t,0,An)}function i(t,r){n.moveTo(t,r),e.point=o}function o(t,e){n.lineTo(t,e)}function u(){e.point=r}function a(){n.closePath()}return e}function Cr(n){var t=.5,e=Math.cos(30*zn),r=16;function i(n){return(r?u:o)(n)}function o(t){return Tr(t,function(e,r){e=n(e,r),t.point(e[0],e[1])})}function u(t){var e,i,o,u,l,c,s,f,h,p,d,g,v={point:m,lineStart:y,lineEnd:M,polygonStart:function(){t.polygonStart(),v.lineStart=b},polygonEnd:function(){t.polygonEnd(),v.lineStart=y}};function m(e,r){e=n(e,r),t.point(e[0],e[1])}function y(){f=NaN,v.point=x,t.lineStart()}function x(e,i){var o=Oe([e,i]),u=n(e,i);a(f,h,s,p,d,g,f=u[0],h=u[1],s=e,p=o[0],d=o[1],g=o[2],r,t),t.point(f,h)}function M(){v.point=m,t.lineEnd()}function b(){y(),v.point=w,v.lineEnd=_}function w(n,t){x(e=n,t),i=f,o=h,u=p,l=d,c=g,v.point=x}function _(){a(f,h,s,p,d,g,i,o,e,u,l,c,r,t),v.lineEnd=M,M()}return v}function a(r,i,o,u,l,c,s,f,h,p,d,g,v,m){var y=s-r,x=f-i,M=y*y+x*x;if(M>4*t&&v--){var b=u+p,_=l+d,S=c+g,k=Math.sqrt(b*b+_*_+S*S),E=Math.asin(S/=k),N=w(w(S)-1)<Nn||w(o-h)<Nn?(o+h)/2:Math.atan2(_,b),L=n(N,E),C=L[0],A=L[1],T=C-r,R=A-i,z=x*T-y*R;(z*z/M>t||w((y*T+x*R)/M-.5)>.3||u*p+l*d+c*g<e)&&(a(r,i,o,u,l,c,C,A,N,b/=k,_/=k,S,v,m),m.point(C,A),a(C,A,N,b,_,S,s,f,h,p,d,g,v,m))}}return i.precision=function(n){return arguments.length?(r=(t=n*n)>0&&16,i):Math.sqrt(t)},i}function Ar(n){this.stream=n}function Tr(n,t){return{point:t,sphere:function(){n.sphere()},lineStart:function(){n.lineStart()},lineEnd:function(){n.lineEnd()},polygonStart:function(){n.polygonStart()},polygonEnd:function(){n.polygonEnd()}}}function Rr(n){return zr(function(){return n})()}function zr(n){var t,e,r,i,u,a,l=Cr(function(n,e){return[(n=t(n,e))[0]*c+i,u-n[1]*c]}),c=150,s=480,f=250,h=0,p=0,d=0,g=0,v=0,m=or,y=z,x=null,M=null;function b(n){return[(n=r(n[0]*zn,n[1]*zn))[0]*c+i,u-n[1]*c]}function _(n){return(n=r.invert((n[0]-i)/c,(u-n[1])/c))&&[n[0]*Dn,n[1]*Dn]}function S(){r=We(e=Pr(d,g,v),t);var n=t(h,p);return i=s-n[0]*c,u=f+n[1]*c,k()}function k(){return a&&(a.valid=!1,a=null),b}return b.stream=function(n){return a&&(a.valid=!1),(a=Dr(m(e,l(y(n))))).valid=!0,a},b.clipAngle=function(n){return arguments.length?(m=null==n?(x=n,or):function(n){var t=Math.cos(n),e=t>0,r=w(t)>Nn;return tr(i,function(n){var t,a,l,c,s;return{lineStart:function(){c=l=!1,s=1},point:function(f,h){var p,d=[f,h],g=i(f,h),v=e?g?0:u(f,h):g?u(f+(f<0?Cn:-Cn),h):0;if(!t&&(c=l=g)&&n.lineStart(),g!==l&&(p=o(t,d),(Fe(t,p)||Fe(d,p))&&(d[0]+=Nn,d[1]+=Nn,g=i(d[0],d[1]))),g!==l)s=0,g?(n.lineStart(),p=o(d,t),n.point(p[0],p[1])):(p=o(t,d),n.point(p[0],p[1]),n.lineEnd()),t=p;else if(r&&t&&e^g){var m;v&a||!(m=o(d,t,!0))||(s=0,e?(n.lineStart(),n.point(m[0][0],m[0][1]),n.point(m[1][0],m[1][1]),n.lineEnd()):(n.point(m[1][0],m[1][1]),n.lineEnd(),n.lineStart(),n.point(m[0][0],m[0][1])))}!g||t&&Fe(t,d)||n.point(d[0],d[1]),t=d,l=g,a=v},lineEnd:function(){l&&n.lineEnd(),t=null},clean:function(){return s|(c&&l)<<1}}},Ir(n,6*zn),e?[0,-n]:[-Cn,n-Cn]);function i(n,e){return Math.cos(n)*Math.cos(e)>t}function o(n,e,r){var i=[1,0,0],o=Ue(Oe(n),Oe(e)),u=Pe(o,o),a=o[0],l=u-a*a;if(!l)return!r&&n;var c=t*u/l,s=-t*a/l,f=Ue(i,o),h=je(i,c);He(h,je(o,s));var p=f,d=Pe(h,p),g=Pe(p,p),v=d*d-g*(Pe(h,h)-1);if(!(v<0)){var m=Math.sqrt(v),y=je(p,(-d-m)/g);if(He(y,h),y=$e(y),!r)return y;var x,M=n[0],b=e[0],_=n[1],S=e[1];b<M&&(x=M,M=b,b=x);var k=b-M,E=w(k-Cn)<Nn;if(!E&&S<_&&(x=_,_=S,S=x),E||k<Nn?E?_+S>0^y[1]<(w(y[0]-M)<Nn?_:S):_<=y[1]&&y[1]<=S:k>Cn^(M<=y[0]&&y[0]<=b)){var N=je(p,(-d+m)/g);return He(N,h),[y,$e(N)]}}}function u(t,r){var i=e?n:Cn-n,o=0;return t<-i?o|=1:t>i&&(o|=2),r<-i?o|=4:r>i&&(o|=8),o}}((x=+n)*zn),k()):x},b.clipExtent=function(n){return arguments.length?(M=n,y=n?lr(n[0][0],n[0][1],n[1][0],n[1][1]):z,k()):M},b.scale=function(n){return arguments.length?(c=+n,S()):c},b.translate=function(n){return arguments.length?(s=+n[0],f=+n[1],S()):[s,f]},b.center=function(n){return arguments.length?(h=n[0]%360*zn,p=n[1]%360*zn,S()):[h*Dn,p*Dn]},b.rotate=function(n){return arguments.length?(d=n[0]%360*zn,g=n[1]%360*zn,v=n.length>2?n[2]%360*zn:0,S()):[d*Dn,g*Dn,v*Dn]},o.rebind(b,l,"precision"),function(){return t=n.apply(this,arguments),b.invert=t.invert&&_,S()}}function Dr(n){return Tr(n,function(t,e){n.point(t*zn,e*zn)})}function qr(n,t){return[n,t]}function Or(n,t){return[n>Cn?n-An:n<-Cn?n+An:n,t]}function Pr(n,t,e){return n?t||e?We(Hr(n),jr(t,e)):Hr(n):t||e?jr(t,e):Or}function Ur(n){return function(t,e){return[(t+=n)>Cn?t-An:t<-Cn?t+An:t,e]}}function Hr(n){var t=Ur(n);return t.invert=Ur(-n),t}function jr(n,t){var e=Math.cos(n),r=Math.sin(n),i=Math.cos(t),o=Math.sin(t);function u(n,t){var u=Math.cos(t),a=Math.cos(n)*u,l=Math.sin(n)*u,c=Math.sin(t),s=c*e+a*r;return[Math.atan2(l*i-s*o,a*e-c*r),Un(s*i+l*o)]}return u.invert=function(n,t){var u=Math.cos(t),a=Math.cos(n)*u,l=Math.sin(n)*u,c=Math.sin(t),s=c*i-l*o;return[Math.atan2(l*i+c*o,a*e+s*r),Un(s*e-a*r)]},u}function Ir(n,t){var e=Math.cos(n),r=Math.sin(n);return function(i,o,u,a){var l=u*t;null!=i?(i=$r(e,i),o=$r(e,o),(u>0?i<o:i>o)&&(i+=u*An)):(i=n+u*An,o=n-.5*l);for(var c,s=i;u>0?s>o:s<o;s-=l)a.point((c=$e([e,-r*Math.cos(s),-r*Math.sin(s)]))[0],c[1])}}function $r(n,t){var e=Oe(t);e[0]-=n,Ie(e);var r=Pn(-e[1]);return((-e[2]<0?-r:r)+2*Math.PI-Nn)%(2*Math.PI)}function Fr(n,t,e){var r=o.range(n,t-Nn,e).concat(t);return function(n){return r.map(function(t){return[n,t]})}}function Br(n,t,e){var r=o.range(n,t-Nn,e).concat(t);return function(n){return r.map(function(t){return[t,n]})}}function Gr(n){return n.source}function Yr(n){return n.target}o.geo.path=function(){var n,t,e,r,i,u=4.5;function a(n){return n&&("function"==typeof u&&r.pointRadius(+u.apply(this,arguments)),i&&i.valid||(i=e(r)),o.geo.stream(n,i)),r.result()}function l(){return i=null,a}return a.area=function(n){return fr=0,o.geo.stream(n,e(mr)),fr},a.centroid=function(n){return Se=ke=Ee=Ne=Le=Ce=Ae=Te=Re=0,o.geo.stream(n,e(_r)),Re?[Ae/Re,Te/Re]:Ce?[Ne/Ce,Le/Ce]:Ee?[Se/Ee,ke/Ee]:[NaN,NaN]},a.bounds=function(n){return gr=vr=-(pr=dr=1/0),o.geo.stream(n,e(xr)),[[pr,dr],[gr,vr]]},a.projection=function(t){return arguments.length?(e=(n=t)?t.stream||(r=t,i=Cr(function(n,t){return r([n*Dn,t*Dn])}),function(n){return Dr(i(n))}):z,l()):n;var r,i},a.context=function(n){return arguments.length?(r=null==(t=n)?new Mr:new Lr(n),"function"!=typeof u&&r.pointRadius(u),l()):t},a.pointRadius=function(n){return arguments.length?(u="function"==typeof n?n:(r.pointRadius(+n),+n),a):u},a.projection(o.geo.albersUsa()).context(null)},o.geo.transform=function(n){return{stream:function(t){var e=new Ar(t);for(var r in n)e[r]=n[r];return e}}},Ar.prototype={point:function(n,t){this.stream.point(n,t)},sphere:function(){this.stream.sphere()},lineStart:function(){this.stream.lineStart()},lineEnd:function(){this.stream.lineEnd()},polygonStart:function(){this.stream.polygonStart()},polygonEnd:function(){this.stream.polygonEnd()}},o.geo.projection=Rr,o.geo.projectionMutator=zr,(o.geo.equirectangular=function(){return Rr(qr)}).raw=qr.invert=qr,o.geo.rotation=function(n){function t(t){return(t=n(t[0]*zn,t[1]*zn))[0]*=Dn,t[1]*=Dn,t}return n=Pr(n[0]%360*zn,n[1]*zn,n.length>2?n[2]*zn:0),t.invert=function(t){return(t=n.invert(t[0]*zn,t[1]*zn))[0]*=Dn,t[1]*=Dn,t},t},Or.invert=qr,o.geo.circle=function(){var n,t,e=[0,0],r=6;function i(){var n="function"==typeof e?e.apply(this,arguments):e,r=Pr(-n[0]*zn,-n[1]*zn,0).invert,i=[];return t(null,null,1,{point:function(n,t){i.push(n=r(n,t)),n[0]*=Dn,n[1]*=Dn}}),{type:"Polygon",coordinates:[i]}}return i.origin=function(n){return arguments.length?(e=n,i):e},i.angle=function(e){return arguments.length?(t=Ir((n=+e)*zn,r*zn),i):n},i.precision=function(e){return arguments.length?(t=Ir(n*zn,(r=+e)*zn),i):r},i.angle(90)},o.geo.distance=function(n,t){var e,r=(t[0]-n[0])*zn,i=n[1]*zn,o=t[1]*zn,u=Math.sin(r),a=Math.cos(r),l=Math.sin(i),c=Math.cos(i),s=Math.sin(o),f=Math.cos(o);return Math.atan2(Math.sqrt((e=f*u)*e+(e=c*s-l*f*a)*e),l*s+c*f*a)},o.geo.graticule=function(){var n,t,e,r,i,u,a,l,c,s,f,h,p=10,d=p,g=90,v=360,m=2.5;function y(){return{type:"MultiLineString",coordinates:x()}}function x(){return o.range(Math.ceil(r/g)*g,e,g).map(f).concat(o.range(Math.ceil(l/v)*v,a,v).map(h)).concat(o.range(Math.ceil(t/p)*p,n,p).filter(function(n){return w(n%g)>Nn}).map(c)).concat(o.range(Math.ceil(u/d)*d,i,d).filter(function(n){return w(n%v)>Nn}).map(s))}return y.lines=function(){return x().map(function(n){return{type:"LineString",coordinates:n}})},y.outline=function(){return{type:"Polygon",coordinates:[f(r).concat(h(a).slice(1),f(e).reverse().slice(1),h(l).reverse().slice(1))]}},y.extent=function(n){return arguments.length?y.majorExtent(n).minorExtent(n):y.minorExtent()},y.majorExtent=function(n){return arguments.length?(r=+n[0][0],e=+n[1][0],l=+n[0][1],a=+n[1][1],r>e&&(n=r,r=e,e=n),l>a&&(n=l,l=a,a=n),y.precision(m)):[[r,l],[e,a]]},y.minorExtent=function(e){return arguments.length?(t=+e[0][0],n=+e[1][0],u=+e[0][1],i=+e[1][1],t>n&&(e=t,t=n,n=e),u>i&&(e=u,u=i,i=e),y.precision(m)):[[t,u],[n,i]]},y.step=function(n){return arguments.length?y.majorStep(n).minorStep(n):y.minorStep()},y.majorStep=function(n){return arguments.length?(g=+n[0],v=+n[1],y):[g,v]},y.minorStep=function(n){return arguments.length?(p=+n[0],d=+n[1],y):[p,d]},y.precision=function(o){return arguments.length?(m=+o,c=Fr(u,i,90),s=Br(t,n,m),f=Fr(l,a,90),h=Br(r,e,m),y):m},y.majorExtent([[-180,-90+Nn],[180,90-Nn]]).minorExtent([[-180,-80-Nn],[180,80+Nn]])},o.geo.greatArc=function(){var n,t,e=Gr,r=Yr;function i(){return{type:"LineString",coordinates:[n||e.apply(this,arguments),t||r.apply(this,arguments)]}}return i.distance=function(){return o.geo.distance(n||e.apply(this,arguments),t||r.apply(this,arguments))},i.source=function(t){return arguments.length?(e=t,n="function"==typeof t?null:t,i):e},i.target=function(n){return arguments.length?(r=n,t="function"==typeof n?null:n,i):r},i.precision=function(){return arguments.length?i:0},i},o.geo.interpolate=function(n,t){return e=n[0]*zn,r=n[1]*zn,i=t[0]*zn,o=t[1]*zn,u=Math.cos(r),a=Math.sin(r),l=Math.cos(o),c=Math.sin(o),s=u*Math.cos(e),f=u*Math.sin(e),h=l*Math.cos(i),p=l*Math.sin(i),d=2*Math.asin(Math.sqrt(jn(o-r)+u*l*jn(i-e))),g=1/Math.sin(d),(v=d?function(n){var t=Math.sin(n*=d)*g,e=Math.sin(d-n)*g,r=e*s+t*h,i=e*f+t*p,o=e*a+t*c;return[Math.atan2(i,r)*Dn,Math.atan2(o,Math.sqrt(r*r+i*i))*Dn]}:function(){return[e*Dn,r*Dn]}).distance=d,v;var e,r,i,o,u,a,l,c,s,f,h,p,d,g,v},o.geo.length=function(n){return wr=0,o.geo.stream(n,Zr),wr};var Zr={sphere:P,point:P,lineStart:function(){var n,t,e;function r(r,i){var o=Math.sin(i*=zn),u=Math.cos(i),a=w((r*=zn)-n),l=Math.cos(a);wr+=Math.atan2(Math.sqrt((a=u*Math.sin(a))*a+(a=e*o-t*u*l)*a),t*o+e*u*l),n=r,t=o,e=u}Zr.point=function(i,o){n=i*zn,t=Math.sin(o*=zn),e=Math.cos(o),Zr.point=r},Zr.lineEnd=function(){Zr.point=Zr.lineEnd=P}},lineEnd:P,polygonStart:P,polygonEnd:P};function Vr(n,t){function e(t,e){var r=Math.cos(t),i=Math.cos(e),o=n(r*i);return[o*i*Math.sin(t),o*Math.sin(e)]}return e.invert=function(n,e){var r=Math.sqrt(n*n+e*e),i=t(r),o=Math.sin(i),u=Math.cos(i);return[Math.atan2(n*o,r*u),Math.asin(r&&e*o/r)]},e}var Xr=Vr(function(n){return Math.sqrt(2/(1+n))},function(n){return 2*Math.asin(n/2)});(o.geo.azimuthalEqualArea=function(){return Rr(Xr)}).raw=Xr;var Wr=Vr(function(n){var t=Math.acos(n);return t&&t/Math.sin(t)},z);function Jr(n,t){var e=Math.cos(n),r=function(n){return Math.tan(Cn/4+n/2)},i=n===t?Math.sin(n):Math.log(e/Math.cos(t))/Math.log(r(t)/r(n)),o=e*Math.pow(r(n),i)/i;if(!i)return ni;function u(n,t){o>0?t<-Rn+Nn&&(t=-Rn+Nn):t>Rn-Nn&&(t=Rn-Nn);var e=o/Math.pow(r(t),i);return[e*Math.sin(i*n),o-e*Math.cos(i*n)]}return u.invert=function(n,t){var e=o-t,r=qn(i)*Math.sqrt(n*n+e*e);return[Math.atan2(n,e)/i,2*Math.atan(Math.pow(o/r,1/i))-Rn]},u}function Kr(n,t){var e=Math.cos(n),r=n===t?Math.sin(n):(e-Math.cos(t))/(t-n),i=e/r+n;if(w(r)<Nn)return qr;function o(n,t){var e=i-t;return[e*Math.sin(r*n),i-e*Math.cos(r*n)]}return o.invert=function(n,t){var e=i-t;return[Math.atan2(n,e)/r,i-qn(r)*Math.sqrt(n*n+e*e)]},o}(o.geo.azimuthalEquidistant=function(){return Rr(Wr)}).raw=Wr,(o.geo.conicConformal=function(){return cr(Jr)}).raw=Jr,(o.geo.conicEquidistant=function(){return cr(Kr)}).raw=Kr;var Qr=Vr(function(n){return 1/n},Math.atan);function ni(n,t){return[n,Math.log(Math.tan(Cn/4+t/2))]}function ti(n){var t,e=Rr(n),r=e.scale,i=e.translate,o=e.clipExtent;return e.scale=function(){var n=r.apply(e,arguments);return n===e?t?e.clipExtent(null):e:n},e.translate=function(){var n=i.apply(e,arguments);return n===e?t?e.clipExtent(null):e:n},e.clipExtent=function(n){var u=o.apply(e,arguments);if(u===e){if(t=null==n){var a=Cn*r(),l=i();o([[l[0]-a,l[1]-a],[l[0]+a,l[1]+a]])}}else t&&(u=null);return u},e.clipExtent(null)}(o.geo.gnomonic=function(){return Rr(Qr)}).raw=Qr,ni.invert=function(n,t){return[n,2*Math.atan(Math.exp(t))-Rn]},(o.geo.mercator=function(){return ti(ni)}).raw=ni;var ei=Vr(function(){return 1},Math.asin);(o.geo.orthographic=function(){return Rr(ei)}).raw=ei;var ri=Vr(function(n){return 1/(1+n)},function(n){return 2*Math.atan(n)});function ii(n,t){return[Math.log(Math.tan(Cn/4+t/2)),-n]}function oi(n){return n[0]}function ui(n){return n[1]}function ai(n){for(var t=n.length,e=[0,1],r=2,i=2;i<t;i++){for(;r>1&&On(n[e[r-2]],n[e[r-1]],n[i])<=0;)--r;e[r++]=i}return e.slice(0,r)}function li(n,t){return n[0]-t[0]||n[1]-t[1]}(o.geo.stereographic=function(){return Rr(ri)}).raw=ri,ii.invert=function(n,t){return[-t,2*Math.atan(Math.exp(n))-Rn]},(o.geo.transverseMercator=function(){var n=ti(ii),t=n.center,e=n.rotate;return n.center=function(n){return n?t([-n[1],n[0]]):[(n=t())[1],-n[0]]},n.rotate=function(n){return n?e([n[0],n[1],n.length>2?n[2]+90:90]):[(n=e())[0],n[1],n[2]-90]},e([0,0,90])}).raw=ii,o.geom={},o.geom.hull=function(n){var t=oi,e=ui;if(arguments.length)return r(n);function r(n){if(n.length<3)return[];var r,i=xt(t),o=xt(e),u=n.length,a=[],l=[];for(r=0;r<u;r++)a.push([+i.call(this,n[r],r),+o.call(this,n[r],r),r]);for(a.sort(li),r=0;r<u;r++)l.push([a[r][0],-a[r][1]]);var c=ai(a),s=ai(l),f=s[0]===c[0],h=s[s.length-1]===c[c.length-1],p=[];for(r=c.length-1;r>=0;--r)p.push(n[a[c[r]][2]]);for(r=+f;r<s.length-h;++r)p.push(n[a[s[r]][2]]);return p}return r.x=function(n){return arguments.length?(t=n,r):t},r.y=function(n){return arguments.length?(e=n,r):e},r},o.geom.polygon=function(n){return B(n,ci),n};var ci=o.geom.polygon.prototype=[];function si(n,t,e){return(e[0]-t[0])*(n[1]-t[1])<(e[1]-t[1])*(n[0]-t[0])}function fi(n,t,e,r){var i=n[0],o=e[0],u=t[0]-i,a=r[0]-o,l=n[1],c=e[1],s=t[1]-l,f=r[1]-c,h=(a*(l-c)-f*(i-o))/(f*u-a*s);return[i+h*u,l+h*s]}function hi(n){var t=n[0],e=n[n.length-1];return!(t[0]-e[0]||t[1]-e[1])}ci.area=function(){for(var n,t=-1,e=this.length,r=this[e-1],i=0;++t<e;)n=r,r=this[t],i+=n[1]*r[0]-n[0]*r[1];return.5*i},ci.centroid=function(n){var t,e,r=-1,i=this.length,o=0,u=0,a=this[i-1];for(arguments.length||(n=-1/(6*this.area()));++r<i;)t=a,a=this[r],e=t[0]*a[1]-a[0]*t[1],o+=(t[0]+a[0])*e,u+=(t[1]+a[1])*e;return[o*n,u*n]},ci.clip=function(n){for(var t,e,r,i,o,u,a=hi(n),l=-1,c=this.length-hi(this),s=this[c-1];++l<c;){for(t=n.slice(),n.length=0,i=this[l],o=t[(r=t.length-a)-1],e=-1;++e<r;)si(u=t[e],s,i)?(si(o,s,i)||n.push(fi(o,u,s,i)),n.push(u)):si(o,s,i)&&n.push(fi(o,u,s,i)),o=u;a&&n.push(n[0]),s=i}return n};var pi,di,gi,vi,mi,yi=[],xi=[];function Mi(){Hi(this),this.edge=this.site=this.circle=null}function bi(n){var t=yi.pop()||new Mi;return t.site=n,t}function wi(n){Ti(n),gi.remove(n),yi.push(n),Hi(n)}function _i(n){var t=n.circle,e=t.x,r=t.cy,i={x:e,y:r},o=n.P,u=n.N,a=[n];wi(n);for(var l=o;l.circle&&w(e-l.circle.x)<Nn&&w(r-l.circle.cy)<Nn;)o=l.P,a.unshift(l),wi(l),l=o;a.unshift(l),Ti(l);for(var c=u;c.circle&&w(e-c.circle.x)<Nn&&w(r-c.circle.cy)<Nn;)u=c.N,a.push(c),wi(c),c=u;a.push(c),Ti(c);var s,f=a.length;for(s=1;s<f;++s)c=a[s],l=a[s-1],Oi(c.edge,l.site,c.site,i);l=a[0],(c=a[f-1]).edge=Di(l.site,c.site,null,i),Ai(l),Ai(c)}function Si(n){for(var t,e,r,i,o=n.x,u=n.y,a=gi._;a;)if((r=ki(a,u)-o)>Nn)a=a.L;else{if(!((i=o-Ei(a,u))>Nn)){r>-Nn?(t=a.P,e=a):i>-Nn?(t=a,e=a.N):t=e=a;break}if(!a.R){t=a;break}a=a.R}var l=bi(n);if(gi.insert(t,l),t||e){if(t===e)return Ti(t),e=bi(t.site),gi.insert(l,e),l.edge=e.edge=Di(t.site,l.site),Ai(t),void Ai(e);if(e){Ti(t),Ti(e);var c=t.site,s=c.x,f=c.y,h=n.x-s,p=n.y-f,d=e.site,g=d.x-s,v=d.y-f,m=2*(h*v-p*g),y=h*h+p*p,x=g*g+v*v,M={x:(v*y-p*x)/m+s,y:(h*x-g*y)/m+f};Oi(e.edge,c,d,M),l.edge=Di(c,n,null,M),e.edge=Di(n,d,null,M),Ai(t),Ai(e)}else l.edge=Di(t.site,l.site)}}function ki(n,t){var e=n.site,r=e.x,i=e.y,o=i-t;if(!o)return r;var u=n.P;if(!u)return-1/0;var a=(e=u.site).x,l=e.y,c=l-t;if(!c)return a;var s=a-r,f=1/o-1/c,h=s/c;return f?(-h+Math.sqrt(h*h-2*f*(s*s/(-2*c)-l+c/2+i-o/2)))/f+r:(r+a)/2}function Ei(n,t){var e=n.N;if(e)return ki(e,t);var r=n.site;return r.y===t?r.x:1/0}function Ni(n){this.site=n,this.edges=[]}function Li(n,t){return t.angle-n.angle}function Ci(){Hi(this),this.x=this.y=this.arc=this.site=this.cy=null}function Ai(n){var t=n.P,e=n.N;if(t&&e){var r=t.site,i=n.site,o=e.site;if(r!==o){var u=i.x,a=i.y,l=r.x-u,c=r.y-a,s=o.x-u,f=2*(l*(v=o.y-a)-c*s);if(!(f>=-Ln)){var h=l*l+c*c,p=s*s+v*v,d=(v*h-c*p)/f,g=(l*p-s*h)/f,v=g+a,m=xi.pop()||new Ci;m.arc=n,m.site=i,m.x=d+u,m.y=v+Math.sqrt(d*d+g*g),m.cy=v,n.circle=m;for(var y=null,x=mi._;x;)if(m.y<x.y||m.y===x.y&&m.x<=x.x){if(!x.L){y=x.P;break}x=x.L}else{if(!x.R){y=x;break}x=x.R}mi.insert(y,m),y||(vi=m)}}}}function Ti(n){var t=n.circle;t&&(t.P||(vi=t.N),mi.remove(t),xi.push(t),Hi(t),n.circle=null)}function Ri(n,t){var e=n.b;if(e)return!0;var r,i,o=n.a,u=t[0][0],a=t[1][0],l=t[0][1],c=t[1][1],s=n.l,f=n.r,h=s.x,p=s.y,d=f.x,g=f.y,v=(h+d)/2,m=(p+g)/2;if(g===p){if(v<u||v>=a)return;if(h>d){if(o){if(o.y>=c)return}else o={x:v,y:l};e={x:v,y:c}}else{if(o){if(o.y<l)return}else o={x:v,y:c};e={x:v,y:l}}}else if(i=m-(r=(h-d)/(g-p))*v,r<-1||r>1)if(h>d){if(o){if(o.y>=c)return}else o={x:(l-i)/r,y:l};e={x:(c-i)/r,y:c}}else{if(o){if(o.y<l)return}else o={x:(c-i)/r,y:c};e={x:(l-i)/r,y:l}}else if(p<g){if(o){if(o.x>=a)return}else o={x:u,y:r*u+i};e={x:a,y:r*a+i}}else{if(o){if(o.x<u)return}else o={x:a,y:r*a+i};e={x:u,y:r*u+i}}return n.a=o,n.b=e,!0}function zi(n,t){this.l=n,this.r=t,this.a=this.b=null}function Di(n,t,e,r){var i=new zi(n,t);return pi.push(i),e&&Oi(i,n,t,e),r&&Oi(i,t,n,r),di[n.i].edges.push(new Pi(i,n,t)),di[t.i].edges.push(new Pi(i,t,n)),i}function qi(n,t,e){var r=new zi(n,null);return r.a=t,r.b=e,pi.push(r),r}function Oi(n,t,e,r){n.a||n.b?n.l===e?n.b=r:n.a=r:(n.a=r,n.l=t,n.r=e)}function Pi(n,t,e){var r=n.a,i=n.b;this.edge=n,this.site=t,this.angle=e?Math.atan2(e.y-t.y,e.x-t.x):n.l===t?Math.atan2(i.x-r.x,r.y-i.y):Math.atan2(r.x-i.x,i.y-r.y)}function Ui(){this._=null}function Hi(n){n.U=n.C=n.L=n.R=n.P=n.N=null}function ji(n,t){var e=t,r=t.R,i=e.U;i?i.L===e?i.L=r:i.R=r:n._=r,r.U=i,e.U=r,e.R=r.L,e.R&&(e.R.U=e),r.L=e}function Ii(n,t){var e=t,r=t.L,i=e.U;i?i.L===e?i.L=r:i.R=r:n._=r,r.U=i,e.U=r,e.L=r.R,e.L&&(e.L.U=e),r.R=e}function $i(n){for(;n.L;)n=n.L;return n}function Fi(n,t){var e,r,i,o=n.sort(Bi).pop();for(pi=[],di=new Array(n.length),gi=new Ui,mi=new Ui;;)if(i=vi,o&&(!i||o.y<i.y||o.y===i.y&&o.x<i.x))o.x===e&&o.y===r||(di[o.i]=new Ni(o),Si(o),e=o.x,r=o.y),o=n.pop();else{if(!i)break;_i(i.arc)}t&&(function(n){for(var t,e=pi,r=ur(n[0][0],n[0][1],n[1][0],n[1][1]),i=e.length;i--;)(!Ri(t=e[i],n)||!r(t)||w(t.a.x-t.b.x)<Nn&&w(t.a.y-t.b.y)<Nn)&&(t.a=t.b=null,e.splice(i,1))}(t),function(n){for(var t,e,r,i,o,u,a,l,c,s,f=n[0][0],h=n[1][0],p=n[0][1],d=n[1][1],g=di,v=g.length;v--;)if((o=g[v])&&o.prepare())for(l=(a=o.edges).length,u=0;u<l;)r=(s=a[u].end()).x,i=s.y,t=(c=a[++u%l].start()).x,e=c.y,(w(r-t)>Nn||w(i-e)>Nn)&&(a.splice(u,0,new Pi(qi(o.site,s,w(r-f)<Nn&&d-i>Nn?{x:f,y:w(t-f)<Nn?e:d}:w(i-d)<Nn&&h-r>Nn?{x:w(e-d)<Nn?t:h,y:d}:w(r-h)<Nn&&i-p>Nn?{x:h,y:w(t-h)<Nn?e:p}:w(i-p)<Nn&&r-f>Nn?{x:w(e-p)<Nn?t:f,y:p}:null),o.site,null)),++l)}(t));var u={cells:di,edges:pi};return gi=mi=pi=di=null,u}function Bi(n,t){return t.y-n.y||t.x-n.x}Ni.prototype.prepare=function(){for(var n,t=this.edges,e=t.length;e--;)(n=t[e].edge).b&&n.a||t.splice(e,1);return t.sort(Li),t.length},Pi.prototype={start:function(){return this.edge.l===this.site?this.edge.a:this.edge.b},end:function(){return this.edge.l===this.site?this.edge.b:this.edge.a}},Ui.prototype={insert:function(n,t){var e,r,i;if(n){if(t.P=n,t.N=n.N,n.N&&(n.N.P=t),n.N=t,n.R){for(n=n.R;n.L;)n=n.L;n.L=t}else n.R=t;e=n}else this._?(n=$i(this._),t.P=null,t.N=n,n.P=n.L=t,e=n):(t.P=t.N=null,this._=t,e=null);for(t.L=t.R=null,t.U=e,t.C=!0,n=t;e&&e.C;)e===(r=e.U).L?(i=r.R)&&i.C?(e.C=i.C=!1,r.C=!0,n=r):(n===e.R&&(ji(this,e),e=(n=e).U),e.C=!1,r.C=!0,Ii(this,r)):(i=r.L)&&i.C?(e.C=i.C=!1,r.C=!0,n=r):(n===e.L&&(Ii(this,e),e=(n=e).U),e.C=!1,r.C=!0,ji(this,r)),e=n.U;this._.C=!1},remove:function(n){n.N&&(n.N.P=n.P),n.P&&(n.P.N=n.N),n.N=n.P=null;var t,e,r,i=n.U,o=n.L,u=n.R;if(e=o?u?$i(u):o:u,i?i.L===n?i.L=e:i.R=e:this._=e,o&&u?(r=e.C,e.C=n.C,e.L=o,o.U=e,e!==u?(i=e.U,e.U=n.U,n=e.R,i.L=n,e.R=u,u.U=e):(e.U=i,i=e,n=e.R)):(r=n.C,n=e),n&&(n.U=i),!r)if(n&&n.C)n.C=!1;else{do{if(n===this._)break;if(n===i.L){if((t=i.R).C&&(t.C=!1,i.C=!0,ji(this,i),t=i.R),t.L&&t.L.C||t.R&&t.R.C){t.R&&t.R.C||(t.L.C=!1,t.C=!0,Ii(this,t),t=i.R),t.C=i.C,i.C=t.R.C=!1,ji(this,i),n=this._;break}}else if((t=i.L).C&&(t.C=!1,i.C=!0,Ii(this,i),t=i.L),t.L&&t.L.C||t.R&&t.R.C){t.L&&t.L.C||(t.R.C=!1,t.C=!0,ji(this,t),t=i.L),t.C=i.C,i.C=t.L.C=!1,Ii(this,i),n=this._;break}t.C=!0,n=i,i=i.U}while(!n.C);n&&(n.C=!1)}}},o.geom.voronoi=function(n){var t=oi,e=ui,r=t,i=e,o=Gi;if(n)return u(n);function u(n){var t=new Array(n.length),e=o[0][0],r=o[0][1],i=o[1][0],u=o[1][1];return Fi(a(n),o).cells.forEach(function(o,a){var l=o.edges,c=o.site;(t[a]=l.length?l.map(function(n){var t=n.start();return[t.x,t.y]}):c.x>=e&&c.x<=i&&c.y>=r&&c.y<=u?[[e,u],[i,u],[i,r],[e,r]]:[]).point=n[a]}),t}function a(n){return n.map(function(n,t){return{x:Math.round(r(n,t)/Nn)*Nn,y:Math.round(i(n,t)/Nn)*Nn,i:t}})}return u.links=function(n){return Fi(a(n)).edges.filter(function(n){return n.l&&n.r}).map(function(t){return{source:n[t.l.i],target:n[t.r.i]}})},u.triangles=function(n){var t=[];return Fi(a(n)).cells.forEach(function(e,r){for(var i,o=e.site,u=e.edges.sort(Li),a=-1,l=u.length,c=u[l-1].edge,s=c.l===o?c.r:c.l;++a<l;)i=s,s=(c=u[a].edge).l===o?c.r:c.l,r<i.i&&r<s.i&&Yi(o,i,s)<0&&t.push([n[r],n[i.i],n[s.i]])}),t},u.x=function(n){return arguments.length?(r=xt(t=n),u):t},u.y=function(n){return arguments.length?(i=xt(e=n),u):e},u.clipExtent=function(n){return arguments.length?(o=null==n?Gi:n,u):o===Gi?null:o},u.size=function(n){return arguments.length?u.clipExtent(n&&[[0,0],n]):o===Gi?null:o&&o[1]},u};var Gi=[[-1e6,-1e6],[1e6,1e6]];function Yi(n,t,e){return(n.x-e.x)*(t.y-n.y)-(n.x-t.x)*(e.y-n.y)}function Zi(n){return n.x}function Vi(n){return n.y}function Xi(n,t,e,r,i,o){if(!n(t,e,r,i,o)){var u=.5*(e+i),a=.5*(r+o),l=t.nodes;l[0]&&Xi(n,l[0],e,r,u,a),l[1]&&Xi(n,l[1],u,r,i,a),l[2]&&Xi(n,l[2],e,a,u,o),l[3]&&Xi(n,l[3],u,a,i,o)}}function Wi(n,t){n=o.rgb(n),t=o.rgb(t);var e=n.r,r=n.g,i=n.b,u=t.r-e,a=t.g-r,l=t.b-i;return function(n){return"#"+ht(Math.round(e+u*n))+ht(Math.round(r+a*n))+ht(Math.round(i+l*n))}}function Ji(n,t){var e,r={},i={};for(e in n)e in t?r[e]=eo(n[e],t[e]):i[e]=n[e];for(e in t)e in n||(i[e]=t[e]);return function(n){for(e in r)i[e]=r[e](n);return i}}function Ki(n,t){return n=+n,t=+t,function(e){return n*(1-e)+t*e}}function Qi(n,t){var e,r,i,o=no.lastIndex=to.lastIndex=0,u=-1,a=[],l=[];for(n+="",t+="";(e=no.exec(n))&&(r=to.exec(t));)(i=r.index)>o&&(i=t.slice(o,i),a[u]?a[u]+=i:a[++u]=i),(e=e[0])===(r=r[0])?a[u]?a[u]+=r:a[++u]=r:(a[++u]=null,l.push({i:u,x:Ki(e,r)})),o=to.lastIndex;return o<t.length&&(i=t.slice(o),a[u]?a[u]+=i:a[++u]=i),a.length<2?l[0]?(t=l[0].x,function(n){return t(n)+""}):function(){return t}:(t=l.length,function(n){for(var e,r=0;r<t;++r)a[(e=l[r]).i]=e.x(n);return a.join("")})}o.geom.delaunay=function(n){return o.geom.voronoi().triangles(n)},o.geom.quadtree=function(n,t,e,r,i){var o,u=oi,a=ui;if(o=arguments.length)return u=Zi,a=Vi,3===o&&(i=e,r=t,e=t=0),l(n);function l(n){var l,c,s,f,h,p,d,g,v,m=xt(u),y=xt(a);if(null!=t)p=t,d=e,g=r,v=i;else if(g=v=-(p=d=1/0),c=[],s=[],h=n.length,o)for(f=0;f<h;++f)(l=n[f]).x<p&&(p=l.x),l.y<d&&(d=l.y),l.x>g&&(g=l.x),l.y>v&&(v=l.y),c.push(l.x),s.push(l.y);else for(f=0;f<h;++f){var x=+m(l=n[f],f),M=+y(l,f);x<p&&(p=x),M<d&&(d=M),x>g&&(g=x),M>v&&(v=M),c.push(x),s.push(M)}var b=g-p,_=v-d;function S(n,t,e,r,i,o,u,a){if(!isNaN(e)&&!isNaN(r))if(n.leaf){var l=n.x,c=n.y;if(null!=l)if(w(l-e)+w(c-r)<.01)k(n,t,e,r,i,o,u,a);else{var s=n.point;n.x=n.y=n.point=null,k(n,s,l,c,i,o,u,a),k(n,t,e,r,i,o,u,a)}else n.x=e,n.y=r,n.point=t}else k(n,t,e,r,i,o,u,a)}function k(n,t,e,r,i,o,u,a){var l=.5*(i+u),c=.5*(o+a),s=e>=l,f=r>=c,h=f<<1|s;n.leaf=!1,s?i=l:u=l,f?o=c:a=c,S(n=n.nodes[h]||(n.nodes[h]={leaf:!0,nodes:[],point:null,x:null,y:null}),t,e,r,i,o,u,a)}b>_?v=d+b:g=p+_;var E={leaf:!0,nodes:[],point:null,x:null,y:null,add:function(n){S(E,n,+m(n,++f),+y(n,f),p,d,g,v)}};if(E.visit=function(n){Xi(n,E,p,d,g,v)},E.find=function(n){return function(n,t,e,r,i,o,u){var a,l=1/0;return function n(c,s,f,h,p){if(!(s>o||f>u||h<r||p<i)){if(d=c.point){var d,g=t-c.x,v=e-c.y,m=g*g+v*v;if(m<l){var y=Math.sqrt(l=m);r=t-y,i=e-y,o=t+y,u=e+y,a=d}}for(var x=c.nodes,M=.5*(s+h),b=.5*(f+p),w=(e>=b)<<1|t>=M,_=w+4;w<_;++w)if(c=x[3&w])switch(3&w){case 0:n(c,s,f,M,b);break;case 1:n(c,M,f,h,b);break;case 2:n(c,s,b,M,p);break;case 3:n(c,M,b,h,p)}}}(n,r,i,o,u),a}(E,n[0],n[1],p,d,g,v)},f=-1,null==t){for(;++f<h;)S(E,n[f],c[f],s[f],p,d,g,v);--f}else n.forEach(E.add);return c=s=n=l=null,E}return l.x=function(n){return arguments.length?(u=n,l):u},l.y=function(n){return arguments.length?(a=n,l):a},l.extent=function(n){return arguments.length?(null==n?t=e=r=i=null:(t=+n[0][0],e=+n[0][1],r=+n[1][0],i=+n[1][1]),l):null==t?null:[[t,e],[r,i]]},l.size=function(n){return arguments.length?(null==n?t=e=r=i=null:(t=e=0,r=+n[0],i=+n[1]),l):null==t?null:[r-t,i-e]},l},o.interpolateRgb=Wi,o.interpolateObject=Ji,o.interpolateNumber=Ki,o.interpolateString=Qi;var no=/[-+]?(?:\d+\.?\d*|\.?\d+)(?:[eE][-+]?\d+)?/g,to=new RegExp(no.source,"g");function eo(n,t){for(var e,r=o.interpolators.length;--r>=0&&!(e=o.interpolators[r](n,t)););return e}function ro(n,t){var e,r=[],i=[],o=n.length,u=t.length,a=Math.min(n.length,t.length);for(e=0;e<a;++e)r.push(eo(n[e],t[e]));for(;e<o;++e)i[e]=n[e];for(;e<u;++e)i[e]=t[e];return function(n){for(e=0;e<a;++e)i[e]=r[e](n);return i}}o.interpolate=eo,o.interpolators=[function(n,t){var e=typeof t;return("string"===e?yt.has(t.toLowerCase())||/^(#|rgb\(|hsl\()/i.test(t)?Wi:Qi:t instanceof Gn?Wi:Array.isArray(t)?ro:"object"===e&&isNaN(t)?Ji:Ki)(n,t)}],o.interpolateArray=ro;var io=function(){return z},oo=o.map({linear:io,poly:function(n){return function(t){return Math.pow(t,n)}},quad:function(){return co},cubic:function(){return so},sin:function(){return ho},exp:function(){return po},circle:function(){return go},elastic:function(n,t){var e;return arguments.length<2&&(t=.45),arguments.length?e=t/An*Math.asin(1/n):(n=1,e=t/4),function(r){return 1+n*Math.pow(2,-10*r)*Math.sin((r-e)*An/t)}},back:function(n){return n||(n=1.70158),function(t){return t*t*((n+1)*t-n)}},bounce:function(){return vo}}),uo=o.map({in:z,out:ao,"in-out":lo,"out-in":function(n){return lo(ao(n))}});function ao(n){return function(t){return 1-n(1-t)}}function lo(n){return function(t){return.5*(t<.5?n(2*t):2-n(2-2*t))}}function co(n){return n*n}function so(n){return n*n*n}function fo(n){if(n<=0)return 0;if(n>=1)return 1;var t=n*n,e=t*n;return 4*(n<.5?e:3*(n-t)+e-.75)}function ho(n){return 1-Math.cos(n*Rn)}function po(n){return Math.pow(2,10*(n-1))}function go(n){return 1-Math.sqrt(1-n*n)}function vo(n){return n<1/2.75?7.5625*n*n:n<2/2.75?7.5625*(n-=1.5/2.75)*n+.75:n<2.5/2.75?7.5625*(n-=2.25/2.75)*n+.9375:7.5625*(n-=2.625/2.75)*n+.984375}function mo(n,t){return t-=n,function(e){return Math.round(n+t*e)}}function yo(n){var t,e,r,i=[n.a,n.b],o=[n.c,n.d],u=Mo(i),a=xo(i,o),l=Mo(((t=o)[0]+=(r=-a)*(e=i)[0],t[1]+=r*e[1],t))||0;i[0]*o[1]<o[0]*i[1]&&(i[0]*=-1,i[1]*=-1,u*=-1,a*=-1),this.rotate=(u?Math.atan2(i[1],i[0]):Math.atan2(-o[0],o[1]))*Dn,this.translate=[n.e,n.f],this.scale=[u,l],this.skew=l?Math.atan2(a,l)*Dn:0}function xo(n,t){return n[0]*t[0]+n[1]*t[1]}function Mo(n){var t=Math.sqrt(xo(n,n));return t&&(n[0]/=t,n[1]/=t),t}o.ease=function(n){var t,e=n.indexOf("-"),r=e>=0?n.slice(0,e):n,i=e>=0?n.slice(e+1):"in";return r=oo.get(r)||io,i=uo.get(i)||z,t=i(r.apply(null,u.call(arguments,1))),function(n){return n<=0?0:n>=1?1:t(n)}},o.interpolateHcl=function(n,t){n=o.hcl(n),t=o.hcl(t);var e=n.h,r=n.c,i=n.l,u=t.h-e,a=t.c-r,l=t.l-i;return isNaN(a)&&(a=0,r=isNaN(r)?t.c:r),isNaN(u)?(u=0,e=isNaN(e)?t.h:e):u>180?u-=360:u<-180&&(u+=360),function(n){return Jn(e+u*n,r+a*n,i+l*n)+""}},o.interpolateHsl=function(n,t){n=o.hsl(n),t=o.hsl(t);var e=n.h,r=n.s,i=n.l,u=t.h-e,a=t.s-r,l=t.l-i;return isNaN(a)&&(a=0,r=isNaN(r)?t.s:r),isNaN(u)?(u=0,e=isNaN(e)?t.h:e):u>180?u-=360:u<-180&&(u+=360),function(n){return Vn(e+u*n,r+a*n,i+l*n)+""}},o.interpolateLab=function(n,t){n=o.lab(n),t=o.lab(t);var e=n.l,r=n.a,i=n.b,u=t.l-e,a=t.a-r,l=t.b-i;return function(n){return it(e+u*n,r+a*n,i+l*n)+""}},o.interpolateRound=mo,o.transform=function(n){var t=l.createElementNS(o.ns.prefix.svg,"g");return(o.transform=function(n){if(null!=n){t.setAttribute("transform",n);var e=t.transform.baseVal.consolidate()}return new yo(e?e.matrix:bo)})(n)},yo.prototype.toString=function(){return"translate("+this.translate+")rotate("+this.rotate+")skewX("+this.skew+")scale("+this.scale+")"};var bo={a:1,b:0,c:0,d:1,e:0,f:0};function wo(n){return n.length?n.pop()+",":""}function _o(n,t){var e=[],r=[];return n=o.transform(n),t=o.transform(t),function(n,t,e,r){if(n[0]!==t[0]||n[1]!==t[1]){var i=e.push("translate(",null,",",null,")");r.push({i:i-4,x:Ki(n[0],t[0])},{i:i-2,x:Ki(n[1],t[1])})}else(t[0]||t[1])&&e.push("translate("+t+")")}(n.translate,t.translate,e,r),function(n,t,e,r){n!==t?(n-t>180?t+=360:t-n>180&&(n+=360),r.push({i:e.push(wo(e)+"rotate(",null,")")-2,x:Ki(n,t)})):t&&e.push(wo(e)+"rotate("+t+")")}(n.rotate,t.rotate,e,r),function(n,t,e,r){n!==t?r.push({i:e.push(wo(e)+"skewX(",null,")")-2,x:Ki(n,t)}):t&&e.push(wo(e)+"skewX("+t+")")}(n.skew,t.skew,e,r),function(n,t,e,r){if(n[0]!==t[0]||n[1]!==t[1]){var i=e.push(wo(e)+"scale(",null,",",null,")");r.push({i:i-4,x:Ki(n[0],t[0])},{i:i-2,x:Ki(n[1],t[1])})}else 1===t[0]&&1===t[1]||e.push(wo(e)+"scale("+t+")")}(n.scale,t.scale,e,r),n=t=null,function(n){for(var t,i=-1,o=r.length;++i<o;)e[(t=r[i]).i]=t.x(n);return e.join("")}}function So(n,t){return t=(t-=n=+n)||1/t,function(e){return(e-n)/t}}function ko(n,t){return t=(t-=n=+n)||1/t,function(e){return Math.max(0,Math.min(1,(e-n)/t))}}function Eo(n){for(var t=n.source,e=n.target,r=function(n,t){if(n===t)return n;for(var e=No(n),r=No(t),i=e.pop(),o=r.pop(),u=null;i===o;)u=i,i=e.pop(),o=r.pop();return u}(t,e),i=[t];t!==r;)t=t.parent,i.push(t);for(var o=i.length;e!==r;)i.splice(o,0,e),e=e.parent;return i}function No(n){for(var t=[],e=n.parent;null!=e;)t.push(n),n=e,e=e.parent;return t.push(n),t}function Lo(n){n.fixed|=2}function Co(n){n.fixed&=-7}function Ao(n){n.fixed|=4,n.px=n.x,n.py=n.y}function To(n){n.fixed&=-5}function Ro(n,t,e){var r=0,i=0;if(n.charge=0,!n.leaf)for(var o,u=n.nodes,a=u.length,l=-1;++l<a;)null!=(o=u[l])&&(Ro(o,t,e),n.charge+=o.charge,r+=o.charge*o.cx,i+=o.charge*o.cy);if(n.point){n.leaf||(n.point.x+=Math.random()-.5,n.point.y+=Math.random()-.5);var c=t*e[n.point.index];n.charge+=n.pointCharge=c,r+=c*n.point.x,i+=c*n.point.y}n.cx=r/n.charge,n.cy=i/n.charge}o.interpolateTransform=_o,o.layout={},o.layout.bundle=function(){return function(n){for(var t=[],e=-1,r=n.length;++e<r;)t.push(Eo(n[e]));return t}},o.layout.chord=function(){var n,t,e,r,i,u,a,l={},c=0;function s(){var l,s,h,p,d,g={},v=[],m=o.range(r),y=[];for(n=[],t=[],l=0,p=-1;++p<r;){for(s=0,d=-1;++d<r;)s+=e[p][d];v.push(s),y.push(o.range(r)),l+=s}for(i&&m.sort(function(n,t){return i(v[n],v[t])}),u&&y.forEach(function(n,t){n.sort(function(n,r){return u(e[t][n],e[t][r])})}),l=(An-c*r)/l,s=0,p=-1;++p<r;){for(h=s,d=-1;++d<r;){var x=m[p],M=y[x][d],b=e[x][M],w=s,_=s+=b*l;g[x+"-"+M]={index:x,subindex:M,startAngle:w,endAngle:_,value:b}}t[x]={index:x,startAngle:h,endAngle:s,value:v[x]},s+=c}for(p=-1;++p<r;)for(d=p-1;++d<r;){var S=g[p+"-"+d],k=g[d+"-"+p];(S.value||k.value)&&n.push(S.value<k.value?{source:k,target:S}:{source:S,target:k})}a&&f()}function f(){n.sort(function(n,t){return a((n.source.value+n.target.value)/2,(t.source.value+t.target.value)/2)})}return l.matrix=function(i){return arguments.length?(r=(e=i)&&e.length,n=t=null,l):e},l.padding=function(e){return arguments.length?(c=e,n=t=null,l):c},l.sortGroups=function(e){return arguments.length?(i=e,n=t=null,l):i},l.sortSubgroups=function(t){return arguments.length?(u=t,n=null,l):u},l.sortChords=function(t){return arguments.length?(a=t,n&&f(),l):a},l.chords=function(){return n||s(),n},l.groups=function(){return t||s(),t},l},o.layout.force=function(){var n,t,e,r,i,u,a={},l=o.dispatch("start","tick","end"),c=[1,1],s=.9,f=zo,h=Do,p=-30,d=qo,g=.1,v=.64,m=[],y=[];function x(n){return function(t,e,r,i){if(t.point!==n){var o=t.cx-n.x,u=t.cy-n.y,a=i-e,l=o*o+u*u;if(a*a/v<l){if(l<d){var c=t.charge/l;n.px-=o*c,n.py-=u*c}return!0}t.point&&l&&l<d&&(c=t.pointCharge/l,n.px-=o*c,n.py-=u*c)}return!t.charge}}function M(n){n.px=o.event.x,n.py=o.event.y,a.resume()}return a.tick=function(){if((e*=.99)<.005)return n=null,l.end({type:"end",alpha:e=0}),!0;var t,a,f,h,d,v,M,b,w,_=m.length,S=y.length;for(a=0;a<S;++a)h=(f=y[a]).source,(v=(b=(d=f.target).x-h.x)*b+(w=d.y-h.y)*w)&&(b*=v=e*i[a]*((v=Math.sqrt(v))-r[a])/v,w*=v,d.x-=b*(M=h.weight+d.weight?h.weight/(h.weight+d.weight):.5),d.y-=w*M,h.x+=b*(M=1-M),h.y+=w*M);if((M=e*g)&&(b=c[0]/2,w=c[1]/2,a=-1,M))for(;++a<_;)(f=m[a]).x+=(b-f.x)*M,f.y+=(w-f.y)*M;if(p)for(Ro(t=o.geom.quadtree(m),e,u),a=-1;++a<_;)(f=m[a]).fixed||t.visit(x(f));for(a=-1;++a<_;)(f=m[a]).fixed?(f.x=f.px,f.y=f.py):(f.x-=(f.px-(f.px=f.x))*s,f.y-=(f.py-(f.py=f.y))*s);l.tick({type:"tick",alpha:e})},a.nodes=function(n){return arguments.length?(m=n,a):m},a.links=function(n){return arguments.length?(y=n,a):y},a.size=function(n){return arguments.length?(c=n,a):c},a.linkDistance=function(n){return arguments.length?(f="function"==typeof n?n:+n,a):f},a.distance=a.linkDistance,a.linkStrength=function(n){return arguments.length?(h="function"==typeof n?n:+n,a):h},a.friction=function(n){return arguments.length?(s=+n,a):s},a.charge=function(n){return arguments.length?(p="function"==typeof n?n:+n,a):p},a.chargeDistance=function(n){return arguments.length?(d=n*n,a):Math.sqrt(d)},a.gravity=function(n){return arguments.length?(g=+n,a):g},a.theta=function(n){return arguments.length?(v=n*n,a):Math.sqrt(v)},a.alpha=function(t){return arguments.length?(t=+t,e?t>0?e=t:(n.c=null,n.t=NaN,n=null,l.end({type:"end",alpha:e=0})):t>0&&(l.start({type:"start",alpha:e=t}),n=Nt(a.tick)),a):e},a.start=function(){var n,t,e,o=m.length,l=y.length,s=c[0],d=c[1];for(n=0;n<o;++n)(e=m[n]).index=n,e.weight=0;for(n=0;n<l;++n)"number"==typeof(e=y[n]).source&&(e.source=m[e.source]),"number"==typeof e.target&&(e.target=m[e.target]),++e.source.weight,++e.target.weight;for(n=0;n<o;++n)e=m[n],isNaN(e.x)&&(e.x=g("x",s)),isNaN(e.y)&&(e.y=g("y",d)),isNaN(e.px)&&(e.px=e.x),isNaN(e.py)&&(e.py=e.y);if(r=[],"function"==typeof f)for(n=0;n<l;++n)r[n]=+f.call(this,y[n],n);else for(n=0;n<l;++n)r[n]=f;if(i=[],"function"==typeof h)for(n=0;n<l;++n)i[n]=+h.call(this,y[n],n);else for(n=0;n<l;++n)i[n]=h;if(u=[],"function"==typeof p)for(n=0;n<o;++n)u[n]=+p.call(this,m[n],n);else for(n=0;n<o;++n)u[n]=p;function g(e,r){if(!t){for(t=new Array(o),c=0;c<o;++c)t[c]=[];for(c=0;c<l;++c){var i=y[c];t[i.source.index].push(i.target),t[i.target.index].push(i.source)}}for(var u,a=t[n],c=-1,s=a.length;++c<s;)if(!isNaN(u=a[c][e]))return u;return Math.random()*r}return a.resume()},a.resume=function(){return a.alpha(.1)},a.stop=function(){return a.alpha(0)},a.drag=function(){if(t||(t=o.behavior.drag().origin(z).on("dragstart.force",Lo).on("drag.force",M).on("dragend.force",Co)),!arguments.length)return t;this.on("mouseover.force",Ao).on("mouseout.force",To).call(t)},o.rebind(a,l,"on")};var zo=20,Do=1,qo=1/0;function Oo(n,t){return o.rebind(n,t,"sort","children","value"),n.nodes=n,n.links=$o,n}function Po(n,t){for(var e=[n];null!=(n=e.pop());)if(t(n),(i=n.children)&&(r=i.length))for(var r,i;--r>=0;)e.push(i[r])}function Uo(n,t){for(var e=[n],r=[];null!=(n=e.pop());)if(r.push(n),(o=n.children)&&(i=o.length))for(var i,o,u=-1;++u<i;)e.push(o[u]);for(;null!=(n=r.pop());)t(n)}function Ho(n){return n.children}function jo(n){return n.value}function Io(n,t){return t.value-n.value}function $o(n){return o.merge(n.map(function(n){return(n.children||[]).map(function(t){return{source:n,target:t}})}))}o.layout.hierarchy=function(){var n=Io,t=Ho,e=jo;function r(i){var o,u=[i],a=[];for(i.depth=0;null!=(o=u.pop());)if(a.push(o),(c=t.call(r,o,o.depth))&&(l=c.length)){for(var l,c,s;--l>=0;)u.push(s=c[l]),s.parent=o,s.depth=o.depth+1;e&&(o.value=0),o.children=c}else e&&(o.value=+e.call(r,o,o.depth)||0),delete o.children;return Uo(i,function(t){var r,i;n&&(r=t.children)&&r.sort(n),e&&(i=t.parent)&&(i.value+=t.value)}),a}return r.sort=function(t){return arguments.length?(n=t,r):n},r.children=function(n){return arguments.length?(t=n,r):t},r.value=function(n){return arguments.length?(e=n,r):e},r.revalue=function(n){return e&&(Po(n,function(n){n.children&&(n.value=0)}),Uo(n,function(n){var t;n.children||(n.value=+e.call(r,n,n.depth)||0),(t=n.parent)&&(t.value+=n.value)})),n},r},o.layout.partition=function(){var n=o.layout.hierarchy(),t=[1,1];function e(n,t,r,i){var o=n.children;if(n.x=t,n.y=n.depth*i,n.dx=r,n.dy=i,o&&(u=o.length)){var u,a,l,c=-1;for(r=n.value?r/n.value:0;++c<u;)e(a=o[c],t,l=a.value*r,i),t+=l}}function r(n){var t=n.children,e=0;if(t&&(i=t.length))for(var i,o=-1;++o<i;)e=Math.max(e,r(t[o]));return 1+e}function i(i,o){var u=n.call(this,i,o);return e(u[0],0,t[0],t[1]/r(u[0])),u}return i.size=function(n){return arguments.length?(t=n,i):t},Oo(i,n)},o.layout.pie=function(){var n=Number,t=Fo,e=0,r=An,i=0;function u(a){var l,c=a.length,s=a.map(function(t,e){return+n.call(u,t,e)}),f=+("function"==typeof e?e.apply(this,arguments):e),h=("function"==typeof r?r.apply(this,arguments):r)-f,p=Math.min(Math.abs(h)/c,+("function"==typeof i?i.apply(this,arguments):i)),d=p*(h<0?-1:1),g=o.sum(s),v=g?(h-c*d)/g:0,m=o.range(c),y=[];return null!=t&&m.sort(t===Fo?function(n,t){return s[t]-s[n]}:function(n,e){return t(a[n],a[e])}),m.forEach(function(n){y[n]={data:a[n],value:l=s[n],startAngle:f,endAngle:f+=l*v+d,padAngle:p}}),y}return u.value=function(t){return arguments.length?(n=t,u):n},u.sort=function(n){return arguments.length?(t=n,u):t},u.startAngle=function(n){return arguments.length?(e=n,u):e},u.endAngle=function(n){return arguments.length?(r=n,u):r},u.padAngle=function(n){return arguments.length?(i=n,u):i},u};var Fo={};function Bo(n){return n.x}function Go(n){return n.y}function Yo(n,t,e){n.y0=t,n.y=e}o.layout.stack=function(){var n=z,t=Xo,e=Wo,r=Yo,i=Bo,u=Go;function a(l,c){if(!(p=l.length))return l;var s=l.map(function(t,e){return n.call(a,t,e)}),f=s.map(function(n){return n.map(function(n,t){return[i.call(a,n,t),u.call(a,n,t)]})}),h=t.call(a,f,c);s=o.permute(s,h),f=o.permute(f,h);var p,d,g,v,m=e.call(a,f,c),y=s[0].length;for(g=0;g<y;++g)for(r.call(a,s[0][g],v=m[g],f[0][g][1]),d=1;d<p;++d)r.call(a,s[d][g],v+=f[d-1][g][1],f[d][g][1]);return l}return a.values=function(t){return arguments.length?(n=t,a):n},a.order=function(n){return arguments.length?(t="function"==typeof n?n:Zo.get(n)||Xo,a):t},a.offset=function(n){return arguments.length?(e="function"==typeof n?n:Vo.get(n)||Wo,a):e},a.x=function(n){return arguments.length?(i=n,a):i},a.y=function(n){return arguments.length?(u=n,a):u},a.out=function(n){return arguments.length?(r=n,a):r},a};var Zo=o.map({"inside-out":function(n){var t,e,r=n.length,i=n.map(Jo),u=n.map(Ko),a=o.range(r).sort(function(n,t){return i[n]-i[t]}),l=0,c=0,s=[],f=[];for(t=0;t<r;++t)e=a[t],l<c?(l+=u[e],s.push(e)):(c+=u[e],f.push(e));return f.reverse().concat(s)},reverse:function(n){return o.range(n.length).reverse()},default:Xo}),Vo=o.map({silhouette:function(n){var t,e,r,i=n.length,o=n[0].length,u=[],a=0,l=[];for(e=0;e<o;++e){for(t=0,r=0;t<i;t++)r+=n[t][e][1];r>a&&(a=r),u.push(r)}for(e=0;e<o;++e)l[e]=(a-u[e])/2;return l},wiggle:function(n){var t,e,r,i,o,u,a,l,c,s=n.length,f=n[0],h=f.length,p=[];for(p[0]=l=c=0,e=1;e<h;++e){for(t=0,i=0;t<s;++t)i+=n[t][e][1];for(t=0,o=0,a=f[e][0]-f[e-1][0];t<s;++t){for(r=0,u=(n[t][e][1]-n[t][e-1][1])/(2*a);r<t;++r)u+=(n[r][e][1]-n[r][e-1][1])/a;o+=u*n[t][e][1]}p[e]=l-=i?o/i*a:0,l<c&&(c=l)}for(e=0;e<h;++e)p[e]-=c;return p},expand:function(n){var t,e,r,i=n.length,o=n[0].length,u=1/i,a=[];for(e=0;e<o;++e){for(t=0,r=0;t<i;t++)r+=n[t][e][1];if(r)for(t=0;t<i;t++)n[t][e][1]/=r;else for(t=0;t<i;t++)n[t][e][1]=u}for(e=0;e<o;++e)a[e]=0;return a},zero:Wo});function Xo(n){return o.range(n.length)}function Wo(n){for(var t=-1,e=n[0].length,r=[];++t<e;)r[t]=0;return r}function Jo(n){for(var t,e=1,r=0,i=n[0][1],o=n.length;e<o;++e)(t=n[e][1])>i&&(r=e,i=t);return r}function Ko(n){return n.reduce(Qo,0)}function Qo(n,t){return n+t[1]}function nu(n,t){return tu(n,Math.ceil(Math.log(t.length)/Math.LN2+1))}function tu(n,t){for(var e=-1,r=+n[0],i=(n[1]-r)/t,o=[];++e<=t;)o[e]=i*e+r;return o}function eu(n){return[o.min(n),o.max(n)]}function ru(n,t){return n.value-t.value}function iu(n,t){var e=n._pack_next;n._pack_next=t,t._pack_prev=n,t._pack_next=e,e._pack_prev=t}function ou(n,t){n._pack_next=t,t._pack_prev=n}function uu(n,t){var e=t.x-n.x,r=t.y-n.y,i=n.r+t.r;return.999*i*i>e*e+r*r}function au(n){if((t=n.children)&&(l=t.length)){var t,e,r,i,o,u,a,l,c=1/0,s=-1/0,f=1/0,h=-1/0;if(t.forEach(lu),(e=t[0]).x=-e.r,e.y=0,x(e),l>1&&((r=t[1]).x=r.r,r.y=0,x(r),l>2))for(fu(e,r,i=t[2]),x(i),iu(e,i),e._pack_prev=i,iu(i,r),r=e._pack_next,o=3;o<l;o++){fu(e,r,i=t[o]);var p=0,d=1,g=1;for(u=r._pack_next;u!==r;u=u._pack_next,d++)if(uu(u,i)){p=1;break}if(1==p)for(a=e._pack_prev;a!==u._pack_prev&&!uu(a,i);a=a._pack_prev,g++);p?(d<g||d==g&&r.r<e.r?ou(e,r=u):ou(e=a,r),o--):(iu(e,i),r=i,x(i))}var v=(c+s)/2,m=(f+h)/2,y=0;for(o=0;o<l;o++)(i=t[o]).x-=v,i.y-=m,y=Math.max(y,i.r+Math.sqrt(i.x*i.x+i.y*i.y));n.r=y,t.forEach(cu)}function x(n){c=Math.min(n.x-n.r,c),s=Math.max(n.x+n.r,s),f=Math.min(n.y-n.r,f),h=Math.max(n.y+n.r,h)}}function lu(n){n._pack_next=n._pack_prev=n}function cu(n){delete n._pack_next,delete n._pack_prev}function su(n,t,e,r){var i=n.children;if(n.x=t+=r*n.x,n.y=e+=r*n.y,n.r*=r,i)for(var o=-1,u=i.length;++o<u;)su(i[o],t,e,r)}function fu(n,t,e){var r=n.r+e.r,i=t.x-n.x,o=t.y-n.y;if(r&&(i||o)){var u=t.r+e.r,a=i*i+o*o,l=.5+((r*=r)-(u*=u))/(2*a),c=Math.sqrt(Math.max(0,2*u*(r+a)-(r-=a)*r-u*u))/(2*a);e.x=n.x+l*i+c*o,e.y=n.y+l*o-c*i}else e.x=n.x+r,e.y=n.y}function hu(n,t){return n.parent==t.parent?1:2}function pu(n){var t=n.children;return t.length?t[0]:n.t}function du(n){var t,e=n.children;return(t=e.length)?e[t-1]:n.t}function gu(n,t,e){var r=e/(t.i-n.i);t.c-=r,t.s+=e,n.c+=r,t.z+=e,t.m+=e}function vu(n,t,e){return n.a.parent===t.parent?n.a:e}function mu(n){var t=n.children;return t&&t.length?mu(t[0]):n}function yu(n){var t,e=n.children;return e&&(t=e.length)?yu(e[t-1]):n}function xu(n){return{x:n.x,y:n.y,dx:n.dx,dy:n.dy}}function Mu(n,t){var e=n.x+t[3],r=n.y+t[0],i=n.dx-t[1]-t[3],o=n.dy-t[0]-t[2];return i<0&&(e+=i/2,i=0),o<0&&(r+=o/2,o=0),{x:e,y:r,dx:i,dy:o}}function bu(n){var t=n[0],e=n[n.length-1];return t<e?[t,e]:[e,t]}function wu(n){return n.rangeExtent?n.rangeExtent():bu(n.range())}function _u(n,t,e,r){var i=e(n[0],n[1]),o=r(t[0],t[1]);return function(n){return o(i(n))}}function Su(n,t){var e,r=0,i=n.length-1,o=n[r],u=n[i];return u<o&&(e=r,r=i,i=e,e=o,o=u,u=e),n[r]=t.floor(o),n[i]=t.ceil(u),n}function ku(n){return n?{floor:function(t){return Math.floor(t/n)*n},ceil:function(t){return Math.ceil(t/n)*n}}:Eu}o.layout.histogram=function(){var n=!0,t=Number,e=eu,r=nu;function i(i,u){for(var a,l,c=[],s=i.map(t,this),f=e.call(this,s,u),h=r.call(this,f,s,u),p=(u=-1,s.length),d=h.length-1,g=n?1:1/p;++u<d;)(a=c[u]=[]).dx=h[u+1]-(a.x=h[u]),a.y=0;if(d>0)for(u=-1;++u<p;)(l=s[u])>=f[0]&&l<=f[1]&&((a=c[o.bisect(h,l,1,d)-1]).y+=g,a.push(i[u]));return c}return i.value=function(n){return arguments.length?(t=n,i):t},i.range=function(n){return arguments.length?(e=xt(n),i):e},i.bins=function(n){return arguments.length?(r="number"==typeof n?function(t){return tu(t,n)}:xt(n),i):r},i.frequency=function(t){return arguments.length?(n=!!t,i):n},i},o.layout.pack=function(){var n,t=o.layout.hierarchy().sort(ru),e=0,r=[1,1];function i(i,o){var u=t.call(this,i,o),a=u[0],l=r[0],c=r[1],s=null==n?Math.sqrt:"function"==typeof n?n:function(){return n};if(a.x=a.y=0,Uo(a,function(n){n.r=+s(n.value)}),Uo(a,au),e){var f=e*(n?1:Math.max(2*a.r/l,2*a.r/c))/2;Uo(a,function(n){n.r+=f}),Uo(a,au),Uo(a,function(n){n.r-=f})}return su(a,l/2,c/2,n?1:1/Math.max(2*a.r/l,2*a.r/c)),u}return i.size=function(n){return arguments.length?(r=n,i):r},i.radius=function(t){return arguments.length?(n=null==t||"function"==typeof t?t:+t,i):n},i.padding=function(n){return arguments.length?(e=+n,i):e},Oo(i,t)},o.layout.tree=function(){var n=o.layout.hierarchy().sort(null).value(null),t=hu,e=[1,1],r=null;function i(i,o){var c=n.call(this,i,o),s=c[0],f=function(n){for(var t,e={A:null,children:[n]},r=[e];null!=(t=r.pop());)for(var i,o=t.children,u=0,a=o.length;u<a;++u)r.push((o[u]=i={_:o[u],parent:t,children:(i=o[u].children)&&i.slice()||[],A:null,a:null,z:0,m:0,c:0,s:0,t:null,i:u}).a=i);return e.children[0]}(s);if(Uo(f,u),f.parent.m=-f.z,Po(f,a),r)Po(s,l);else{var h=s,p=s,d=s;Po(s,function(n){n.x<h.x&&(h=n),n.x>p.x&&(p=n),n.depth>d.depth&&(d=n)});var g=t(h,p)/2-h.x,v=e[0]/(p.x+t(p,h)/2+g),m=e[1]/(d.depth||1);Po(s,function(n){n.x=(n.x+g)*v,n.y=n.depth*m})}return c}function u(n){var e=n.children,r=n.parent.children,i=n.i?r[n.i-1]:null;if(e.length){!function(n){for(var t,e=0,r=0,i=n.children,o=i.length;--o>=0;)(t=i[o]).z+=e,t.m+=e,e+=t.s+(r+=t.c)}(n);var o=(e[0].z+e[e.length-1].z)/2;i?(n.z=i.z+t(n._,i._),n.m=n.z-o):n.z=o}else i&&(n.z=i.z+t(n._,i._));n.parent.A=function(n,e,r){if(e){for(var i,o=n,u=n,a=e,l=o.parent.children[0],c=o.m,s=u.m,f=a.m,h=l.m;a=du(a),o=pu(o),a&&o;)l=pu(l),(u=du(u)).a=n,(i=a.z+f-o.z-c+t(a._,o._))>0&&(gu(vu(a,n,r),n,i),c+=i,s+=i),f+=a.m,c+=o.m,h+=l.m,s+=u.m;a&&!du(u)&&(u.t=a,u.m+=f-s),o&&!pu(l)&&(l.t=o,l.m+=c-h,r=n)}return r}(n,i,n.parent.A||r[0])}function a(n){n._.x=n.z+n.parent.m,n.m+=n.parent.m}function l(n){n.x*=e[0],n.y=n.depth*e[1]}return i.separation=function(n){return arguments.length?(t=n,i):t},i.size=function(n){return arguments.length?(r=null==(e=n)?l:null,i):r?null:e},i.nodeSize=function(n){return arguments.length?(r=null==(e=n)?null:l,i):r?e:null},Oo(i,n)},o.layout.cluster=function(){var n=o.layout.hierarchy().sort(null).value(null),t=hu,e=[1,1],r=!1;function i(i,u){var a,l=n.call(this,i,u),c=l[0],s=0;Uo(c,function(n){var e=n.children;e&&e.length?(n.x=function(n){return n.reduce(function(n,t){return n+t.x},0)/n.length}(e),n.y=function(n){return 1+o.max(n,function(n){return n.y})}(e)):(n.x=a?s+=t(n,a):0,n.y=0,a=n)});var f=mu(c),h=yu(c),p=f.x-t(f,h)/2,d=h.x+t(h,f)/2;return Uo(c,r?function(n){n.x=(n.x-c.x)*e[0],n.y=(c.y-n.y)*e[1]}:function(n){n.x=(n.x-p)/(d-p)*e[0],n.y=(1-(c.y?n.y/c.y:1))*e[1]}),l}return i.separation=function(n){return arguments.length?(t=n,i):t},i.size=function(n){return arguments.length?(r=null==(e=n),i):r?null:e},i.nodeSize=function(n){return arguments.length?(r=null!=(e=n),i):r?e:null},Oo(i,n)},o.layout.treemap=function(){var n,t=o.layout.hierarchy(),e=Math.round,r=[1,1],i=null,u=xu,a=!1,l="squarify",c=.5*(1+Math.sqrt(5));function s(n,t){for(var e,r,i=-1,o=n.length;++i<o;)r=(e=n[i]).value*(t<0?0:t),e.area=isNaN(r)||r<=0?0:r}function f(n){var t=n.children;if(t&&t.length){var e,r,i,o=u(n),a=[],c=t.slice(),h=1/0,g="slice"===l?o.dx:"dice"===l?o.dy:"slice-dice"===l?1&n.depth?o.dy:o.dx:Math.min(o.dx,o.dy);for(s(c,o.dx*o.dy/n.value),a.area=0;(i=c.length)>0;)a.push(e=c[i-1]),a.area+=e.area,"squarify"!==l||(r=p(a,g))<=h?(c.pop(),h=r):(a.area-=a.pop().area,d(a,g,o,!1),g=Math.min(o.dx,o.dy),a.length=a.area=0,h=1/0);a.length&&(d(a,g,o,!0),a.length=a.area=0),t.forEach(f)}}function h(n){var t=n.children;if(t&&t.length){var e,r=u(n),i=t.slice(),o=[];for(s(i,r.dx*r.dy/n.value),o.area=0;e=i.pop();)o.push(e),o.area+=e.area,null!=e.z&&(d(o,e.z?r.dx:r.dy,r,!i.length),o.length=o.area=0);t.forEach(h)}}function p(n,t){for(var e,r=n.area,i=0,o=1/0,u=-1,a=n.length;++u<a;)(e=n[u].area)&&(e<o&&(o=e),e>i&&(i=e));return t*=t,(r*=r)?Math.max(t*i*c/r,r/(t*o*c)):1/0}function d(n,t,r,i){var o,u=-1,a=n.length,l=r.x,c=r.y,s=t?e(n.area/t):0;if(t==r.dx){for((i||s>r.dy)&&(s=r.dy);++u<a;)(o=n[u]).x=l,o.y=c,o.dy=s,l+=o.dx=Math.min(r.x+r.dx-l,s?e(o.area/s):0);o.z=!0,o.dx+=r.x+r.dx-l,r.y+=s,r.dy-=s}else{for((i||s>r.dx)&&(s=r.dx);++u<a;)(o=n[u]).x=l,o.y=c,o.dx=s,c+=o.dy=Math.min(r.y+r.dy-c,s?e(o.area/s):0);o.z=!1,o.dy+=r.y+r.dy-c,r.x+=s,r.dx-=s}}function g(e){var i=n||t(e),o=i[0];return o.x=o.y=0,o.value?(o.dx=r[0],o.dy=r[1]):o.dx=o.dy=0,n&&t.revalue(o),s([o],o.dx*o.dy/o.value),(n?h:f)(o),a&&(n=i),i}return g.size=function(n){return arguments.length?(r=n,g):r},g.padding=function(n){if(!arguments.length)return i;function t(t){return Mu(t,n)}var e;return u=null==(i=n)?xu:"function"==(e=typeof n)?function(t){var e=n.call(g,t,t.depth);return null==e?xu(t):Mu(t,"number"==typeof e?[e,e,e,e]:e)}:"number"===e?(n=[n,n,n,n],t):t,g},g.round=function(n){return arguments.length?(e=n?Math.round:Number,g):e!=Number},g.sticky=function(t){return arguments.length?(a=t,n=null,g):a},g.ratio=function(n){return arguments.length?(c=n,g):c},g.mode=function(n){return arguments.length?(l=n+"",g):l},Oo(g,t)},o.random={normal:function(n,t){var e=arguments.length;return e<2&&(t=1),e<1&&(n=0),function(){var e,r,i;do{i=(e=2*Math.random()-1)*e+(r=2*Math.random()-1)*r}while(!i||i>1);return n+t*e*Math.sqrt(-2*Math.log(i)/i)}},logNormal:function(){var n=o.random.normal.apply(o,arguments);return function(){return Math.exp(n())}},bates:function(n){var t=o.random.irwinHall(n);return function(){return t()/n}},irwinHall:function(n){return function(){for(var t=0,e=0;e<n;e++)t+=Math.random();return t}}},o.scale={};var Eu={floor:z,ceil:z};function Nu(n,t,e,r){var i=[],u=[],a=0,l=Math.min(n.length,t.length)-1;for(n[l]<n[0]&&(n=n.slice().reverse(),t=t.slice().reverse());++a<=l;)i.push(e(n[a-1],n[a])),u.push(r(t[a-1],t[a]));return function(t){var e=o.bisect(n,t,1,l)-1;return u[e](i[e](t))}}function Lu(n,t,e,r){var i,o;function u(){var u=Math.min(n.length,t.length)>2?Nu:_u,l=r?ko:So;return i=u(n,t,l,e),o=u(t,n,l,eo),a}function a(n){return i(n)}return a.invert=function(n){return o(n)},a.domain=function(t){return arguments.length?(n=t.map(Number),u()):n},a.range=function(n){return arguments.length?(t=n,u()):t},a.rangeRound=function(n){return a.range(n).interpolate(mo)},a.clamp=function(n){return arguments.length?(r=n,u()):r},a.interpolate=function(n){return arguments.length?(e=n,u()):e},a.ticks=function(t){return Ru(n,t)},a.tickFormat=function(t,e){return zu(n,t,e)},a.nice=function(t){return Au(n,t),u()},a.copy=function(){return Lu(n,t,e,r)},u()}function Cu(n,t){return o.rebind(n,t,"range","rangeRound","interpolate","clamp")}function Au(n,t){return Su(n,ku(Tu(n,t)[2])),Su(n,ku(Tu(n,t)[2])),n}function Tu(n,t){null==t&&(t=10);var e=bu(n),r=e[1]-e[0],i=Math.pow(10,Math.floor(Math.log(r/t)/Math.LN10)),o=t/r*i;return o<=.15?i*=10:o<=.35?i*=5:o<=.75&&(i*=2),e[0]=Math.ceil(e[0]/i)*i,e[1]=Math.floor(e[1]/i)*i+.5*i,e[2]=i,e}function Ru(n,t){return o.range.apply(o,Tu(n,t))}function zu(n,t,e){var r=Tu(n,t);if(e){var i=Dt.exec(e);if(i.shift(),"s"===i[8]){var u=o.formatPrefix(Math.max(w(r[0]),w(r[1])));return i[7]||(i[7]="."+qu(u.scale(r[2]))),i[8]="f",e=o.format(i.join("")),function(n){return e(u.scale(n))+u.symbol}}i[7]||(i[7]="."+function(n,t){var e=qu(t[2]);return n in Du?Math.abs(e-qu(Math.max(w(t[0]),w(t[1]))))+ +("e"!==n):e-2*("%"===n)}(i[8],r)),e=i.join("")}else e=",."+qu(r[2])+"f";return o.format(e)}o.scale.linear=function(){return Lu([0,1],[0,1],eo,!1)};var Du={s:1,g:1,p:1,r:1,e:1};function qu(n){return-Math.floor(Math.log(n)/Math.LN10+.01)}function Ou(n,t,e,r){function i(n){return(e?Math.log(n<0?0:n):-Math.log(n>0?0:-n))/Math.log(t)}function u(n){return e?Math.pow(t,n):-Math.pow(t,-n)}function a(t){return n(i(t))}return a.invert=function(t){return u(n.invert(t))},a.domain=function(t){return arguments.length?(e=t[0]>=0,n.domain((r=t.map(Number)).map(i)),a):r},a.base=function(e){return arguments.length?(t=+e,n.domain(r.map(i)),a):t},a.nice=function(){var t=Su(r.map(i),e?Math:Uu);return n.domain(t),r=t.map(u),a},a.ticks=function(){var n=bu(r),o=[],a=n[0],l=n[1],c=Math.floor(i(a)),s=Math.ceil(i(l)),f=t%1?2:t;if(isFinite(s-c)){if(e){for(;c<s;c++)for(var h=1;h<f;h++)o.push(u(c)*h);o.push(u(c))}else for(o.push(u(c));c++<s;)for(h=f-1;h>0;h--)o.push(u(c)*h);for(c=0;o[c]<a;c++);for(s=o.length;o[s-1]>l;s--);o=o.slice(c,s)}return o},a.tickFormat=function(n,e){if(!arguments.length)return Pu;arguments.length<2?e=Pu:"function"!=typeof e&&(e=o.format(e));var r=Math.max(1,t*n/a.ticks().length);return function(n){var o=n/u(Math.round(i(n)));return o*t<t-.5&&(o*=t),o<=r?e(n):""}},a.copy=function(){return Ou(n.copy(),t,e,r)},Cu(a,n)}o.scale.log=function(){return Ou(o.scale.linear().domain([0,1]),10,!0,[1,10])};var Pu=o.format(".0e"),Uu={floor:function(n){return-Math.ceil(-n)},ceil:function(n){return-Math.floor(-n)}};function Hu(n,t,e){var r=ju(t),i=ju(1/t);function o(t){return n(r(t))}return o.invert=function(t){return i(n.invert(t))},o.domain=function(t){return arguments.length?(n.domain((e=t.map(Number)).map(r)),o):e},o.ticks=function(n){return Ru(e,n)},o.tickFormat=function(n,t){return zu(e,n,t)},o.nice=function(n){return o.domain(Au(e,n))},o.exponent=function(u){return arguments.length?(r=ju(t=u),i=ju(1/t),n.domain(e.map(r)),o):t},o.copy=function(){return Hu(n.copy(),t,e)},Cu(o,n)}function ju(n){return function(t){return t<0?-Math.pow(-t,n):Math.pow(t,n)}}function Iu(n,t){var e,r,i;function u(i){return r[((e.get(i)||("range"===t.t?e.set(i,n.push(i)):NaN))-1)%r.length]}function a(t,e){return o.range(n.length).map(function(n){return t+e*n})}return u.domain=function(r){if(!arguments.length)return n;n=[],e=new S;for(var i,o=-1,a=r.length;++o<a;)e.has(i=r[o])||e.set(i,n.push(i));return u[t.t].apply(u,t.a)},u.range=function(n){return arguments.length?(r=n,i=0,t={t:"range",a:arguments},u):r},u.rangePoints=function(e,o){arguments.length<2&&(o=0);var l=e[0],c=e[1],s=n.length<2?(l=(l+c)/2,0):(c-l)/(n.length-1+o);return r=a(l+s*o/2,s),i=0,t={t:"rangePoints",a:arguments},u},u.rangeRoundPoints=function(e,o){arguments.length<2&&(o=0);var l=e[0],c=e[1],s=n.length<2?(l=c=Math.round((l+c)/2),0):(c-l)/(n.length-1+o)|0;return r=a(l+Math.round(s*o/2+(c-l-(n.length-1+o)*s)/2),s),i=0,t={t:"rangeRoundPoints",a:arguments},u},u.rangeBands=function(e,o,l){arguments.length<2&&(o=0),arguments.length<3&&(l=o);var c=e[1]<e[0],s=e[c-0],f=(e[1-c]-s)/(n.length-o+2*l);return r=a(s+f*l,f),c&&r.reverse(),i=f*(1-o),t={t:"rangeBands",a:arguments},u},u.rangeRoundBands=function(e,o,l){arguments.length<2&&(o=0),arguments.length<3&&(l=o);var c=e[1]<e[0],s=e[c-0],f=e[1-c],h=Math.floor((f-s)/(n.length-o+2*l));return r=a(s+Math.round((f-s-(n.length-o)*h)/2),h),c&&r.reverse(),i=Math.round(h*(1-o)),t={t:"rangeRoundBands",a:arguments},u},u.rangeBand=function(){return i},u.rangeExtent=function(){return bu(t.a[0])},u.copy=function(){return Iu(n,t)},u.domain(n)}o.scale.pow=function(){return Hu(o.scale.linear(),1,[0,1])},o.scale.sqrt=function(){return o.scale.pow().exponent(.5)},o.scale.ordinal=function(){return Iu([],{t:"range",a:[[]]})},o.scale.category10=function(){return o.scale.ordinal().range($u)},o.scale.category20=function(){return o.scale.ordinal().range(Fu)},o.scale.category20b=function(){return o.scale.ordinal().range(Bu)},o.scale.category20c=function(){return o.scale.ordinal().range(Gu)};var $u=[2062260,16744206,2924588,14034728,9725885,9197131,14907330,8355711,12369186,1556175].map(st),Fu=[2062260,11454440,16744206,16759672,2924588,10018698,14034728,16750742,9725885,12955861,9197131,12885140,14907330,16234194,8355711,13092807,12369186,14408589,1556175,10410725].map(st),Bu=[3750777,5395619,7040719,10264286,6519097,9216594,11915115,13556636,9202993,12426809,15186514,15190932,8666169,11356490,14049643,15177372,8077683,10834324,13528509,14589654].map(st),Gu=[3244733,7057110,10406625,13032431,15095053,16616764,16625259,16634018,3253076,7652470,10607003,13101504,7695281,10394312,12369372,14342891,6513507,9868950,12434877,14277081].map(st);function Yu(n,t){var e;function r(){var r=0,u=t.length;for(e=[];++r<u;)e[r-1]=o.quantile(n,r/u);return i}function i(n){if(!isNaN(n=+n))return t[o.bisect(e,n)]}return i.domain=function(t){return arguments.length?(n=t.map(m).filter(y).sort(v),r()):n},i.range=function(n){return arguments.length?(t=n,r()):t},i.quantiles=function(){return e},i.invertExtent=function(r){return(r=t.indexOf(r))<0?[NaN,NaN]:[r>0?e[r-1]:n[0],r<e.length?e[r]:n[n.length-1]]},i.copy=function(){return Yu(n,t)},r()}function Zu(n,t,e){var r,i;function o(t){return e[Math.max(0,Math.min(i,Math.floor(r*(t-n))))]}function u(){return r=e.length/(t-n),i=e.length-1,o}return o.domain=function(e){return arguments.length?(n=+e[0],t=+e[e.length-1],u()):[n,t]},o.range=function(n){return arguments.length?(e=n,u()):e},o.invertExtent=function(t){return[t=(t=e.indexOf(t))<0?NaN:t/r+n,t+1/r]},o.copy=function(){return Zu(n,t,e)},u()}function Vu(n,t){function e(e){if(e<=e)return t[o.bisect(n,e)]}return e.domain=function(t){return arguments.length?(n=t,e):n},e.range=function(n){return arguments.length?(t=n,e):t},e.invertExtent=function(e){return e=t.indexOf(e),[n[e-1],n[e]]},e.copy=function(){return Vu(n,t)},e}function Xu(n){function t(n){return+n}return t.invert=t,t.domain=t.range=function(e){return arguments.length?(n=e.map(t),t):n},t.ticks=function(t){return Ru(n,t)},t.tickFormat=function(t,e){return zu(n,t,e)},t.copy=function(){return Xu(n)},t}function Wu(){return 0}o.scale.quantile=function(){return Yu([],[])},o.scale.quantize=function(){return Zu(0,1,[0,1])},o.scale.threshold=function(){return Vu([.5],[0,1])},o.scale.identity=function(){return Xu([0,1])},o.svg={},o.svg.arc=function(){var n=Ku,t=Qu,e=Wu,r=Ju,i=na,o=ta,u=ea;function a(){var a=Math.max(0,+n.apply(this,arguments)),c=Math.max(0,+t.apply(this,arguments)),s=i.apply(this,arguments)-Rn,f=o.apply(this,arguments)-Rn,h=Math.abs(f-s),p=s>f?0:1;if(c<a&&(d=c,c=a,a=d),h>=Tn)return l(c,p)+(a?l(a,1-p):"")+"Z";var d,g,v,m,y,x,M,b,w,_,S,k,E=0,N=0,L=[];if((m=(+u.apply(this,arguments)||0)/2)&&(v=r===Ju?Math.sqrt(a*a+c*c):+r.apply(this,arguments),p||(N*=-1),c&&(N=Un(v/c*Math.sin(m))),a&&(E=Un(v/a*Math.sin(m)))),c){y=c*Math.cos(s+N),x=c*Math.sin(s+N),M=c*Math.cos(f-N),b=c*Math.sin(f-N);var C=Math.abs(f-s-2*N)<=Cn?0:1;if(N&&ra(y,x,M,b)===p^C){var A=(s+f)/2;y=c*Math.cos(A),x=c*Math.sin(A),M=b=null}}else y=x=0;if(a){w=a*Math.cos(f-E),_=a*Math.sin(f-E),S=a*Math.cos(s+E),k=a*Math.sin(s+E);var T=Math.abs(s-f+2*E)<=Cn?0:1;if(E&&ra(w,_,S,k)===1-p^T){var R=(s+f)/2;w=a*Math.cos(R),_=a*Math.sin(R),S=k=null}}else w=_=0;if(h>Nn&&(d=Math.min(Math.abs(c-a)/2,+e.apply(this,arguments)))>.001){g=a<c^p?0:1;var z=d,D=d;if(h<Cn){var q=null==S?[w,_]:null==M?[y,x]:fi([y,x],[S,k],[M,b],[w,_]),O=y-q[0],P=x-q[1],U=M-q[0],H=b-q[1],j=1/Math.sin(Math.acos((O*U+P*H)/(Math.sqrt(O*O+P*P)*Math.sqrt(U*U+H*H)))/2),I=Math.sqrt(q[0]*q[0]+q[1]*q[1]);D=Math.min(d,(a-I)/(j-1)),z=Math.min(d,(c-I)/(j+1))}if(null!=M){var $=ia(null==S?[w,_]:[S,k],[y,x],c,z,p),F=ia([M,b],[w,_],c,z,p);d===z?L.push("M",$[0],"A",z,",",z," 0 0,",g," ",$[1],"A",c,",",c," 0 ",1-p^ra($[1][0],$[1][1],F[1][0],F[1][1]),",",p," ",F[1],"A",z,",",z," 0 0,",g," ",F[0]):L.push("M",$[0],"A",z,",",z," 0 1,",g," ",F[0])}else L.push("M",y,",",x);if(null!=S){var B=ia([y,x],[S,k],a,-D,p),G=ia([w,_],null==M?[y,x]:[M,b],a,-D,p);d===D?L.push("L",G[0],"A",D,",",D," 0 0,",g," ",G[1],"A",a,",",a," 0 ",p^ra(G[1][0],G[1][1],B[1][0],B[1][1]),",",1-p," ",B[1],"A",D,",",D," 0 0,",g," ",B[0]):L.push("L",G[0],"A",D,",",D," 0 0,",g," ",B[0])}else L.push("L",w,",",_)}else L.push("M",y,",",x),null!=M&&L.push("A",c,",",c," 0 ",C,",",p," ",M,",",b),L.push("L",w,",",_),null!=S&&L.push("A",a,",",a," 0 ",T,",",1-p," ",S,",",k);return L.push("Z"),L.join("")}function l(n,t){return"M0,"+n+"A"+n+","+n+" 0 1,"+t+" 0,"+-n+"A"+n+","+n+" 0 1,"+t+" 0,"+n}return a.innerRadius=function(t){return arguments.length?(n=xt(t),a):n},a.outerRadius=function(n){return arguments.length?(t=xt(n),a):t},a.cornerRadius=function(n){return arguments.length?(e=xt(n),a):e},a.padRadius=function(n){return arguments.length?(r=n==Ju?Ju:xt(n),a):r},a.startAngle=function(n){return arguments.length?(i=xt(n),a):i},a.endAngle=function(n){return arguments.length?(o=xt(n),a):o},a.padAngle=function(n){return arguments.length?(u=xt(n),a):u},a.centroid=function(){var e=(+n.apply(this,arguments)+ +t.apply(this,arguments))/2,r=(+i.apply(this,arguments)+ +o.apply(this,arguments))/2-Rn;return[Math.cos(r)*e,Math.sin(r)*e]},a};var Ju="auto";function Ku(n){return n.innerRadius}function Qu(n){return n.outerRadius}function na(n){return n.startAngle}function ta(n){return n.endAngle}function ea(n){return n&&n.padAngle}function ra(n,t,e,r){return(n-e)*t-(t-r)*n>0?0:1}function ia(n,t,e,r,i){var o=n[0]-t[0],u=n[1]-t[1],a=(i?r:-r)/Math.sqrt(o*o+u*u),l=a*u,c=-a*o,s=n[0]+l,f=n[1]+c,h=t[0]+l,p=t[1]+c,d=(s+h)/2,g=(f+p)/2,v=h-s,m=p-f,y=v*v+m*m,x=e-r,M=s*p-h*f,b=(m<0?-1:1)*Math.sqrt(Math.max(0,x*x*y-M*M)),w=(M*m-v*b)/y,_=(-M*v-m*b)/y,S=(M*m+v*b)/y,k=(-M*v+m*b)/y,E=w-d,N=_-g,L=S-d,C=k-g;return E*E+N*N>L*L+C*C&&(w=S,_=k),[[w-l,_-c],[w*e/x,_*e/x]]}function oa(n){var t=oi,e=ui,r=Je,i=aa,o=i.key,u=.7;function a(o){var a,l=[],c=[],s=-1,f=o.length,h=xt(t),p=xt(e);function d(){l.push("M",i(n(c),u))}for(;++s<f;)r.call(this,a=o[s],s)?c.push([+h.call(this,a,s),+p.call(this,a,s)]):c.length&&(d(),c=[]);return c.length&&d(),l.length?l.join(""):null}return a.x=function(n){return arguments.length?(t=n,a):t},a.y=function(n){return arguments.length?(e=n,a):e},a.defined=function(n){return arguments.length?(r=n,a):r},a.interpolate=function(n){return arguments.length?(o="function"==typeof n?i=n:(i=ua.get(n)||aa).key,a):o},a.tension=function(n){return arguments.length?(u=n,a):u},a}o.svg.line=function(){return oa(z)};var ua=o.map({linear:aa,"linear-closed":la,step:function(n){for(var t=0,e=n.length,r=n[0],i=[r[0],",",r[1]];++t<e;)i.push("H",(r[0]+(r=n[t])[0])/2,"V",r[1]);return e>1&&i.push("H",r[0]),i.join("")},"step-before":ca,"step-after":sa,basis:pa,"basis-open":function(n){if(n.length<4)return aa(n);for(var t,e=[],r=-1,i=n.length,o=[0],u=[0];++r<3;)t=n[r],o.push(t[0]),u.push(t[1]);for(e.push(da(ma,o)+","+da(ma,u)),--r;++r<i;)t=n[r],o.shift(),o.push(t[0]),u.shift(),u.push(t[1]),ya(e,o,u);return e.join("")},"basis-closed":function(n){for(var t,e,r=-1,i=n.length,o=i+4,u=[],a=[];++r<4;)e=n[r%i],u.push(e[0]),a.push(e[1]);for(t=[da(ma,u),",",da(ma,a)],--r;++r<o;)e=n[r%i],u.shift(),u.push(e[0]),a.shift(),a.push(e[1]),ya(t,u,a);return t.join("")},bundle:function(n,t){var e=n.length-1;if(e)for(var r,i,o=n[0][0],u=n[0][1],a=n[e][0]-o,l=n[e][1]-u,c=-1;++c<=e;)i=c/e,(r=n[c])[0]=t*r[0]+(1-t)*(o+i*a),r[1]=t*r[1]+(1-t)*(u+i*l);return pa(n)},cardinal:function(n,t){return n.length<3?aa(n):n[0]+fa(n,ha(n,t))},"cardinal-open":function(n,t){return n.length<4?aa(n):n[1]+fa(n.slice(1,-1),ha(n,t))},"cardinal-closed":function(n,t){return n.length<3?la(n):n[0]+fa((n.push(n[0]),n),ha([n[n.length-2]].concat(n,[n[1]]),t))},monotone:function(n){return n.length<3?aa(n):n[0]+fa(n,function(n){for(var t,e,r,i,o=[],u=function(n){for(var t=0,e=n.length-1,r=[],i=n[0],o=n[1],u=r[0]=xa(i,o);++t<e;)r[t]=(u+(u=xa(i=o,o=n[t+1])))/2;return r[t]=u,r}(n),a=-1,l=n.length-1;++a<l;)t=xa(n[a],n[a+1]),w(t)<Nn?u[a]=u[a+1]=0:(i=(e=u[a]/t)*e+(r=u[a+1]/t)*r)>9&&(i=3*t/Math.sqrt(i),u[a]=i*e,u[a+1]=i*r);for(a=-1;++a<=l;)i=(n[Math.min(l,a+1)][0]-n[Math.max(0,a-1)][0])/(6*(1+u[a]*u[a])),o.push([i||0,u[a]*i||0]);return o}(n))}});function aa(n){return n.length>1?n.join("L"):n+"Z"}function la(n){return n.join("L")+"Z"}function ca(n){for(var t=0,e=n.length,r=n[0],i=[r[0],",",r[1]];++t<e;)i.push("V",(r=n[t])[1],"H",r[0]);return i.join("")}function sa(n){for(var t=0,e=n.length,r=n[0],i=[r[0],",",r[1]];++t<e;)i.push("H",(r=n[t])[0],"V",r[1]);return i.join("")}function fa(n,t){if(t.length<1||n.length!=t.length&&n.length!=t.length+2)return aa(n);var e=n.length!=t.length,r="",i=n[0],o=n[1],u=t[0],a=u,l=1;if(e&&(r+="Q"+(o[0]-2*u[0]/3)+","+(o[1]-2*u[1]/3)+","+o[0]+","+o[1],i=n[1],l=2),t.length>1){a=t[1],o=n[l],l++,r+="C"+(i[0]+u[0])+","+(i[1]+u[1])+","+(o[0]-a[0])+","+(o[1]-a[1])+","+o[0]+","+o[1];for(var c=2;c<t.length;c++,l++)o=n[l],a=t[c],r+="S"+(o[0]-a[0])+","+(o[1]-a[1])+","+o[0]+","+o[1]}if(e){var s=n[l];r+="Q"+(o[0]+2*a[0]/3)+","+(o[1]+2*a[1]/3)+","+s[0]+","+s[1]}return r}function ha(n,t){for(var e,r=[],i=(1-t)/2,o=n[0],u=n[1],a=1,l=n.length;++a<l;)e=o,o=u,u=n[a],r.push([i*(u[0]-e[0]),i*(u[1]-e[1])]);return r}function pa(n){if(n.length<3)return aa(n);var t=1,e=n.length,r=n[0],i=r[0],o=r[1],u=[i,i,i,(r=n[1])[0]],a=[o,o,o,r[1]],l=[i,",",o,"L",da(ma,u),",",da(ma,a)];for(n.push(n[e-1]);++t<=e;)r=n[t],u.shift(),u.push(r[0]),a.shift(),a.push(r[1]),ya(l,u,a);return n.pop(),l.push("L",r),l.join("")}function da(n,t){return n[0]*t[0]+n[1]*t[1]+n[2]*t[2]+n[3]*t[3]}ua.forEach(function(n,t){t.key=n,t.closed=/-closed$/.test(n)});var ga=[0,2/3,1/3,0],va=[0,1/3,2/3,0],ma=[0,1/6,2/3,1/6];function ya(n,t,e){n.push("C",da(ga,t),",",da(ga,e),",",da(va,t),",",da(va,e),",",da(ma,t),",",da(ma,e))}function xa(n,t){return(t[1]-n[1])/(t[0]-n[0])}function Ma(n){for(var t,e,r,i=-1,o=n.length;++i<o;)e=(t=n[i])[0],r=t[1]-Rn,t[0]=e*Math.cos(r),t[1]=e*Math.sin(r);return n}function ba(n){var t=oi,e=oi,r=0,i=ui,o=Je,u=aa,a=u.key,l=u,c="L",s=.7;function f(a){var f,h,p,d=[],g=[],v=[],m=-1,y=a.length,x=xt(t),M=xt(r),b=t===e?function(){return h}:xt(e),w=r===i?function(){return p}:xt(i);function _(){d.push("M",u(n(v),s),c,l(n(g.reverse()),s),"Z")}for(;++m<y;)o.call(this,f=a[m],m)?(g.push([h=+x.call(this,f,m),p=+M.call(this,f,m)]),v.push([+b.call(this,f,m),+w.call(this,f,m)])):g.length&&(_(),g=[],v=[]);return g.length&&_(),d.length?d.join(""):null}return f.x=function(n){return arguments.length?(t=e=n,f):e},f.x0=function(n){return arguments.length?(t=n,f):t},f.x1=function(n){return arguments.length?(e=n,f):e},f.y=function(n){return arguments.length?(r=i=n,f):i},f.y0=function(n){return arguments.length?(r=n,f):r},f.y1=function(n){return arguments.length?(i=n,f):i},f.defined=function(n){return arguments.length?(o=n,f):o},f.interpolate=function(n){return arguments.length?(a="function"==typeof n?u=n:(u=ua.get(n)||aa).key,l=u.reverse||u,c=u.closed?"M":"L",f):a},f.tension=function(n){return arguments.length?(s=n,f):s},f}function wa(n){return n.radius}function _a(n){return[n.x,n.y]}function Sa(){return 64}function ka(){return"circle"}function Ea(n){var t=Math.sqrt(n/Cn);return"M0,"+t+"A"+t+","+t+" 0 1,1 0,"+-t+"A"+t+","+t+" 0 1,1 0,"+t+"Z"}o.svg.line.radial=function(){var n=oa(Ma);return n.radius=n.x,delete n.x,n.angle=n.y,delete n.y,n},ca.reverse=sa,sa.reverse=ca,o.svg.area=function(){return ba(z)},o.svg.area.radial=function(){var n=ba(Ma);return n.radius=n.x,delete n.x,n.innerRadius=n.x0,delete n.x0,n.outerRadius=n.x1,delete n.x1,n.angle=n.y,delete n.y,n.startAngle=n.y0,delete n.y0,n.endAngle=n.y1,delete n.y1,n},o.svg.chord=function(){var n=Gr,t=Yr,e=wa,r=na,i=ta;function o(e,r){var i,o,c=u(this,n,e,r),s=u(this,t,e,r);return"M"+c.p0+a(c.r,c.p1,c.a1-c.a0)+(o=s,((i=c).a0==o.a0&&i.a1==o.a1?l(c.r,c.p1,c.r,c.p0):l(c.r,c.p1,s.r,s.p0)+a(s.r,s.p1,s.a1-s.a0)+l(s.r,s.p1,c.r,c.p0))+"Z")}function u(n,t,o,u){var a=t.call(n,o,u),l=e.call(n,a,u),c=r.call(n,a,u)-Rn,s=i.call(n,a,u)-Rn;return{r:l,a0:c,a1:s,p0:[l*Math.cos(c),l*Math.sin(c)],p1:[l*Math.cos(s),l*Math.sin(s)]}}function a(n,t,e){return"A"+n+","+n+" 0 "+ +(e>Cn)+",1 "+t}function l(n,t,e,r){return"Q 0,0 "+r}return o.radius=function(n){return arguments.length?(e=xt(n),o):e},o.source=function(t){return arguments.length?(n=xt(t),o):n},o.target=function(n){return arguments.length?(t=xt(n),o):t},o.startAngle=function(n){return arguments.length?(r=xt(n),o):r},o.endAngle=function(n){return arguments.length?(i=xt(n),o):i},o},o.svg.diagonal=function(){var n=Gr,t=Yr,e=_a;function r(r,i){var o=n.call(this,r,i),u=t.call(this,r,i),a=(o.y+u.y)/2,l=[o,{x:o.x,y:a},{x:u.x,y:a},u];return"M"+(l=l.map(e))[0]+"C"+l[1]+" "+l[2]+" "+l[3]}return r.source=function(t){return arguments.length?(n=xt(t),r):n},r.target=function(n){return arguments.length?(t=xt(n),r):t},r.projection=function(n){return arguments.length?(e=n,r):e},r},o.svg.diagonal.radial=function(){var n=o.svg.diagonal(),t=_a,e=n.projection;return n.projection=function(n){return arguments.length?e(function(n){return function(){var t=n.apply(this,arguments),e=t[0],r=t[1]-Rn;return[e*Math.cos(r),e*Math.sin(r)]}}(t=n)):t},n},o.svg.symbol=function(){var n=ka,t=Sa;function e(e,r){return(Na.get(n.call(this,e,r))||Ea)(t.call(this,e,r))}return e.type=function(t){return arguments.length?(n=xt(t),e):n},e.size=function(n){return arguments.length?(t=xt(n),e):t},e};var Na=o.map({circle:Ea,cross:function(n){var t=Math.sqrt(n/5)/2;return"M"+-3*t+","+-t+"H"+-t+"V"+-3*t+"H"+t+"V"+-t+"H"+3*t+"V"+t+"H"+t+"V"+3*t+"H"+-t+"V"+t+"H"+-3*t+"Z"},diamond:function(n){var t=Math.sqrt(n/(2*Ca)),e=t*Ca;return"M0,"+-t+"L"+e+",0 0,"+t+" "+-e+",0Z"},square:function(n){var t=Math.sqrt(n)/2;return"M"+-t+","+-t+"L"+t+","+-t+" "+t+","+t+" "+-t+","+t+"Z"},"triangle-down":function(n){var t=Math.sqrt(n/La),e=t*La/2;return"M0,"+e+"L"+t+","+-e+" "+-t+","+-e+"Z"},"triangle-up":function(n){var t=Math.sqrt(n/La),e=t*La/2;return"M0,"+-e+"L"+t+","+e+" "+-t+","+e+"Z"}});o.svg.symbolTypes=Na.keys();var La=Math.sqrt(3),Ca=Math.tan(30*zn);X.transition=function(n){for(var t,e,r=za||++Oa,i=Ha(n),o=[],u=Da||{time:Date.now(),ease:fo,delay:0,duration:250},a=-1,l=this.length;++a<l;){o.push(t=[]);for(var c=this[a],s=-1,f=c.length;++s<f;)(e=c[s])&&ja(e,s,i,r,u),t.push(e)}return Ra(o,i,r)},X.interrupt=function(n){return this.each(null==n?Aa:Ta(Ha(n)))};var Aa=Ta(Ha());function Ta(n){return function(){var t,e,r;(t=this[n])&&(r=t[e=t.active])&&(r.timer.c=null,r.timer.t=NaN,--t.count?delete t[e]:delete this[n],t.active+=.5,r.event&&r.event.interrupt.call(this,this.__data__,r.index))}}function Ra(n,t,e){return B(n,qa),n.namespace=t,n.id=e,n}var za,Da,qa=[],Oa=0;function Pa(n,t,e,r){var i=n.id,o=n.namespace;return dn(n,"function"==typeof e?function(n,u,a){n[o][i].tween.set(t,r(e.call(n,n.__data__,u,a)))}:(e=r(e),function(n){n[o][i].tween.set(t,e)}))}function Ua(n){return null==n&&(n=""),function(){this.textContent=n}}function Ha(n){return null==n?"__transition__":"__transition_"+n+"__"}function ja(n,t,e,r,i){var o,u,a,l,c,s=n[e]||(n[e]={active:0,count:0}),f=s[r];function h(e){var i=s.active,h=s[i];for(var d in h&&(h.timer.c=null,h.timer.t=NaN,--s.count,delete s[i],h.event&&h.event.interrupt.call(n,n.__data__,h.index)),s)if(+d<r){var g=s[d];g.timer.c=null,g.timer.t=NaN,--s.count,delete s[d]}u.c=p,Nt(function(){return u.c&&p(e||1)&&(u.c=null,u.t=NaN),1},0,o),s.active=r,f.event&&f.event.start.call(n,n.__data__,t),c=[],f.tween.forEach(function(e,r){(r=r.call(n,n.__data__,t))&&c.push(r)}),l=f.ease,a=f.duration}function p(i){for(var o=i/a,u=l(o),h=c.length;h>0;)c[--h].call(n,u);if(o>=1)return f.event&&f.event.end.call(n,n.__data__,t),--s.count?delete s[r]:delete n[e],1}f||(o=i.time,u=Nt(function(n){var t=f.delay;if(u.t=t+o,t<=n)return h(n-t);u.c=h},0,o),f=s[r]={tween:new S,time:o,timer:u,delay:i.delay,duration:i.duration,ease:i.ease,index:t},i=null,++s.count)}qa.call=X.call,qa.empty=X.empty,qa.node=X.node,qa.size=X.size,o.transition=function(n,t){return n&&n.transition?za?n.transition(t):n:o.selection().transition(n)},o.transition.prototype=qa,qa.select=function(n){var t,e,r,i=this.id,o=this.namespace,u=[];n=W(n);for(var a=-1,l=this.length;++a<l;){u.push(t=[]);for(var c=this[a],s=-1,f=c.length;++s<f;)(r=c[s])&&(e=n.call(r,r.__data__,s,a))?("__data__"in r&&(e.__data__=r.__data__),ja(e,s,o,i,r[o][i]),t.push(e)):t.push(null)}return Ra(u,o,i)},qa.selectAll=function(n){var t,e,r,i,o,u=this.id,a=this.namespace,l=[];n=J(n);for(var c=-1,s=this.length;++c<s;)for(var f=this[c],h=-1,p=f.length;++h<p;)if(r=f[h]){o=r[a][u],e=n.call(r,r.__data__,h,c),l.push(t=[]);for(var d=-1,g=e.length;++d<g;)(i=e[d])&&ja(i,d,a,u,o),t.push(i)}return Ra(l,a,u)},qa.filter=function(n){var t,e,r=[];"function"!=typeof n&&(n=hn(n));for(var i=0,o=this.length;i<o;i++){r.push(t=[]);for(var u,a=0,l=(u=this[i]).length;a<l;a++)(e=u[a])&&n.call(e,e.__data__,a,i)&&t.push(e)}return Ra(r,this.namespace,this.id)},qa.tween=function(n,t){var e=this.id,r=this.namespace;return arguments.length<2?this.node()[r][e].tween.get(n):dn(this,null==t?function(t){t[r][e].tween.remove(n)}:function(i){i[r][e].tween.set(n,t)})},qa.attr=function(n,t){if(arguments.length<2){for(t in n)this.attr(t,n[t]);return this}var e="transform"==n?_o:eo,r=o.ns.qualify(n);function i(){this.removeAttribute(r)}function u(){this.removeAttributeNS(r.space,r.local)}return Pa(this,"attr."+n,t,r.local?function(n){return null==n?u:(n+="",function(){var t,i=this.getAttributeNS(r.space,r.local);return i!==n&&(t=e(i,n),function(n){this.setAttributeNS(r.space,r.local,t(n))})})}:function(n){return null==n?i:(n+="",function(){var t,i=this.getAttribute(r);return i!==n&&(t=e(i,n),function(n){this.setAttribute(r,t(n))})})})},qa.attrTween=function(n,t){var e=o.ns.qualify(n);return this.tween("attr."+n,e.local?function(n,r){var i=t.call(this,n,r,this.getAttributeNS(e.space,e.local));return i&&function(n){this.setAttributeNS(e.space,e.local,i(n))}}:function(n,r){var i=t.call(this,n,r,this.getAttribute(e));return i&&function(n){this.setAttribute(e,i(n))}})},qa.style=function(n,t,e){var r=arguments.length;if(r<3){if("string"!=typeof n){for(e in r<2&&(t=""),n)this.style(e,n[e],t);return this}e=""}function i(){this.style.removeProperty(n)}return Pa(this,"style."+n,t,function(t){return null==t?i:(t+="",function(){var r,i=s(this).getComputedStyle(this,null).getPropertyValue(n);return i!==t&&(r=eo(i,t),function(t){this.style.setProperty(n,r(t),e)})})})},qa.styleTween=function(n,t,e){return arguments.length<3&&(e=""),this.tween("style."+n,function(r,i){var o=t.call(this,r,i,s(this).getComputedStyle(this,null).getPropertyValue(n));return o&&function(t){this.style.setProperty(n,o(t),e)}})},qa.text=function(n){return Pa(this,"text",n,Ua)},qa.remove=function(){var n=this.namespace;return this.each("end.transition",function(){var t;this[n].count<2&&(t=this.parentNode)&&t.removeChild(this)})},qa.ease=function(n){var t=this.id,e=this.namespace;return arguments.length<1?this.node()[e][t].ease:("function"!=typeof n&&(n=o.ease.apply(o,arguments)),dn(this,function(r){r[e][t].ease=n}))},qa.delay=function(n){var t=this.id,e=this.namespace;return arguments.length<1?this.node()[e][t].delay:dn(this,"function"==typeof n?function(r,i,o){r[e][t].delay=+n.call(r,r.__data__,i,o)}:(n=+n,function(r){r[e][t].delay=n}))},qa.duration=function(n){var t=this.id,e=this.namespace;return arguments.length<1?this.node()[e][t].duration:dn(this,"function"==typeof n?function(r,i,o){r[e][t].duration=Math.max(1,n.call(r,r.__data__,i,o))}:(n=Math.max(1,n),function(r){r[e][t].duration=n}))},qa.each=function(n,t){var e=this.id,r=this.namespace;if(arguments.length<2){var i=Da,u=za;try{za=e,dn(this,function(t,i,o){Da=t[r][e],n.call(t,t.__data__,i,o)})}finally{Da=i,za=u}}else dn(this,function(i){var u=i[r][e];(u.event||(u.event=o.dispatch("start","end","interrupt"))).on(n,t)});return this},qa.transition=function(){for(var n,t,e,r=this.id,i=++Oa,o=this.namespace,u=[],a=0,l=this.length;a<l;a++){u.push(n=[]);for(var c,s=0,f=(c=this[a]).length;s<f;s++)(t=c[s])&&ja(t,s,o,i,{time:(e=t[o][r]).time,ease:e.ease,delay:e.delay+e.duration,duration:e.duration}),n.push(t)}return Ra(u,o,i)},o.svg.axis=function(){var n,t=o.scale.linear(),e=Ia,r=6,i=6,u=3,l=[10],c=null;function s(a){a.each(function(){var a,s=o.select(this),f=this.__chart__||t,h=this.__chart__=t.copy(),p=null==c?h.ticks?h.ticks.apply(h,l):h.domain():c,d=null==n?h.tickFormat?h.tickFormat.apply(h,l):z:n,g=s.selectAll(".tick").data(p,h),v=g.enter().insert("g",".domain").attr("class","tick").style("opacity",Nn),m=o.transition(g.exit()).style("opacity",Nn).remove(),y=o.transition(g.order()).style("opacity",1),x=Math.max(r,0)+u,M=wu(h),b=s.selectAll(".domain").data([0]),w=(b.enter().append("path").attr("class","domain"),o.transition(b));v.append("line"),v.append("text");var _,S,k,E,N=v.select("line"),L=y.select("line"),C=g.select("text").text(d),A=v.select("text"),T=y.select("text"),R="top"===e||"left"===e?-1:1;if("bottom"===e||"top"===e?(a=Fa,_="x",k="y",S="x2",E="y2",C.attr("dy",R<0?"0em":".71em").style("text-anchor","middle"),w.attr("d","M"+M[0]+","+R*i+"V0H"+M[1]+"V"+R*i)):(a=Ba,_="y",k="x",S="y2",E="x2",C.attr("dy",".32em").style("text-anchor",R<0?"end":"start"),w.attr("d","M"+R*i+","+M[0]+"H0V"+M[1]+"H"+R*i)),N.attr(E,R*r),A.attr(k,R*x),L.attr(S,0).attr(E,R*r),T.attr(_,0).attr(k,R*x),h.rangeBand){var D=h,q=D.rangeBand()/2;f=h=function(n){return D(n)+q}}else f.rangeBand?f=h:m.call(a,h,f);v.call(a,f,h),y.call(a,h,h)})}return s.scale=function(n){return arguments.length?(t=n,s):t},s.orient=function(n){return arguments.length?(e=n in $a?n+"":Ia,s):e},s.ticks=function(){return arguments.length?(l=a(arguments),s):l},s.tickValues=function(n){return arguments.length?(c=n,s):c},s.tickFormat=function(t){return arguments.length?(n=t,s):n},s.tickSize=function(n){var t=arguments.length;return t?(r=+n,i=+arguments[t-1],s):r},s.innerTickSize=function(n){return arguments.length?(r=+n,s):r},s.outerTickSize=function(n){return arguments.length?(i=+n,s):i},s.tickPadding=function(n){return arguments.length?(u=+n,s):u},s.tickSubdivide=function(){return arguments.length&&s},s};var Ia="bottom",$a={top:1,right:1,bottom:1,left:1};function Fa(n,t,e){n.attr("transform",function(n){var r=t(n);return"translate("+(isFinite(r)?r:e(n))+",0)"})}function Ba(n,t,e){n.attr("transform",function(n){var r=t(n);return"translate(0,"+(isFinite(r)?r:e(n))+")"})}o.svg.brush=function(){var n,t,e=$(h,"brushstart","brush","brushend"),r=null,i=null,u=[0,0],a=[0,0],l=!0,c=!0,f=Ya[0];function h(n){n.each(function(){var n=o.select(this).style("pointer-events","all").style("-webkit-tap-highlight-color","rgba(0,0,0,0)").on("mousedown.brush",v).on("touchstart.brush",v),t=n.selectAll(".background").data([0]);t.enter().append("rect").attr("class","background").style("visibility","hidden").style("cursor","crosshair"),n.selectAll(".extent").data([0]).enter().append("rect").attr("class","extent").style("cursor","move");var e=n.selectAll(".resize").data(f,z);e.exit().remove(),e.enter().append("g").attr("class",function(n){return"resize "+n}).style("cursor",function(n){return Ga[n]}).append("rect").attr("x",function(n){return/[ew]$/.test(n)?-3:null}).attr("y",function(n){return/^[ns]/.test(n)?-3:null}).attr("width",6).attr("height",6).style("visibility","hidden"),e.style("display",h.empty()?"none":null);var u,a=o.transition(n),l=o.transition(t);r&&(u=wu(r),l.attr("x",u[0]).attr("width",u[1]-u[0]),d(a)),i&&(u=wu(i),l.attr("y",u[0]).attr("height",u[1]-u[0]),g(a)),p(a)})}function p(n){n.selectAll(".resize").attr("transform",function(n){return"translate("+u[+/e$/.test(n)]+","+a[+/^s/.test(n)]+")"})}function d(n){n.select(".extent").attr("x",u[0]),n.selectAll(".extent,.n>rect,.s>rect").attr("width",u[1]-u[0])}function g(n){n.select(".extent").attr("y",a[0]),n.selectAll(".extent,.e>rect,.w>rect").attr("height",a[1]-a[0])}function v(){var f,v,m=this,y=o.select(o.event.target),x=e.of(m,arguments),M=o.select(m),b=y.datum(),w=!/^(n|s)$/.test(b)&&r,_=!/^(e|w)$/.test(b)&&i,S=y.classed("extent"),k=_n(m),E=o.mouse(m),N=o.select(s(m)).on("keydown.brush",function(){32==o.event.keyCode&&(S||(f=null,E[0]-=u[1],E[1]-=a[1],S=2),j())}).on("keyup.brush",function(){32==o.event.keyCode&&2==S&&(E[0]+=u[1],E[1]+=a[1],S=0,j())});if(o.event.changedTouches?N.on("touchmove.brush",A).on("touchend.brush",R):N.on("mousemove.brush",A).on("mouseup.brush",R),M.interrupt().selectAll("*").interrupt(),S)E[0]=u[0]-E[0],E[1]=a[0]-E[1];else if(b){var L=+/w$/.test(b),C=+/^n/.test(b);v=[u[1-L]-E[0],a[1-C]-E[1]],E[0]=u[L],E[1]=a[C]}else o.event.altKey&&(f=E.slice());function A(){var n=o.mouse(m),t=!1;v&&(n[0]+=v[0],n[1]+=v[1]),S||(o.event.altKey?(f||(f=[(u[0]+u[1])/2,(a[0]+a[1])/2]),E[0]=u[+(n[0]<f[0])],E[1]=a[+(n[1]<f[1])]):f=null),w&&T(n,r,0)&&(d(M),t=!0),_&&T(n,i,1)&&(g(M),t=!0),t&&(p(M),x({type:"brush",mode:S?"move":"resize"}))}function T(e,r,i){var o,s,h=wu(r),p=h[0],d=h[1],g=E[i],v=i?a:u,m=v[1]-v[0];if(S&&(p-=g,d-=m+g),o=(i?c:l)?Math.max(p,Math.min(d,e[i])):e[i],S?s=(o+=g)+m:(f&&(g=Math.max(p,Math.min(d,2*f[i]-o))),g<o?(s=o,o=g):s=g),v[0]!=o||v[1]!=s)return i?t=null:n=null,v[0]=o,v[1]=s,!0}function R(){A(),M.style("pointer-events","all").selectAll(".resize").style("display",h.empty()?"none":null),o.select("body").style("cursor",null),N.on("mousemove.brush",null).on("mouseup.brush",null).on("touchmove.brush",null).on("touchend.brush",null).on("keydown.brush",null).on("keyup.brush",null),k(),x({type:"brushend"})}M.style("pointer-events","none").selectAll(".resize").style("display",null),o.select("body").style("cursor",y.style("cursor")),x({type:"brushstart"}),A()}return h.event=function(r){r.each(function(){var r=e.of(this,arguments),i={x:u,y:a,i:n,j:t},l=this.__chart__||i;this.__chart__=i,za?o.select(this).transition().each("start.brush",function(){n=l.i,t=l.j,u=l.x,a=l.y,r({type:"brushstart"})}).tween("brush:brush",function(){var e=ro(u,i.x),o=ro(a,i.y);return n=t=null,function(n){u=i.x=e(n),a=i.y=o(n),r({type:"brush",mode:"resize"})}}).each("end.brush",function(){n=i.i,t=i.j,r({type:"brush",mode:"resize"}),r({type:"brushend"})}):(r({type:"brushstart"}),r({type:"brush",mode:"resize"}),r({type:"brushend"}))})},h.x=function(n){return arguments.length?(f=Ya[!(r=n)<<1|!i],h):r},h.y=function(n){return arguments.length?(f=Ya[!r<<1|!(i=n)],h):i},h.clamp=function(n){return arguments.length?(r&&i?(l=!!n[0],c=!!n[1]):r?l=!!n:i&&(c=!!n),h):r&&i?[l,c]:r?l:i?c:null},h.extent=function(e){var o,l,c,s,f;return arguments.length?(r&&(o=e[0],l=e[1],i&&(o=o[0],l=l[0]),n=[o,l],r.invert&&(o=r(o),l=r(l)),l<o&&(f=o,o=l,l=f),o==u[0]&&l==u[1]||(u=[o,l])),i&&(c=e[0],s=e[1],r&&(c=c[1],s=s[1]),t=[c,s],i.invert&&(c=i(c),s=i(s)),s<c&&(f=c,c=s,s=f),c==a[0]&&s==a[1]||(a=[c,s])),h):(r&&(n?(o=n[0],l=n[1]):(o=u[0],l=u[1],r.invert&&(o=r.invert(o),l=r.invert(l)),l<o&&(f=o,o=l,l=f))),i&&(t?(c=t[0],s=t[1]):(c=a[0],s=a[1],i.invert&&(c=i.invert(c),s=i.invert(s)),s<c&&(f=c,c=s,s=f))),r&&i?[[o,c],[l,s]]:r?[o,l]:i&&[c,s])},h.clear=function(){return h.empty()||(u=[0,0],a=[0,0],n=t=null),h},h.empty=function(){return!!r&&u[0]==u[1]||!!i&&a[0]==a[1]},o.rebind(h,e,"on")};var Ga={n:"ns-resize",e:"ew-resize",s:"ns-resize",w:"ew-resize",nw:"nwse-resize",ne:"nesw-resize",se:"nwse-resize",sw:"nesw-resize"},Ya=[["n","e","s","w","nw","ne","se","sw"],["e","w"],["n","s"],[]],Za=Pt.format=he.timeFormat,Va=Za.utc,Xa=Va("%Y-%m-%dT%H:%M:%S.%LZ");function Wa(n){return n.toISOString()}function Ja(n,t,e){function r(t){return n(t)}function i(n,e){var r=(n[1]-n[0])/e,i=o.bisect(Qa,r);return i==Qa.length?[t.year,Tu(n.map(function(n){return n/31536e6}),e)[2]]:i?t[r/Qa[i-1]<Qa[i]/r?i-1:i]:[el,Tu(n,e)[2]]}return r.invert=function(t){return Ka(n.invert(t))},r.domain=function(t){return arguments.length?(n.domain(t),r):n.domain().map(Ka)},r.nice=function(n,t){var e=r.domain(),o=bu(e),u=null==n?i(o,10):"number"==typeof n&&i(o,n);function a(e){return!isNaN(e)&&!n.range(e,Ka(+e+1),t).length}return u&&(n=u[0],t=u[1]),r.domain(Su(e,t>1?{floor:function(t){for(;a(t=n.floor(t));)t=Ka(t-1);return t},ceil:function(t){for(;a(t=n.ceil(t));)t=Ka(+t+1);return t}}:n))},r.ticks=function(n,t){var e=bu(r.domain()),o=null==n?i(e,10):"number"==typeof n?i(e,n):!n.range&&[{range:n},t];return o&&(n=o[0],t=o[1]),n.range(e[0],Ka(+e[1]+1),t<1?1:t)},r.tickFormat=function(){return e},r.copy=function(){return Ja(n.copy(),t,e)},Cu(r,n)}function Ka(n){return new Date(n)}Za.iso=Date.prototype.toISOString&&+new Date("2000-01-01T00:00:00.000Z")?Wa:Xa,Wa.parse=function(n){var t=new Date(n);return isNaN(t)?null:t},Wa.toString=Xa.toString,Pt.second=It(function(n){return new Ut(1e3*Math.floor(n/1e3))},function(n,t){n.setTime(n.getTime()+1e3*Math.floor(t))},function(n){return n.getSeconds()}),Pt.seconds=Pt.second.range,Pt.seconds.utc=Pt.second.utc.range,Pt.minute=It(function(n){return new Ut(6e4*Math.floor(n/6e4))},function(n,t){n.setTime(n.getTime()+6e4*Math.floor(t))},function(n){return n.getMinutes()}),Pt.minutes=Pt.minute.range,Pt.minutes.utc=Pt.minute.utc.range,Pt.hour=It(function(n){var t=n.getTimezoneOffset()/60;return new Ut(36e5*(Math.floor(n/36e5-t)+t))},function(n,t){n.setTime(n.getTime()+36e5*Math.floor(t))},function(n){return n.getHours()}),Pt.hours=Pt.hour.range,Pt.hours.utc=Pt.hour.utc.range,Pt.month=It(function(n){return(n=Pt.day(n)).setDate(1),n},function(n,t){n.setMonth(n.getMonth()+t)},function(n){return n.getMonth()}),Pt.months=Pt.month.range,Pt.months.utc=Pt.month.utc.range;var Qa=[1e3,5e3,15e3,3e4,6e4,3e5,9e5,18e5,36e5,108e5,216e5,432e5,864e5,1728e5,6048e5,2592e6,7776e6,31536e6],nl=[[Pt.second,1],[Pt.second,5],[Pt.second,15],[Pt.second,30],[Pt.minute,1],[Pt.minute,5],[Pt.minute,15],[Pt.minute,30],[Pt.hour,1],[Pt.hour,3],[Pt.hour,6],[Pt.hour,12],[Pt.day,1],[Pt.day,2],[Pt.week,1],[Pt.month,1],[Pt.month,3],[Pt.year,1]],tl=Za.multi([[".%L",function(n){return n.getMilliseconds()}],[":%S",function(n){return n.getSeconds()}],["%I:%M",function(n){return n.getMinutes()}],["%I %p",function(n){return n.getHours()}],["%a %d",function(n){return n.getDay()&&1!=n.getDate()}],["%b %d",function(n){return 1!=n.getDate()}],["%B",function(n){return n.getMonth()}],["%Y",Je]]),el={range:function(n,t,e){return o.range(Math.ceil(n/e)*e,+t,e).map(Ka)},floor:z,ceil:z};nl.year=Pt.year,Pt.scale=function(){return Ja(o.scale.linear(),nl,tl)};var rl=nl.map(function(n){return[n[0].utc,n[1]]}),il=Va.multi([[".%L",function(n){return n.getUTCMilliseconds()}],[":%S",function(n){return n.getUTCSeconds()}],["%I:%M",function(n){return n.getUTCMinutes()}],["%I %p",function(n){return n.getUTCHours()}],["%a %d",function(n){return n.getUTCDay()&&1!=n.getUTCDate()}],["%b %d",function(n){return 1!=n.getUTCDate()}],["%B",function(n){return n.getUTCMonth()}],["%Y",Je]]);function ol(n){return JSON.parse(n.responseText)}function ul(n){var t=l.createRange();return t.selectNode(l.body),t.createContextualFragment(n.responseText)}rl.year=Pt.year.utc,Pt.scale.utc=function(){return Ja(o.scale.linear(),rl,il)},o.text=Mt(function(n){return n.responseText}),o.json=function(n,t){return bt(n,"application/json",ol,t)},o.html=function(n,t){return bt(n,"text/html",ul,t)},o.xml=Mt(function(n){return n.responseXML}),this.d3=o,void 0===(i="function"==typeof(r=o)?r.call(t,e,t,n):r)||(n.exports=i)}()}},t={};function e(r){var i=t[r];if(void 0!==i)return i.exports;var o=t[r]={exports:{}};return n[r](o,o.exports,e),o.exports}e.n=n=>{var t=n&&n.__esModule?()=>n.default:()=>n;return e.d(t,{a:t}),t},e.d=(n,t)=>{for(var r in t)e.o(t,r)&&!e.o(n,r)&&Object.defineProperty(n,r,{enumerable:!0,get:t[r]})},e.o=(n,t)=>Object.prototype.hasOwnProperty.call(n,t),(()=>{"use strict";e(4143);var n,t,r=null,i=null,o=null,u=null,a=null,l=null,c=null,s=1,f=0,h=null,p=null,d=null,g=null,v=null,m=null,y=null,x=null,M=null,b=null,w=null,_=void 0!==(t=(n=document.getElementsByTagName("script"))[n.length-1]).getAttribute.length?t.src:t.getAttribute("src",-1);function S(){for(var n=[],t=[],e=0;e<s+1;e++)n.push(0),t.push(0);for(a.forEach(function(t){n[t.group]=n[t.group]+1}),e=0;e<a.length;e++)t[a[e].group]=t[a[e].group]+1,a[e].x=t[a[e].group]*((r-60)/(n[a[e].group]+1)),a[e].px=t[a[e].group]*((r-60)/(n[a[e].group]+1));H(),N(),forceOn&&ln(),L()}function k(n){d3.select(this).classed(n.fixed=!0)}function E(){x=null,M=null,y=null}function N(){for(var n=0;n<a.length;n++)a[n].y=i-30-a[n].group*f,a[n].py=i-30-a[n].group*f;g.attr("transform",function(n){return n.x>r-15?n.x=r-15:n.x<15&&(n.x=15),"translate("+n.x+","+n.y+")"}),d.attr("d",function(n){var t=n.target.x-n.source.x,e=n.target.y-n.source.y,o=Math.sqrt(t*t+e*e),u=t/o,a=e/o,l=n.left?17:12,c=n.right?17:12,s=n.source.x+l*u,f=n.source.y+l*a,h=n.target.x-c*u,p=n.target.y-c*a;return s>r-15?s=r-15:s<15&&(s=15),h>r-15?h=r-15:h<15&&(h=15),f>i-15?f=i-15:f<15&&(f=15),p>i-15?p=i-15:p<15&&(p=15),"M"+s+","+f+"L"+h+","+p})}function L(){(d=d.data(c)).classed("highlighted",function(n){return n.highlighted}).classed("selected",function(n){return n===m}).style("marker-start",function(n){return n.left?"url(#start-arrow)":""}).style("marker-end",function(n){return n.right?"url(#end-arrow)":""}),d.enter().append("svg:path").attr("class","link").classed("highlighted",function(n){return n.highlighted}).classed("selected",function(n){return n===m}).style("marker-start",function(n){return n.left?"url(#start-arrow)":""}).style("marker-end",function(n){return n.right?"url(#end-arrow)":""}).on("mousedown",function(n){!d3.event.shiftKey&&curEdit&&(m=(y=n)===m?null:y,v=null,curHighlight&&P(),L())}),d.exit().remove(),(g=g.data(a,function(n){return n.id})).selectAll("circle").style("fill",function(n){return n===v?d3.rgb(o(n.id)).brighter().toString():n.highlighted?"#FF0000":o(n.id)}).classed("highlighted",function(n){return n.highlighted}).attr("group",function(n){return n.group});var n=g.enter().append("svg:g");n.append("svg:circle").attr("class","node").attr("r",12).style("fill",function(n){return n===v?d3.rgb(o(n.id)).brighter().toString():n.highlighted?"#FF0000":o(n.id)}).style("stroke",function(n){return d3.rgb(o(n.id)).darker().toString()}).classed("highlighted",function(n){return n.highlighted}).on("mouseover",function(n){x&&n!==x&&d3.select(this).attr("transform","scale(1.1)")}).on("mouseout",function(n){x&&n!==x&&d3.select(this).attr("transform","")}).on("mousedown",function(n){d3.event.shiftKey||((x=n)===v?(v=null,curHighlight&&P()):(v=x,curHighlight&&U(v)),m=null,curEdit&&p.style("marker-end","url(#end-arrow)").classed("hidden",!1).attr("d","M"+x.x+","+x.y+"L"+x.x+","+x.y),L())}).on("mouseup",function(n){if(x)if(p.classed("hidden",!0).style("marker-end",""),(M=n)!==x){d3.select(this).attr("transform","");var t=null,e=x.id,r=M.id;if(1!=dataRelMatrix[e][r]){var o=function(n,t,e){if(1==n[e][t])return null;for(var r=J(n),i=[],o=[],u=0;u<r.length;u++)r[u][1]==t&&i.push(r[u][0]),r[u][0]==e&&o.push(r[u][1]);for(u=0;u<i.length;u++)for(var a=0;a<o.length;a++){if(1==n[o[a]][i[u]])return null;n[i[u]][o[a]]=1}return n}(dataRelMatrix,e,r);o?(h.stop(),dataRelMatrix=o.slice(),dataCovRel=K(dataRelMatrix),dataGroupList=j(dataRelMatrix),s=d3.max(dataGroupList),f=(i-60)/d3.max([s,1]),a=[],c=[],L(),a=I(dataLabels,dataGroupList),l=a.length,H(),c=F(a,dataCovRel),h.nodes(a).links(c),N(),E(),rn(),t=c.filter(function(n){return n.source.id==d3.min([e,r])&&n.target.id==d3.max([e,r])})[0],forceOn&&ln()):alert("Creating this relation would violate antisymmetry.")}curEdit&&(m=t),v=null,curHighlight&&P(),L()}else E()}).on("dblclick",function(n){name="";for(var t=/^[/^0-9a-zA-Z_]+$/;""==name;){if(name=prompt("Enter new label name.",n.name),t.test(name)||(alert("Please input alphanumeric characters only with no spaces."),name=""),name==n.name)return;C(name)&&(alert("Sorry, a node with that name already exists."),name="")}"null"!=name&&(n.name=name,d3.select(this.parentNode).select("text").text(function(n){return n.name}),dataLabels[n.id]=name)}),labelsOn&&n.append("svg:text").attr("x",0).attr("y",4).attr("class","id noselect").attr("pointer-events","none").text(function(n){return n.name}),g.exit().remove(),h.start()}function C(n){for(var t=0;t<a.length;t++)if(a[t].name==n)return!0;return!1}function A(){if(u.classed("active",!0),curEdit&&!d3.event.shiftKey&&!x&&!y){for(var n=d3.mouse(this),t=l+1;C(t.toString());)t+=1;t=t.toString();var e={id:l++,group:0,name:t,highlighted:!1};e.x=n[0],e.y=i-30,forceOn||(e.fixed=!0),a.push(e),dataLabels.push(t),dataRelMatrix=function(n){for(var t=[],e=0;e<n.length;e++)n[e].push(0),t.push(0);return t.push(1),n.push(t),n}(dataRelMatrix),rn(),L()}}function T(){x&&(curEdit&&p.attr("d","M"+x.x+","+x.y+"L"+d3.mouse(this)[0]+","+d3.mouse(this)[1]),L())}function R(){x&&p.classed("hidden",!0).style("marker-end",""),u.classed("active",!1),E(),L()}_=_.substring(0,_.length-14);var z=-1;function D(){if(-1===z&&(z=d3.event.keyCode,16===d3.event.keyCode&&(g.call(b),u.classed("shift",!0)),v||m)){switch(d3.event.keyCode){case 8:case 46:if(curEdit&&v)dataLabels.splice(v.id,1),h.stop(),dataRelMatrix=function(n,t){n.splice(t,1);for(var e=0;e<n.length;e++)n[e].splice(t,1);return n}(dataRelMatrix,v.id),dataCovRel=K(dataRelMatrix),dataGroupList=j(dataRelMatrix),s=d3.max(dataGroupList),f=(i-60)/d3.max([s,1]),a=[],c=[],L(),a=I(dataLabels,dataGroupList),l=a.length,H(),c=F(a,dataCovRel),h.nodes(a).links(c),E(),curHighlight&&P(),forceOn&&ln();else if(curEdit&&m){var n=m.source.id,t=m.target.id;dataRelMatrix[n][t]=0,dataRelMatrix[t][n]=0,dataCovRel=K(dataRelMatrix),dataGroupList=j(dataRelMatrix),s=d3.max(dataGroupList),f=(i-60)/d3.max([s,1]),a=[],c=[],L(),a=I(dataLabels,dataGroupList),l=a.length,H(),c=F(a,dataCovRel),h.nodes(a).links(c),E(),curHighlight&&P(),forceOn&&ln()}m=null,curEdit&&(v=null),rn(),L()}L()}}function q(){z=-1,16===d3.event.keyCode&&(g.on("mousedown.drag",null).on("touchstart.drag",null),u.classed("shift",!1))}function O(){g.call(b),u.classed("shift",!0),m=null,L()}function P(){for(var n=0;n<a.length;n++)a[n].highlighted=!1;for(n=0;n<c.length;n++)c[n].highlighted=!1;L()}function U(n){P();for(var t=J(dataRelMatrix),e=[],r=0;r<t.length;r++)t[r][1]==n.id&&e.push(t[r][0]),t[r][0]==n.id&&e.push(t[r][1]);for(r=0;r<e.length;r++)a[e[r]].highlighted=!0;for(r=0;r<c.length;r++)c[r].highlighted=en(c[r].source.id,e)&&en(c[r].target.id,e);L()}function H(){for(var n=0;n<a.length;n++)a[n].fixed=!0}function j(n){return fixExtremalNodes?function(n){for(var t=W(n),e=X(n),r=[],i=0;i<n.length;i++){for(var o=[],u=0;u<t.length;u++)for(var a=0;a<t[u].length;a++)t[u][a]==i&&o.push(t[u]);r.push(o.slice())}var l=[];for(i=0;i<r.length;i++){for(o=[],u=0;u<r[i].length;u++)o.push(r[i][u].length);l.push(o.slice())}var c=[];for(i=0;i<l.length;i++)c.push(d3.max(l[i])-1);var s=function(n){for(var t=n.length,e=Math.abs(n[0]),r=1;r<t;r++){for(var i=Math.abs(n[r]),o=e;e&&i;)e>i?e%=i:i%=e;e=Math.abs(o*n[r])/(e+i)}return e}(c),f=[];for(i=0;i<n.length;i++)f.push(s/c[i]*e[i]);return f}(n):null!=V(dataRelMatrix)?V(n):X(n)}function I(n,t){for(var e=[],o=0;o<n.length;o++)e.push({id:o,name:n[o],group:t[o],highlighted:!1});var u=[],a=[];for(o=0;o<s+1;o++)u.push(0),a.push(0);for(e.forEach(function(n){u[n.group]=u[n.group]+1}),o=0;o<e.length;o++)e[o].y=i-30-e[o].group*f,e[o].py=i-30-e[o].group*f,a[e[o].group]=a[e[o].group]+1,e[o].x=a[e[o].group]*((r-60)/(u[e[o].group]+1)),e[o].px=a[e[o].group]*((r-60)/(u[e[o].group]+1));return e}function F(n,t){for(var e=[],r=0;r<t.length;r++){var i=d3.min(dataCovRel[r]),o=d3.max(dataCovRel[r]);e.push({source:n[i],target:n[o],left:!1,right:!1,highlighted:!1})}return e}function B(){r=menuOpen?window.innerWidth-document.getElementById("side").clientWidth:window.innerWidth-10,i=window.innerHeight-10,f=(i-60)/d3.max([s,1]),u.attr("width",r),u.attr("height",i),forceOn||S(),h.size([r,i]).resume()}function G(n){if(0==n.length)return'error "No constructor for an empty poset."';for(var t=function(n,t){for(var e=[],r=0;r<n.length;r++)e.push([t[n[r][0]],t[n[r][1]]]);return e}(dataCovRel,n),e=function(n){for(var t="{{",e=0;e<n.length;e++){for(var r=0;r<n[e].length;r++)t=t+'"'+n[e][r].toString()+'"',r==n[e].length-1?t+="}":t+=",";e<n.length-1?t+=",{":t+="}"}return t}(t),r="{",i=n.length,o=0;o<i;o++)r=o!=i-1?r+'"'+n[o].toString()+'", ':r+'"'+n[o].toString()+'"}';return"poset("+r+","+e+")"}function Y(){forceOn||ln(),forceCharge=-un.noUiSlider.get(),h.charge(forceCharge).start()}function Z(){forceOn||ln(),forceLinkDist=an.noUiSlider.get(),h.linkDistance(forceLinkDist).start()}function V(n){for(var t=[],e=0;e<n.length;e++)t.push([e,0]);var r=dataCovRel;for(e=0;e<r.length;e++){var i=t[r[e][1]][t[r[e][1]].length-1]-t[r[e][0]][t[r[e][1]].length-1]-1,o=t[r[e][0]][0],u=t[r[e][1]][0];if(o==u){if(0==i)continue;return null}var a=[];if(i>0)for(var l=0;l<t.length;l++)t[l][0]==o?a.push([u,t[l][t[l].length-1]+i]):a.push(t[l]);else for(l=0;l<t.length;l++)t[l][0]==u?a.push([o,t[l][t[l].length-1]-i]):a.push(t[l]);t=a}var c=[];for(e=0;e<t.length;e++)c.push(t[e][1]);return c}function X(n){for(var t=function(n){for(var t=dataCovRel,e=[],r=[],i=0;i<n.length;i++){for(var o=0,u=[],a=0;a<t.length;a++)t[a][1]==i&&(o+=1),t[a][0]==i&&u.push(t[a][1]);e.push(o),r.push(u)}var l=[];for(i=0;i<e.length;i++)0==e[i]&&l.push(i);for(var c=[l.slice()];l.length>0;){var s=[];for(i=0;i<l.length;i++)s.push(r[l[i]]);s=Q(s);var f=[];for(i=0;i<s.length;i++)1==e[s[i]]?f.push(s[i]):e[s[i]]=e[s[i]]-1;l=f.slice(),c.push(f.slice())}return c.splice(c.length-1,1),c}(n),e=[],r=0;r<n.length;r++)for(var i=0;i<t.length;i++)for(var o=0;o<t[i].length;o++)t[i][o]==r&&e.push(i);return e}function W(n){for(var t=function(n){for(var t=[],e=0;e<n.length;e++){for(var r=!0,i=0;i<n.length;i++)if(1==n[i][e]&&e!=i){r=!1;break}r&&t.push(e)}return t}(n),e=[],r=0;r<t.length;r++)e.push([t[r]]);var i=dataCovRel,o=[];for(r=0;r<n.length;r++){for(var u=[],a=0;a<i.length;a++)i[a][0]==r&&u.push(i[a][1]);o.push(u)}for(var l=[];e.length>0;){var c=[];for(r=0;r<e.length;r++){var s=[];if(0!=(u=o[e[r][e[r].length-1]]).length){for(a=0;a<u.length;a++)s.push(e[r].concat(u[a]));c.push(s.slice())}else l.push(e[r].slice())}e=Q(c.slice())}return l}function J(n){for(var t=[],e=n.length,r=0;r<e;r++)for(var i=r;i<e;i++)1==n[r][i]?t.push([r,i]):1==n[i][r]&&t.push([i,r]);return t}function K(n){for(var t=n.length,e=[],r=[],i=0;i<t;i++){for(var o=[],u=0;u<t;u++)i!=u&&1==n[i][u]&&o.push(u);r.push(o)}for(i=0;i<t;i++){var a=[],l=r[i];for(u=0;u<l.length;u++)a.push(r[l[u]]);var c=tn(l,a=nn(Q(a)));for(u=0;u<c.length;u++)e.push([i,c[u]])}return e}function Q(n){return[].concat.apply([],n)}function nn(n){var t,e=n.length,r=[],i={};for(t=0;t<e;t++)i[n[t]]=0;for(t in i)r.push(t);return r}function tn(n,t){for(var e=n.length,r=t.length,i=[],o=0;o<e;o++)for(var u=0;u<r;u++)n[o]==t[u]&&i.push(o);i=nn(i);var a=0;for(o=0;o<i.length;o++)n.splice(i[o]-a,1),a+=1;return n}function en(n,t){for(var e=0;e<t.length;e++)if(t[e]==n)return!0;return!1}function rn(){d3.select("#isAtomic").html("&nbsp;&nbsp; isAtomic"),d3.select("#isBounded").html("&nbsp;&nbsp; isBounded"),d3.select("#isDistributive").html("&nbsp;&nbsp; isDistributive"),d3.select("#isGeometric").html("&nbsp;&nbsp; isGeometric"),d3.select("#isGraded").html("&nbsp;&nbsp; isGraded"),d3.select("#isLattice").html("&nbsp;&nbsp; isLattice"),d3.select("#isLowerSemilattice").html("&nbsp;&nbsp; isLowerSemilattice"),d3.select("#isLowerSemimodular").html("&nbsp;&nbsp; isLowerSemimodular"),d3.select("#isModular").html("&nbsp;&nbsp; isModular"),d3.select("#isRanked").html("&nbsp;&nbsp; isRanked"),d3.select("#isSperner").html("&nbsp;&nbsp; isSperner"),d3.select("#isStrictSperner").html("&nbsp;&nbsp; isStrictSperner"),d3.select("#isUpperSemilattice").html("&nbsp;&nbsp; isUpperSemilattice"),d3.select("#isUpperSemimodular").html("&nbsp;&nbsp; isUpperSemimodular"),d3.select("#dilworthNumber").html("&nbsp;&nbsp; dilworthNumber"),tikzGenerated&&(d3.select("#tikzHolder").node().remove(),tikzGenerated=!1)}function on(n,t,e){var r=function(n,t){var e=new XMLHttpRequest;return"withCredentials"in e?e.open(n,t,!0):"undefined"!=typeof XDomainRequest?(e=new XDomainRequest).open(n,t):e=null,e}(n,t);r?(r.onload=function(){var n;n=r.responseText,"isAtomic"==w?d3.select("#isAtomic").html("&nbsp;&nbsp; isAtomic :: <b>"+n+"</b>"):"isBounded"==w?d3.select("#isBounded").html("&nbsp;&nbsp; isBounded :: <b>"+n+"</b>"):"isDistributive"==w?d3.select("#isDistributive").html("&nbsp;&nbsp; isDistributive :: <b>"+n+"</b>"):"isGeometric"==w?d3.select("#isGeometric").html("&nbsp;&nbsp; isGeometric :: <b>"+n+"</b>"):"isGraded"==w?d3.select("#isGraded").html("&nbsp;&nbsp; isGraded :: <b>"+n+"</b>"):"isLattice"==w?d3.select("#isLattice").html("&nbsp;&nbsp; isLattice :: <b>"+n+"</b>"):"isLowerSemilattice"==w?d3.select("#isLowerSemilattice").html("&nbsp;&nbsp; isLowerSemilattice :: <b>"+n+"</b>"):"isLowerSemimodular"==w?d3.select("#isLowerSemimodular").html("&nbsp;&nbsp; isLowerSemimodular :: <b>"+n+"</b>"):"isModular"==w?d3.select("#isModular").html("&nbsp;&nbsp; isModular :: <b>"+n+"</b>"):"isRanked"==w?d3.select("#isRanked").html("&nbsp;&nbsp; isRanked :: <b>"+n+"</b>"):"isSperner"==w?d3.select("#isSperner").html("&nbsp;&nbsp; isSperner :: <b>"+n+"</b>"):"isStrictSperner"==w?d3.select("#isStrictSperner").html("&nbsp;&nbsp; isStrictSperner :: <b>"+n+"</b>"):"isUpperSemilattice"==w?d3.select("#isUpperSemilattice").html("&nbsp;&nbsp; isUpperSemilattice :: <b>"+n+"</b>"):"isUpperSemimodular"==w?d3.select("#isUpperSemimodular").html("&nbsp;&nbsp; isUpperSemimodular :: <b>"+n+"</b>"):"dilworthNumber"==w&&d3.select("#dilworthNumber").html("&nbsp;&nbsp; dilworthNumber :: <b>"+n+"</b>")},r.send(e)):alert("CORS not supported")}$("#side").BootSideMenu({side:"right",closeOnClick:!1,width:"230px"}),document.getElementsByClassName("toggler")[0].addEventListener("mousedown",function(){menuOpen=!menuOpen,B()},!1),window.addEventListener("resize",B,!1);var un=document.getElementById("charge-slider");noUiSlider.create(un,{start:[1500],range:{min:[0],max:[6e3]}});var an=document.getElementById("linkdist-slider");function ln(){forceOn?(H(),document.getElementById("forceToggle").innerHTML="Turn on force"):(function(){for(var n=0;n<a.length;n++)a[n].fixed=!1}(),L(),document.getElementById("forceToggle").innerHTML="Turn off force"),forceOn=!forceOn}noUiSlider.create(an,{start:[100],range:{min:[0],max:[400]}}),$(document).ready(function(){$("#editToggle").on("click",function(){curEdit?($(this).html("Enable editing"),curEdit=!curEdit,O()):($(this).html("Disable editing"),curEdit=!curEdit,g.on("mousedown.drag",null).on("touchstart.drag",null),u.classed("shift",!1))}),$("#labelToggle").on("click",function(){labelsOn?($(this).html("Show labels"),labelsOn=!labelsOn,g.select("text").remove()):($(this).html("Hide labels"),labelsOn=!labelsOn,g.append("svg:text").attr("x",0).attr("y",4).attr("class","id noselect").attr("pointer-events","none").text(function(n){return n.name}))}),$("#highlightToggle").on("click",function(){curHighlight?($(this).html("Highlight comparable elements"),P(),curHighlight=!curHighlight):($(this).html("Disable highlighting"),dataMaxChains=W(dataRelMatrix),v&&U(v),curHighlight=!curHighlight)}),$("#extremalNodeToggle").on("click",function(){fixExtremalNodes?($(this).html("Fix extremal nodes"),fixExtremalNodes=!fixExtremalNodes):($(this).html("Don't fix extremal nodes"),fixExtremalNodes=!fixExtremalNodes),h.stop(),a=[],c=[],L(),dataGroupList=j(dataRelMatrix),s=d3.max(dataGroupList),f=(i-60)/s,a=I(dataLabels,dataGroupList),H(),c=F(a,dataCovRel),h.nodes(a).links(c),N(),forceOn&&ln(),L()}),$("#reset").on("click",S),$("#forceToggle").on("click",ln),$("#exportTikz").on("click",function(){!function(){for(var n=[],t=0;t<a.length;t++)n[t]=[a[t].x.toString()+"/"+a[t].y.toString()+"/"+a[t].id+"/"+a[t].name];for(var e=[],r=0;r<c.length;r++)e[r]=[c[r].source.id.toString()+"/"+c[r].target.id.toString()];var i,o=function(){for(var n="",t=0;t<5;t++)n+="ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz".charAt(Math.floor(52*Math.random()));return n}();if(i="\\begin{tikzpicture}\n\\newcommand*\\points"+o+"{"+n+"}\n\\newcommand*\\edges"+o+"{"+e+"}\n\\newcommand*\\scale"+o+"{0.02}\n\\foreach \\x/\\y/\\z/\\w in \\points"+o+"\n  \\node (\\z) at (\\scale"+o+"*\\x,-\\scale"+o+"*\\y) [circle,draw,inner sep=0pt] {$\\w$};\n\\foreach \\x/\\y in \\edges"+o+"\n  \\draw (\\x) -- (\\y);\n\\end{tikzpicture}\n% \\points"+o+" is point set in the form x-coord/y-coord/node ID/node label\n% \\edges"+o+" is edge set in the form Source ID/Target ID\n% \\scale"+o+" makes the picture able to be viewed on the page\n",!tikzGenerated){var u=document.createElement("div");u.id="tikzHolder",u.className="list-group-item",u.setAttribute("href","#");var l=document.createElement("textarea");l.value="",l.id="tikzTextBox",l.rows=20,l.style="vertical-align:middle; width: 100%";var s=document.createElement("button");s.id="copyButton",s.style="vertical-align:middle;",s.type="button";var f=document.createElement("i");f.classList.add("fa-solid","fa-clipboard"),s.appendChild(f),u.appendChild(l),u.appendChild(s);var h=document.getElementById("menuList");h.insertBefore(u,h.childNodes[16]),document.getElementById("copyButton").setAttribute("data-clipboard-target","#tikzTextBox"),new ClipboardJS("#copyButton").on("error",function(n){window.alert("Press enter, then CTRL-C or CMD-C to copy")}),tikzGenerated=!0}document.getElementById("tikzTextBox").value=i}()}),$("#isAtomic").on("click",function(){w="isAtomic",on("POST","http://localhost:"+portData+"/isAtomic/",G(dataLabels))}),$("#isBounded").on("click",function(){w="isBounded",on("POST","http://localhost:"+portData+"/isBounded/",G(dataLabels))}),$("#isDistributive").on("click",function(){w="isDistributive",on("POST","http://localhost:"+portData+"/isDistributive/",G(dataLabels))}),$("#isGeometric").on("click",function(){w="isGeometric",on("POST","http://localhost:"+portData+"/isGeometric/",G(dataLabels))}),$("#isGraded").on("click",function(){w="isGraded",on("POST","http://localhost:"+portData+"/isGraded/",G(dataLabels))}),$("#isLattice").on("click",function(){w="isLattice",on("POST","http://localhost:"+portData+"/isLattice/",G(dataLabels))}),$("#isLowerSemilattice").on("click",function(){w="isLowerSemilattice",on("POST","http://localhost:"+portData+"/isLowerSemilattice/",G(dataLabels))}),$("#isLowerSemimodular").on("click",function(){w="isLowerSemimodular",on("POST","http://localhost:"+portData+"/isLowerSemimodular/",G(dataLabels))}),$("#isModular").on("click",function(){w="isModular",on("POST","http://localhost:"+portData+"/isModular/",G(dataLabels))}),$("#isRanked").on("click",function(){w="isRanked",on("POST","http://localhost:"+portData+"/isRanked/",G(dataLabels))}),$("#isSperner").on("click",function(){w="isSperner",on("POST","http://localhost:"+portData+"/isSperner/",G(dataLabels))}),$("#isStrictSperner").on("click",function(){w="isStrictSperner",on("POST","http://localhost:"+portData+"/isStrictSperner/",G(dataLabels))}),$("#isUpperSemilattice").on("click",function(){w="isUpperSemilattice",on("POST","http://localhost:"+portData+"/isUpperSemilattice/",G(dataLabels))}),$("#isUpperSemimodular").on("click",function(){w="isUpperSemimodular",on("POST","http://localhost:"+portData+"/isUpperSemimodular/",G(dataLabels))}),$("#dilworthNumber").on("click",function(){w="dilworthNumber",on("POST","http://localhost:"+portData+"/dilworthNumber/",G(dataLabels))}),$("#endSession").on("click",function(){activeSession&&(document.getElementById("endSession").style.color="white",document.getElementById("endSession").style.backgroundColor="red",document.getElementById("endSession").innerHTML="Session terminated",on("POST","http://localhost:"+portData+"/end/",G(dataLabels)),activeSession=!activeSession)}),r=window.innerWidth-document.getElementById("side").clientWidth,i=window.innerHeight-10,o=d3.scale.category10(),u=d3.select("body").append("svg").attr("width",r).attr("height",i).attr("id","canvasElement2d"),fixExtremalNodes&&(document.getElementById("extremalNodeToggle").text="Don't fix extremal nodes"),dataCovRel=K(dataRelMatrix),dataGroupList=j(dataRelMatrix),s=d3.max(dataGroupList),f=(i-60)/d3.max([s,1]),a=I(dataLabels,dataGroupList),l=a.length,H(),c=F(a,dataCovRel),h=d3.layout.force().nodes(a).links(c).size([r,i]).linkDistance(forceLinkDist).charge(forceCharge).on("tick",N),un.noUiSlider.on("slide",Y),an.noUiSlider.on("slide",Z),u.append("svg:defs").append("svg:marker").attr("id","end-arrow").attr("viewBox","0 -5 10 10").attr("refX",6).attr("markerWidth",3).attr("markerHeight",3).attr("orient","auto").append("svg:path").attr("d","M0,-5L10,0L0,5").attr("fill","#000"),b=h.drag().on("dragstart",k),p=u.append("svg:path").attr("class","link dragline hidden").attr("d","M0,0L0,0"),d=u.append("svg:g").selectAll("path"),g=u.append("svg:g").selectAll("g"),v=null,m=null,y=null,x=null,M=null,u.on("mousedown",A).on("mousemove",T).on("mouseup",R),d3.select(window).on("keydown",D).on("keyup",q),L(),O()})})()})();
